% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AddFacet}
\alias{AddFacet}
\title{AddFacet}
\usage{
AddFacet(p, fv1 = NULL, fv2 = NULL, Exclude = "None", Debug = FALSE)
}
\arguments{
\item{Debug}{= FALSE}

\item{data}{Source data.table}

\item{ShapColNames}{Names of the columns that contain shap values you want included}

\item{FacetVar1}{Column name}

\item{FacetVar2}{Column name}

\item{AggMethod}{A string for aggregating shapely values for importances. Choices include, 'mean', 'absmean', 'meanabs', 'sd', 'median', 'absmedian', 'medianabs'}

\item{TopN}{The number of variables to plot}
}
\description{
Add up to two facet variables for plots
}
\seealso{
Other Graphics: 
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
\keyword{internal}
