% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions_NEW.R
\name{BarPlot3D}
\alias{BarPlot3D}
\title{BarPlot3D}
\usage{
BarPlot3D(
  dt,
  PreAgg = FALSE,
  AggMethod = "mean",
  XVar = NULL,
  YVar = NULL,
  ZVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  ZVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  NumberBins = 21,
  NumLevels_Y = 33,
  NumLevels_X = 33,
  Height = NULL,
  Width = NULL,
  ShowLabels = FALSE,
  Theme = "dark",
  MouseScroll = FALSE,
  TextColor = "white",
  ContainLabel = TRUE,
  title.text = "3D Bar Plot",
  title.subtext = NULL,
  title.link = NULL,
  title.sublink = NULL,
  title.Align = NULL,
  title.top = NULL,
  title.left = NULL,
  title.right = NULL,
  title.bottom = NULL,
  title.padding = NULL,
  title.itemGap = NULL,
  title.backgroundColor = NULL,
  title.borderColor = NULL,
  title.borderWidth = NULL,
  title.borderRadius = NULL,
  title.shadowColor = NULL,
  title.shadowBlur = NULL,
  title.shadowOffsetX = NULL,
  title.shadowOffsetY = NULL,
  title.textStyle.color = NULL,
  title.textStyle.fontStyle = NULL,
  title.textStyle.fontWeight = NULL,
  title.textStyle.fontFamily = NULL,
  title.textStyle.fontSize = NULL,
  title.textStyle.lineHeight = NULL,
  title.textStyle.width = NULL,
  title.textStyle.height = NULL,
  title.textStyle.textBorderColor = NULL,
  title.textStyle.textBorderWidth = NULL,
  title.textStyle.textBorderType = NULL,
  title.textStyle.textBorderDashOffset = NULL,
  title.textStyle.textShadowColor = NULL,
  title.textStyle.textShadowBlur = NULL,
  title.textStyle.textShadowOffsetX = NULL,
  title.textStyle.textShadowOffsetY = NULL,
  title.subtextStyle.color = NULL,
  title.subtextStyle.align = NULL,
  title.subtextStyle.fontStyle = NULL,
  title.subtextStyle.fontWeight = NULL,
  title.subtextStyle.fontFamily = NULL,
  title.subtextStyle.fontSize = NULL,
  title.subtextStyle.lineHeight = NULL,
  title.subtextStyle.width = NULL,
  title.subtextStyle.height = NULL,
  title.subtextStyle.textBorderColor = NULL,
  title.subtextStyle.textBorderWidth = NULL,
  title.subtextStyle.textBorderType = NULL,
  title.subtextStyle.textBorderDashOffset = NULL,
  title.subtextStyle.textShadowColor = NULL,
  title.subtextStyle.textShadowBlur = NULL,
  title.subtextStyle.textShadowOffsetX = NULL,
  title.subtextStyle.textShadowOffsetY = NULL,
  xAxis.title = NULL,
  xAxis.min = NULL,
  xAxis.max = NULL,
  xAxis.position = "bottom",
  xAxis.nameLocation = "center",
  xAxis.axisTick.customValues = NULL,
  xAxis.nameTextStyle.color = NULL,
  xAxis.nameTextStyle.padding = 25,
  xAxis.nameTextStyle.align = NULL,
  xAxis.nameTextStyle.fontStyle = NULL,
  xAxis.nameTextStyle.fontWeight = NULL,
  xAxis.nameTextStyle.fontSize = NULL,
  xAxis.nameTextStyle.fontFamily = NULL,
  xAxis.splitNumber = NULL,
  xAxis.axisLabel.rotate = NULL,
  xAxis.axisLabel.margin = NULL,
  xAxis.axisLabel.color = NULL,
  xAxis.axisLabel.fontStyle = NULL,
  xAxis.axisLabel.fontWeight = NULL,
  xAxis.axisLabel.fontFamily = NULL,
  xAxis.axisLabel.fontSize = 14,
  xAxis.axisLabel.align = NULL,
  xAxis.axisLabel.verticalAlign = NULL,
  xAxis.axisLabel.backgroundColor = NULL,
  xAxis.axisLabel.borderColor = NULL,
  xAxis.axisLabel.borderWidth = NULL,
  xAxis.axisLabel.borderType = NULL,
  xAxis.axisLabel.borderRadius = NULL,
  xAxis.axisLabel.padding = NULL,
  xAxis.axisLabel.shadowColor = NULL,
  xAxis.axisLabel.shadowBlur = NULL,
  xAxis.axisLabel.shadowOffsetX = NULL,
  xAxis.axisLabel.shadowOffsetY = NULL,
  xAxis.axisLabel.textBorderColor = NULL,
  xAxis.axisLabel.textBorderWidth = NULL,
  xAxis.axisLabel.textBorderType = NULL,
  xAxis.axisLabel.textShadowColor = NULL,
  xAxis.axisLabel.textShadowBlur = NULL,
  xAxis.axisLabel.textShadowOffsetX = NULL,
  xAxis.axisLabel.textShadowOffsetY = NULL,
  xAxis.axisLabel.overflow = NULL,
  yAxis.title = NULL,
  yAxis.min = NULL,
  yAxis.max = NULL,
  yAxis.position = "center",
  yAxis.nameLocation = "center",
  yAxis.axisTick.customValues = NULL,
  yAxis.nameTextStyle.color = NULL,
  yAxis.nameTextStyle.padding = 25,
  yAxis.nameTextStyle.align = NULL,
  yAxis.nameTextStyle.fontStyle = NULL,
  yAxis.nameTextStyle.fontWeight = NULL,
  yAxis.nameTextStyle.fontSize = NULL,
  yAxis.nameTextStyle.fontFamily = NULL,
  yAxis.splitNumber = NULL,
  yAxis.axisLabel.rotate = NULL,
  yAxis.axisLabel.margin = NULL,
  yAxis.axisLabel.color = NULL,
  yAxis.axisLabel.fontStyle = NULL,
  yAxis.axisLabel.fontWeight = NULL,
  yAxis.axisLabel.fontFamily = NULL,
  yAxis.axisLabel.fontSize = 14,
  yAxis.axisLabel.align = NULL,
  yAxis.axisLabel.verticalAlign = NULL,
  yAxis.axisLabel.backgroundColor = NULL,
  yAxis.axisLabel.borderColor = NULL,
  yAxis.axisLabel.borderWidth = NULL,
  yAxis.axisLabel.borderType = NULL,
  yAxis.axisLabel.borderRadius = NULL,
  yAxis.axisLabel.padding = NULL,
  yAxis.axisLabel.shadowColor = NULL,
  yAxis.axisLabel.shadowBlur = NULL,
  yAxis.axisLabel.shadowOffsetX = NULL,
  yAxis.axisLabel.shadowOffsetY = NULL,
  yAxis.axisLabel.textBorderColor = NULL,
  yAxis.axisLabel.textBorderWidth = NULL,
  yAxis.axisLabel.textBorderType = NULL,
  yAxis.axisLabel.textShadowColor = NULL,
  yAxis.axisLabel.textShadowBlur = NULL,
  yAxis.axisLabel.textShadowOffsetX = NULL,
  yAxis.axisLabel.textShadowOffsetY = NULL,
  yAxis.axisLabel.overflow = NULL,
  legend.show = TRUE,
  legend.type = "scross",
  legend.selector = NULL,
  legend.icon = NULL,
  legend.align = NULL,
  legend.padding = NULL,
  legend.itemGap = NULL,
  legend.itemWidth = NULL,
  legend.orient = NULL,
  legend.width = NULL,
  legend.height = NULL,
  legend.left = NULL,
  legend.right = NULL,
  legend.top = NULL,
  legend.bottom = NULL,
  legend.backgroundColor = NULL,
  legend.borderColor = NULL,
  legend.borderWidth = NULL,
  legend.borderRadius = NULL,
  legend.shadowBlur = NULL,
  legend.shadowColor = NULL,
  legend.shadowOffsetX = NULL,
  legend.shadowOffsetY = NULL,
  legend.itemStyle.color = NULL,
  legend.itemStyle.borderColor = NULL,
  legend.itemStyle.borderWidth = NULL,
  legend.itemStyle.borderType = NULL,
  legend.itemStyle.shadowBlur = NULL,
  legend.itemStyle.shadowColor = NULL,
  legend.itemStyle.shadowOffsetX = NULL,
  legend.itemStyle.shadowOffsetY = NULL,
  legend.itemStyle.opacity = NULL,
  legend.lineStyle.color = NULL,
  legend.lineStyle.width = NULL,
  legend.lineStyle.type = NULL,
  legend.lineStyle.shadowBlur = NULL,
  legend.lineStyle.shadowColor = NULL,
  legend.lineStyle.shadowOffsetX = NULL,
  legend.lineStyle.shadowOffsetY = NULL,
  legend.lineStyle.opacity = NULL,
  legend.lineStyle.inactiveColor = NULL,
  legend.lineStyle.inactiveWidth = NULL,
  legend.textStyle.color = NULL,
  legend.textStyle.fontStyle = NULL,
  legend.textStyle.fontWeight = NULL,
  legend.textStyle.fontFamily = NULL,
  legend.textStyle.fontSize = NULL,
  legend.textStyle.backgroundColor = NULL,
  legend.textStyle.borderColor = NULL,
  legend.textStyle.borderWidth = NULL,
  legend.textStyle.borderType = NULL,
  legend.textStyle.borderRadius = NULL,
  legend.textStyle.padding = NULL,
  legend.textStyle.shadowColor = NULL,
  legend.textStyle.shadowBlur = NULL,
  legend.textStyle.shadowOffsetX = NULL,
  legend.textStyle.shadowOffsetY = NULL,
  legend.textStyle.width = NULL,
  legend.textStyle.height = NULL,
  legend.textStyle.textBorderColor = NULL,
  legend.textStyle.textBorderWidth = NULL,
  legend.textStyle.textBorderType = NULL,
  legend.textStyle.textShadowColor = NULL,
  legend.textStyle.textShadowBlur = NULL,
  legend.textStyle.textShadowOffsetX = NULL,
  legend.textStyle.textShadowOffsetY = NULL,
  legend.pageTextStyle.color = NULL,
  legend.pageTextStyle.fontStyle = NULL,
  legend.pageTextStyle.fontWeight = NULL,
  legend.pageTextStyle.fontFamily = NULL,
  legend.pageTextStyle.fontSize = NULL,
  legend.pageTextStyle.lineHeight = NULL,
  legend.pageTextStyle.width = NULL,
  legend.pageTextStyle.height = NULL,
  legend.pageTextStyle.textBorderColor = NULL,
  legend.pageTextStyle.textBorderWidth = NULL,
  legend.pageTextStyle.textBorderType = NULL,
  legend.pageTextStyle.textShadowColor = NULL,
  legend.pageTextStyle.textShadowBlur = NULL,
  legend.pageTextStyle.textShadowOffsetX = NULL,
  legend.pageTextStyle.textShadowOffsetY = NULL,
  legend.emphasis.selectorLabel.show = NULL,
  legend.emphasis.selectorLabel.distance = NULL,
  legend.emphasis.selectorLabel.rotate = NULL,
  legend.emphasis.selectorLabel.color = NULL,
  legend.emphasis.selectorLabel.fontStyle = NULL,
  legend.emphasis.selectorLabel.fontWeight = NULL,
  legend.emphasis.selectorLabel.fontFamily = NULL,
  legend.emphasis.selectorLabel.fontSize = NULL,
  legend.emphasis.selectorLabel.align = NULL,
  legend.emphasis.selectorLabel.verticalAlign = NULL,
  legend.emphasis.selectorLabel.lineHeight = NULL,
  legend.emphasis.selectorLabel.backgroundColor = NULL,
  legend.emphasis.selectorLabel.borderColor = NULL,
  legend.emphasis.selectorLabel.borderWidth = NULL,
  legend.emphasis.selectorLabel.borderType = NULL,
  legend.emphasis.selectorLabel.borderRadius = NULL,
  legend.emphasis.selectorLabel.padding = NULL,
  legend.emphasis.selectorLabel.shadowColor = NULL,
  legend.emphasis.selectorLabel.shadowBlur = NULL,
  legend.emphasis.selectorLabel.shadowOffsetX = NULL,
  legend.emphasis.selectorLabel.shadowOffsetY = NULL,
  legend.emphasis.selectorLabel.width = NULL,
  legend.emphasis.selectorLabel.height = NULL,
  legend.emphasis.selectorLabel.textBorderColor = NULL,
  legend.emphasis.selectorLabel.textBorderWidth = NULL,
  legend.emphasis.selectorLabel.textBorderType = NULL,
  legend.emphasis.selectorLabel.textShadowColor = NULL,
  legend.emphasis.selectorLabel.textShadowBlur = NULL,
  legend.emphasis.selectorLabel.textShadowOffsetX = NULL,
  legend.emphasis.selectorLabel.textShadowOffsetY = NULL,
  tooltip.show = TRUE,
  tooltip.trigger = "cross",
  tooltip.backgroundColor = NULL,
  tooltip.borderColor = NULL,
  tooltip.borderWidth = NULL,
  tooltip.padding = NULL,
  tooltip.axisPointer.type = "cross",
  tooltip.axisPointer.lineStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.shadowBlur = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetX = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetY = NULL,
  tooltip.axisPointer.shadowStyle.opacity = NULL,
  tooltip.textStyle.color = NULL,
  tooltip.textStyle.fontStyle = NULL,
  tooltip.textStyle.fontWeight = NULL,
  tooltip.textStyle.fontFamily = NULL,
  tooltip.textStyle.lineHeight = NULL,
  tooltip.textStyle.width = NULL,
  tooltip.textStyle.height = NULL,
  tooltip.textStyle.textBorderColor = NULL,
  tooltip.textStyle.textBorderWidth = NULL,
  tooltip.textStyle.textBorderType = NULL,
  tooltip.textStyle.textShadowColor = NULL,
  tooltip.textStyle.textShadowBlur = NULL,
  tooltip.textStyle.textShadowOffsetX = NULL,
  tooltip.textStyle.textShadowOffsetY = NULL,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{PreAgg}{logical. Is your data pre aggregated}

\item{AggMethod}{'mean', 'median', 'sum', 'sd', 'coeffvar', 'count'}

\item{XVar}{X-Axis variable name}

\item{YVar}{Y-Axis variable name}

\item{ZVar}{Z-Axis variable name}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{ZVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{NumberBins}{= 21}

\item{NumLevels_Y}{= 20}

\item{NumLevels_X}{= 20}

\item{Height}{"400px"}

\item{Width}{"200px"}

\item{ShowLabels}{character}

\item{Theme}{"dark-blue"}

\item{MouseScroll}{logical, zoom via mouse scroll}

\item{TextColor}{character}

\item{ContainLabel}{TRUE}

\item{title.text}{Title name}

\item{title.subtext}{Subtitle name}

\item{title.link}{Title as a link}

\item{title.sublink}{Subtitle as a link}

\item{title.Align}{'auto' 'left' 'right' 'center'}

\item{title.top}{'auto' '20' 'top' 'middle' 'bottom'}

\item{title.left}{distance between title and left side of container}

\item{title.right}{distance between title and right side of container}

\item{title.bottom}{'auto' '20' 'top' 'middle' 'bottom'}

\item{title.padding}{numeric}

\item{title.itemGap}{space between title and subtitle}

\item{title.backgroundColor}{hex or name}

\item{title.borderColor}{hex or name}

\item{title.borderWidth}{numeric}

\item{title.borderRadius}{numeric}

\item{title.shadowColor}{hex or name}

\item{title.shadowBlur}{numeric}

\item{title.shadowOffsetX}{numeric}

\item{title.shadowOffsetY}{numeric}

\item{title.textStyle.color}{hex or name}

\item{title.textStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{title.textStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{title.textStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{title.textStyle.fontSize}{numeric}

\item{title.textStyle.lineHeight}{numeric}

\item{title.textStyle.width}{numeric}

\item{title.textStyle.height}{numeric}

\item{title.textStyle.textBorderColor}{hex or name}

\item{title.textStyle.textBorderWidth}{numeric}

\item{title.textStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{title.textStyle.textBorderDashOffset}{numeric}

\item{title.textStyle.textShadowColor}{hex or name}

\item{title.textStyle.textShadowBlur}{numeric}

\item{title.textStyle.textShadowOffsetX}{numeric}

\item{title.textStyle.textShadowOffsetY}{numeric}

\item{title.subtextStyle.color}{hex or name}

\item{title.subtextStyle.align}{'auto' 'left' 'right' 'center'}

\item{title.subtextStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{title.subtextStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{title.subtextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{title.subtextStyle.fontSize}{numeric}

\item{title.subtextStyle.lineHeight}{numeric}

\item{title.subtextStyle.width}{numeric}

\item{title.subtextStyle.height}{numeric}

\item{title.subtextStyle.textBorderColor}{hex or name}

\item{title.subtextStyle.textBorderWidth}{numeric}

\item{title.subtextStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{title.subtextStyle.textBorderDashOffset}{numeric}

\item{title.subtextStyle.textShadowColor}{numeric}

\item{title.subtextStyle.textShadowBlur}{numeric}

\item{title.subtextStyle.textShadowOffsetX}{numeric}

\item{title.subtextStyle.textShadowOffsetY}{numeric}

\item{xAxis.title}{Axis title}

\item{xAxis.min}{Min value}

\item{xAxis.max}{Max value}

\item{xAxis.position}{'bottom', 'top'}

\item{xAxis.nameLocation}{'center', 'start', 'middle'}

\item{xAxis.axisTick.customValues}{Custom values for the x-axis}

\item{xAxis.nameTextStyle.color}{hex or named color}

\item{xAxis.nameTextStyle.padding}{numeric}

\item{xAxis.nameTextStyle.align}{'center', 'left', 'right'}

\item{xAxis.nameTextStyle.fontStyle}{'italic', 'normal' 'oblique'}

\item{xAxis.nameTextStyle.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{xAxis.nameTextStyle.fontSize}{numeric}

\item{xAxis.nameTextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{xAxis.splitNumber}{numeric. Increments for numeric axis labels}

\item{xAxis.axisLabel.rotate}{numeric}

\item{xAxis.axisLabel.margin}{numeric}

\item{xAxis.axisLabel.color}{hex or named}

\item{xAxis.axisLabel.fontStyle}{'italic', 'normal' 'oblique'}

\item{xAxis.axisLabel.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{xAxis.axisLabel.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{xAxis.axisLabel.fontSize}{numeric}

\item{xAxis.axisLabel.align}{'center', 'left', 'right'}

\item{xAxis.axisLabel.verticalAlign}{'top', 'bottom'}

\item{xAxis.axisLabel.backgroundColor}{hex or named}

\item{xAxis.axisLabel.borderColor}{hex or named}

\item{xAxis.axisLabel.borderWidth}{numeric}

\item{xAxis.axisLabel.borderType}{'solid', 'dashed', 'dotted'}

\item{xAxis.axisLabel.borderRadius}{numeric}

\item{xAxis.axisLabel.padding}{numeric}

\item{xAxis.axisLabel.shadowColor}{hex or named}

\item{xAxis.axisLabel.shadowBlur}{numeric}

\item{xAxis.axisLabel.shadowOffsetX}{numeric}

\item{xAxis.axisLabel.shadowOffsetY}{numeric}

\item{xAxis.axisLabel.textBorderColor}{hex or named}

\item{xAxis.axisLabel.textBorderWidth}{numeric}

\item{xAxis.axisLabel.textBorderType}{'solid', 'dashed', 'dotted'}

\item{xAxis.axisLabel.textShadowColor}{hex or named}

\item{xAxis.axisLabel.textShadowBlur}{numeric}

\item{xAxis.axisLabel.textShadowOffsetX}{numeric}

\item{xAxis.axisLabel.textShadowOffsetY}{numeric}

\item{xAxis.axisLabel.overflow}{'truncate'}

\item{yAxis.title}{Axis title}

\item{yAxis.min}{Min value}

\item{yAxis.max}{Max value}

\item{yAxis.position}{'bottom', 'top'}

\item{yAxis.nameLocation}{'center', 'start', 'middle'}

\item{yAxis.axisTick.customValues}{Custom values for the y-axis}

\item{yAxis.nameTextStyle.color}{hex or named color}

\item{yAxis.nameTextStyle.padding}{numeric}

\item{yAxis.nameTextStyle.align}{'center', 'left', 'right'}

\item{yAxis.nameTextStyle.fontStyle}{'italic', 'normal' 'oblique'}

\item{yAxis.nameTextStyle.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{yAxis.nameTextStyle.fontSize}{numeric}

\item{yAxis.nameTextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{yAxis.splitNumber}{numeric. Increments for numeric axis labels}

\item{yAxis.axisLabel.rotate}{numeric}

\item{yAxis.axisLabel.margin}{numeric}

\item{yAxis.axisLabel.color}{hex or named}

\item{yAxis.axisLabel.fontStyle}{'italic', 'normal' 'oblique'}

\item{yAxis.axisLabel.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{yAxis.axisLabel.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{yAxis.axisLabel.fontSize}{numeric}

\item{yAxis.axisLabel.align}{'center', 'left', 'right'}

\item{yAxis.axisLabel.verticalAlign}{'top', 'bottom'}

\item{yAxis.axisLabel.backgroundColor}{hex or named}

\item{yAxis.axisLabel.borderColor}{hex or named}

\item{yAxis.axisLabel.borderWidth}{numeric}

\item{yAxis.axisLabel.borderType}{'solid', 'dashed', 'dotted'}

\item{yAxis.axisLabel.borderRadius}{numeric}

\item{yAxis.axisLabel.padding}{numeric}

\item{yAxis.axisLabel.shadowColor}{hex or named}

\item{yAxis.axisLabel.shadowBlur}{numeric}

\item{yAxis.axisLabel.shadowOffsetX}{numeric}

\item{yAxis.axisLabel.shadowOffsetY}{numeric}

\item{yAxis.axisLabel.textBorderColor}{hex or named}

\item{yAxis.axisLabel.textBorderWidth}{numeric}

\item{yAxis.axisLabel.textBorderType}{'solid', 'dashed', 'dotted'}

\item{yAxis.axisLabel.textShadowColor}{hex or named}

\item{yAxis.axisLabel.textShadowBlur}{numeric}

\item{yAxis.axisLabel.textShadowOffsetX}{numeric}

\item{yAxis.axisLabel.textShadowOffsetY}{numeric}

\item{yAxis.axisLabel.overflow}{'truncate'}

\item{legend.show}{logical}

\item{legend.type}{'scroll' 'plain'}

\item{legend.selector}{logical}

\item{legend.icon}{'circle', 'rect', 'roundRect', 'triangle', 'diamond', 'pin', 'arrow', 'none'}

\item{legend.align}{'auto' 'left' 'right'}

\item{legend.padding}{numeric}

\item{legend.itemGap}{numeric}

\item{legend.itemWidth}{numeric}

\item{legend.orient}{'vertical' 'horizontal'}

\item{legend.width}{numeric}

\item{legend.height}{numeric}

\item{legend.left}{numeric}

\item{legend.right}{numeric}

\item{legend.top}{numeric}

\item{legend.bottom}{numeric}

\item{legend.backgroundColor}{hex or color name}

\item{legend.borderColor}{hex or color name}

\item{legend.borderWidth}{numeric}

\item{legend.borderRadius}{numeric}

\item{legend.shadowBlur}{numeric}

\item{legend.shadowColor}{hex or color name}

\item{legend.shadowOffsetX}{numeric}

\item{legend.shadowOffsetY}{numeric}

\item{legend.itemStyle.color}{hex or color name}

\item{legend.itemStyle.borderColor}{hex or color name}

\item{legend.itemStyle.borderWidth}{numeric}

\item{legend.itemStyle.borderType}{'solid' 'dashed' 'dotted'}

\item{legend.itemStyle.shadowBlur}{numeric}

\item{legend.itemStyle.shadowColor}{hex or color name}

\item{legend.itemStyle.shadowOffsetX}{numeric}

\item{legend.itemStyle.shadowOffsetY}{numeric}

\item{legend.itemStyle.opacity}{numeric 0 to 1}

\item{legend.lineStyle.color}{hex or color name}

\item{legend.lineStyle.width}{numeric}

\item{legend.lineStyle.type}{'solid' 'dashed' 'dotted'}

\item{legend.lineStyle.shadowBlur}{numeric}

\item{legend.lineStyle.shadowColor}{hex or color name}

\item{legend.lineStyle.shadowOffsetX}{numeric}

\item{legend.lineStyle.shadowOffsetY}{numeric}

\item{legend.lineStyle.opacity}{numeric 0 to 1}

\item{legend.lineStyle.inactiveColor}{hex or color name}

\item{legend.lineStyle.inactiveWidth}{numeric}

\item{legend.textStyle.color}{hex or color name}

\item{legend.textStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{legend.textStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{legend.textStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{legend.textStyle.fontSize}{numeric}

\item{legend.textStyle.backgroundColor}{hex or color name}

\item{legend.textStyle.borderColor}{hex or color name}

\item{legend.textStyle.borderWidth}{numeric}

\item{legend.textStyle.borderType}{'solid' 'dashed' 'dotted'}

\item{legend.textStyle.borderRadius}{numeric}

\item{legend.textStyle.padding}{numeric}

\item{legend.textStyle.shadowColor}{hex or color name}

\item{legend.textStyle.shadowBlur}{numeric}

\item{legend.textStyle.shadowOffsetX}{numeric}

\item{legend.textStyle.shadowOffsetY}{numeric}

\item{legend.textStyle.width}{numeric}

\item{legend.textStyle.height}{numeric}

\item{legend.textStyle.textBorderColor}{hex or color name}

\item{legend.textStyle.textBorderWidth}{numeric}

\item{legend.textStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{legend.textStyle.textShadowColor}{hex or color name}

\item{legend.textStyle.textShadowBlur}{numeric}

\item{legend.textStyle.textShadowOffsetX}{numeric}

\item{legend.textStyle.textShadowOffsetY}{numeric}

\item{legend.pageTextStyle.color}{hex or color name}

\item{legend.pageTextStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{legend.pageTextStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{legend.pageTextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{legend.pageTextStyle.fontSize}{numeric}

\item{legend.pageTextStyle.lineHeight}{numeric}

\item{legend.pageTextStyle.width}{numeric}

\item{legend.pageTextStyle.height}{numeric}

\item{legend.pageTextStyle.textBorderColor}{hex or color name}

\item{legend.pageTextStyle.textBorderWidth}{numeric}

\item{legend.pageTextStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{legend.pageTextStyle.textShadowColor}{hex or color name}

\item{legend.pageTextStyle.textShadowBlur}{numeric}

\item{legend.pageTextStyle.textShadowOffsetX}{numeric}

\item{legend.pageTextStyle.textShadowOffsetY}{numeric}

\item{legend.emphasis.selectorLabel.show}{logical}

\item{legend.emphasis.selectorLabel.distance}{numeric}

\item{legend.emphasis.selectorLabel.rotate}{numeric}

\item{legend.emphasis.selectorLabel.color}{hex or color name}

\item{legend.emphasis.selectorLabel.fontStyle}{'normal' 'italic' 'oblique'}

\item{legend.emphasis.selectorLabel.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{legend.emphasis.selectorLabel.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{legend.emphasis.selectorLabel.fontSize}{numeric}

\item{legend.emphasis.selectorLabel.align}{'left' 'center' 'right'}

\item{legend.emphasis.selectorLabel.verticalAlign}{'top' 'middle' 'bottom'}

\item{legend.emphasis.selectorLabel.lineHeight}{numeric}

\item{legend.emphasis.selectorLabel.backgroundColor}{hex or color name}

\item{legend.emphasis.selectorLabel.borderColor}{hex or color name}

\item{legend.emphasis.selectorLabel.borderWidth}{numeric}

\item{legend.emphasis.selectorLabel.borderType}{'solid' 'dashed' 'dotted'}

\item{legend.emphasis.selectorLabel.borderRadius}{numeric}

\item{legend.emphasis.selectorLabel.padding}{numeric}

\item{legend.emphasis.selectorLabel.shadowColor}{hex or color name}

\item{legend.emphasis.selectorLabel.shadowBlur}{numeric}

\item{legend.emphasis.selectorLabel.shadowOffsetX}{numeric}

\item{legend.emphasis.selectorLabel.shadowOffsetY}{numeric}

\item{legend.emphasis.selectorLabel.width}{numeric}

\item{legend.emphasis.selectorLabel.height}{numeric}

\item{legend.emphasis.selectorLabel.textBorderColor}{hex or color name}

\item{legend.emphasis.selectorLabel.textBorderWidth}{numeric}

\item{legend.emphasis.selectorLabel.textBorderType}{'solid' 'dashed' 'dotted'}

\item{legend.emphasis.selectorLabel.textShadowColor}{hex or color name}

\item{legend.emphasis.selectorLabel.textShadowBlur}{numeric}

\item{legend.emphasis.selectorLabel.textShadowOffsetX}{numeric}

\item{legend.emphasis.selectorLabel.textShadowOffsetY}{numeric}

\item{tooltip.show}{logical}

\item{tooltip.trigger}{"cross" "axis" "item" "none"}

\item{tooltip.backgroundColor}{hex or name}

\item{tooltip.borderColor}{numeric}

\item{tooltip.borderWidth}{numeric}

\item{tooltip.padding}{numeric}

\item{tooltip.axisPointer.type}{"line" or "shadow"}

\item{tooltip.axisPointer.lineStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.shadowBlur}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetX}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetY}{numeric}

\item{tooltip.axisPointer.shadowStyle.opacity}{numeric between 0 and 1}

\item{tooltip.textStyle.color}{hex or name}

\item{tooltip.textStyle.fontStyle}{"normal" "italic" "oblique"}

\item{tooltip.textStyle.fontWeight}{"normal" "bold" "bolder" "lighter"}

\item{tooltip.textStyle.fontFamily}{valid family name}

\item{tooltip.textStyle.lineHeight}{numeric}

\item{tooltip.textStyle.width}{numeric}

\item{tooltip.textStyle.height}{numeric}

\item{tooltip.textStyle.textBorderColor}{hex or name}

\item{tooltip.textStyle.textBorderWidth}{numeric}

\item{tooltip.textStyle.textBorderType}{"solid" "dashed" "dotted"}

\item{tooltip.textStyle.textShadowColor}{hex or name}

\item{tooltip.textStyle.textShadowBlur}{numeric}

\item{tooltip.textStyle.textShadowOffsetX}{numeric}

\item{tooltip.textStyle.textShadowOffsetY}{numeric}

\item{Debug}{Debugging purposes}
}
\value{
plot
}
\description{
Build a 3D Bar Plot
}
\examples{
# Create fake data
data <- AutoPlots::FakeDataGenerator(N = 100000)

# Echarts 3D Bar Chart
AutoPlots::BarPlot3D(
  dt = data,
  PreAgg = FALSE,
  AggMethod = 'mean',
  XVar = "Factor_1",
  YVar = "Factor_2",
  ZVar = "Adrian",
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  ZVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  NumberBins = 21,
  NumLevels_Y = 33,
  NumLevels_X = 33,
  Height = NULL,
  Width = NULL,
  ShowLabels = FALSE,
  Theme = "dark",
  MouseScroll = FALSE,
  TextColor = "black",
  ContainLabel = TRUE,
  Debug = FALSE)

}
\seealso{
Other Standard Plots: 
\code{\link{ACF}()},
\code{\link{Area}()},
\code{\link{Bar}()},
\code{\link{Box}()},
\code{\link{Copula3D}()},
\code{\link{Copula}()},
\code{\link{CorrMatrix}()},
\code{\link{Density}()},
\code{\link{Donut}()},
\code{\link{HeatMap}()},
\code{\link{Histogram}()},
\code{\link{Line}()},
\code{\link{PACF}()},
\code{\link{Parallel}()},
\code{\link{Pie}()},
\code{\link{Plot.ACF}()},
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.Donut}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.PACF}()},
\code{\link{Plot.Parallel}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.ProbabilityPlot}()},
\code{\link{Plot.Radar}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Rosetype}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.WordCloud}()},
\code{\link{ProbabilityPlot}()},
\code{\link{Radar}()},
\code{\link{River}()},
\code{\link{Rosetype}()},
\code{\link{Scatter3D}()},
\code{\link{Scatter}()},
\code{\link{StackedBar}()},
\code{\link{Step}()},
\code{\link{WordCloud}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
