% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions_NEW.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{ConfusionMatrix}
\usage{
ConfusionMatrix(
  dt = NULL,
  PreAgg = TRUE,
  XVar = NULL,
  YVar = NULL,
  ZVar = "N",
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  ZVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  NumberBins = 21,
  NumLevels_X = 50,
  NumLevels_Y = 50,
  Height = NULL,
  Width = NULL,
  ShowLabels = FALSE,
  Theme = "dark",
  MouseScroll = FALSE,
  TimeLine = FALSE,
  AggMethod = "count",
  GroupVar = NULL,
  title.text = NULL,
  yAxis.title = NULL,
  xAxis.title = NULL,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{PreAgg}{FALSE}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{ZVar}{= "N"}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{ZVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{NumberBins}{= 21,}

\item{NumLevels_X}{= NumLevels_Y,}

\item{NumLevels_Y}{= NumLevels_X,}

\item{Height}{"400px"}

\item{Width}{"200px"}

\item{ShowLabels}{character}

\item{Theme}{"auritus", "azul", "bee-inspired", "blue", "caravan", "carp", "chalk", "cool", "dark", "dark-blue", "dark-bold", "dark-digerati", "dark-fresh-cut", "dark-mushroom", "eduardo", "essos", "forest", "fresh-cut", "fruit", "gray", "green", "halloween", "helianthus", "infographic", "inspired", "jazz", "london", "macarons", "macarons2", "mint", "purple-passion", "rainbow", "red", "red-velvet", "roma", "royal", "sakura", "shine", "tech-blue", "v5", "vintage", "walden", "wef", "weforum", "westeros", "wonderland"}

\item{MouseScroll}{logical, zoom via mouse scroll}

\item{TimeLine}{logical}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{GroupVar}{= NULL}

\item{Debug}{Debugging purposes}
}
\value{
plot
}
\description{
Generate variable importance plots
}
\seealso{
Other Model Evaluation: 
\code{\link{BinaryMetricsPlot}()},
\code{\link{Calibration.Box}()},
\code{\link{Calibration.Line}()},
\code{\link{Gains}()},
\code{\link{ImportancePareto}()},
\code{\link{Lift}()},
\code{\link{PartialDependence.Box}()},
\code{\link{PartialDependence.HeatMap}()},
\code{\link{PartialDependence.Line}()},
\code{\link{Plot.BinaryMetrics}()},
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.HeatMap}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.ShapImportance}()},
\code{\link{Plot.VariableImportance}()},
\code{\link{Plot.prROC}()},
\code{\link{ROC}()},
\code{\link{Residuals.Histogram}()},
\code{\link{Residuals.Scatter}()},
\code{\link{ShapImportance}()},
\code{\link{VariableImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
