% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions_NEW.R
\name{Donut}
\alias{Donut}
\title{Donut}
\usage{
Donut(
  dt = NULL,
  PreAgg = FALSE,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  AggMethod = "mean",
  Height = NULL,
  Width = NULL,
  ShowLabels = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  Theme = "dark",
  TimeLine = TRUE,
  TextColor = "white",
  title.text = "Donut Plot",
  title.subtext = NULL,
  title.link = NULL,
  title.sublink = NULL,
  title.Align = NULL,
  title.top = NULL,
  title.left = NULL,
  title.right = NULL,
  title.bottom = NULL,
  title.padding = NULL,
  title.itemGap = NULL,
  title.backgroundColor = NULL,
  title.borderColor = NULL,
  title.borderWidth = NULL,
  title.borderRadius = NULL,
  title.shadowColor = NULL,
  title.shadowBlur = NULL,
  title.shadowOffsetX = NULL,
  title.shadowOffsetY = NULL,
  title.textStyle.color = NULL,
  title.textStyle.fontStyle = NULL,
  title.textStyle.fontWeight = NULL,
  title.textStyle.fontFamily = NULL,
  title.textStyle.fontSize = NULL,
  title.textStyle.lineHeight = NULL,
  title.textStyle.width = NULL,
  title.textStyle.height = NULL,
  title.textStyle.textBorderColor = NULL,
  title.textStyle.textBorderWidth = NULL,
  title.textStyle.textBorderType = NULL,
  title.textStyle.textBorderDashOffset = NULL,
  title.textStyle.textShadowColor = NULL,
  title.textStyle.textShadowBlur = NULL,
  title.textStyle.textShadowOffsetX = NULL,
  title.textStyle.textShadowOffsetY = NULL,
  title.subtextStyle.color = NULL,
  title.subtextStyle.align = NULL,
  title.subtextStyle.fontStyle = NULL,
  title.subtextStyle.fontWeight = NULL,
  title.subtextStyle.fontFamily = NULL,
  title.subtextStyle.fontSize = NULL,
  title.subtextStyle.lineHeight = NULL,
  title.subtextStyle.width = NULL,
  title.subtextStyle.height = NULL,
  title.subtextStyle.textBorderColor = NULL,
  title.subtextStyle.textBorderWidth = NULL,
  title.subtextStyle.textBorderType = NULL,
  title.subtextStyle.textBorderDashOffset = NULL,
  title.subtextStyle.textShadowColor = NULL,
  title.subtextStyle.textShadowBlur = NULL,
  title.subtextStyle.textShadowOffsetX = NULL,
  title.subtextStyle.textShadowOffsetY = NULL,
  xtooltip.show = TRUE,
  tooltip.trigger = "item",
  tooltip.backgroundColor = NULL,
  tooltip.borderColor = NULL,
  tooltip.borderWidth = NULL,
  tooltip.padding = NULL,
  tooltip.axisPointer.type = "cross",
  tooltip.axisPointer.lineStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.shadowBlur = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetX = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetY = NULL,
  tooltip.axisPointer.shadowStyle.opacity = NULL,
  tooltip.textStyle.color = NULL,
  tooltip.textStyle.fontStyle = NULL,
  tooltip.textStyle.fontWeight = NULL,
  tooltip.textStyle.fontFamily = NULL,
  tooltip.textStyle.lineHeight = NULL,
  tooltip.textStyle.width = NULL,
  tooltip.textStyle.height = NULL,
  tooltip.textStyle.textBorderColor = NULL,
  tooltip.textStyle.textBorderWidth = NULL,
  tooltip.textStyle.textBorderType = NULL,
  tooltip.textStyle.textShadowColor = NULL,
  tooltip.textStyle.textShadowBlur = NULL,
  tooltip.textStyle.textShadowOffsetX = NULL,
  tooltip.textStyle.textShadowOffsetY = NULL,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{PreAgg}{logical}

\item{XVar}{X-Axis variable name}

\item{YVar}{Y-Axis variable name}

\item{GroupVar}{Column name of Group Variable for distinct colored histograms by group levels}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{Height}{"400px"}

\item{Width}{"200px"}

\item{ShowLabels}{character}

\item{Title.YAxis}{character}

\item{Title.XAxis}{character}

\item{Theme}{"auritus","azul","bee-inspired","blue","caravan","carp","chalk","cool","dark-bold","dark","eduardo","essos","forest","fresh-cut","fruit","gray","green","halloween","helianthus","infographic","inspired","jazz","london","dark","macarons","macarons2","mint","purple-passion","red-velvet","red","roma","royal","sakura","shine","tech-blue","vintage","walden","wef","weforum","westeros","wonderland"}

\item{TimeLine}{logical}

\item{TextColor}{'darkblue'}

\item{title.text}{Title name}

\item{title.subtext}{Subtitle name}

\item{title.link}{Title as a link}

\item{title.sublink}{Subtitle as a link}

\item{title.Align}{'auto' 'left' 'right' 'center'}

\item{title.top}{'auto' '20' '20%' 'top' 'middle' 'bottom'}

\item{title.left}{distance between title and left side of container}

\item{title.right}{distance between title and right side of container}

\item{title.bottom}{'auto' '20' '20%' 'top' 'middle' 'bottom'}

\item{title.padding}{numeric}

\item{title.itemGap}{space between title and subtitle}

\item{title.backgroundColor}{hex or name}

\item{title.borderColor}{hex or name}

\item{title.borderWidth}{numeric}

\item{title.borderRadius}{numeric}

\item{title.shadowColor}{hex or name}

\item{title.shadowBlur}{numeric}

\item{title.shadowOffsetX}{numeric}

\item{title.shadowOffsetY}{numeric}

\item{title.textStyle.color}{hex or name}

\item{title.textStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{title.textStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{title.textStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{title.textStyle.fontSize}{numeric}

\item{title.textStyle.lineHeight}{numeric}

\item{title.textStyle.width}{numeric}

\item{title.textStyle.height}{numeric}

\item{title.textStyle.textBorderColor}{hex or name}

\item{title.textStyle.textBorderWidth}{numeric}

\item{title.textStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{title.textStyle.textBorderDashOffset}{numeric}

\item{title.textStyle.textShadowColor}{hex or name}

\item{title.textStyle.textShadowBlur}{numeric}

\item{title.textStyle.textShadowOffsetX}{numeric}

\item{title.textStyle.textShadowOffsetY}{numeric}

\item{title.subtextStyle.color}{hex or name}

\item{title.subtextStyle.align}{'auto' 'left' 'right' 'center'}

\item{title.subtextStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{title.subtextStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{title.subtextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{title.subtextStyle.fontSize}{numeric}

\item{title.subtextStyle.lineHeight}{numeric}

\item{title.subtextStyle.width}{numeric}

\item{title.subtextStyle.height}{numeric}

\item{title.subtextStyle.textBorderColor}{hex or name}

\item{title.subtextStyle.textBorderWidth}{numeric}

\item{title.subtextStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{title.subtextStyle.textBorderDashOffset}{numeric}

\item{title.subtextStyle.textShadowColor}{numeric}

\item{title.subtextStyle.textShadowBlur}{numeric}

\item{title.subtextStyle.textShadowOffsetX}{numeric}

\item{title.subtextStyle.textShadowOffsetY}{numeric}

\item{tooltip.trigger}{"cross" "axis" "item" "none"}

\item{tooltip.backgroundColor}{hex or name}

\item{tooltip.borderColor}{numeric}

\item{tooltip.borderWidth}{numeric}

\item{tooltip.padding}{numeric}

\item{tooltip.axisPointer.type}{"line" or "shadow"}

\item{tooltip.axisPointer.lineStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.shadowBlur}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetX}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetY}{numeric}

\item{tooltip.axisPointer.shadowStyle.opacity}{numeric between 0 and 1}

\item{tooltip.textStyle.color}{hex or name}

\item{tooltip.textStyle.fontStyle}{"normal" "italic" "oblique"}

\item{tooltip.textStyle.fontWeight}{"normal" "bold" "bolder" "lighter"}

\item{tooltip.textStyle.fontFamily}{valid family name}

\item{tooltip.textStyle.lineHeight}{numeric}

\item{tooltip.textStyle.width}{numeric}

\item{tooltip.textStyle.height}{numeric}

\item{tooltip.textStyle.textBorderColor}{hex or name}

\item{tooltip.textStyle.textBorderWidth}{numeric}

\item{tooltip.textStyle.textBorderType}{"solid" "dashed" "dotted"}

\item{tooltip.textStyle.textShadowColor}{hex or name}

\item{tooltip.textStyle.textShadowBlur}{numeric}

\item{tooltip.textStyle.textShadowOffsetX}{numeric}

\item{tooltip.textStyle.textShadowOffsetY}{numeric}

\item{Debug}{Debugging purposes}

\item{tooltip.show}{logical}
}
\value{
plot
}
\description{
Build a donut plot by simply passing arguments to a single function
}
\examples{

# Create fake data
dt <- data.table::data.table(Y = qnorm(p = runif(10000)), GV = sample(LETTERS, 1000, TRUE))

# Create plot
AutoPlots::Donut(
  dt = dt,
  PreAgg = FALSE,
  XVar = "GV",
  YVar = "Y",
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  AggMethod = 'mean',
  Height = NULL,
  Width = NULL,
  Title = 'Pie Chart',
  ShowLabels = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  Theme = "dark",
  TimeLine = TRUE,
  TextColor = "black",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = '#63aeff',
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  xaxis.fontSize = 14,
  yaxis.fontSize = 14,
  Debug = FALSE)

}
\seealso{
Other Standard Plots: 
\code{\link{ACF}()},
\code{\link{Area}()},
\code{\link{BarPlot3D}()},
\code{\link{Bar}()},
\code{\link{Box}()},
\code{\link{Copula3D}()},
\code{\link{Copula}()},
\code{\link{CorrMatrix}()},
\code{\link{Density}()},
\code{\link{HeatMap}()},
\code{\link{Histogram}()},
\code{\link{Line}()},
\code{\link{PACF}()},
\code{\link{Parallel}()},
\code{\link{Pie}()},
\code{\link{Plot.ACF}()},
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.Donut}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.PACF}()},
\code{\link{Plot.Parallel}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.ProbabilityPlot}()},
\code{\link{Plot.Radar}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Rosetype}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.WordCloud}()},
\code{\link{ProbabilityPlot}()},
\code{\link{Radar}()},
\code{\link{River}()},
\code{\link{Rosetype}()},
\code{\link{Scatter3D}()},
\code{\link{Scatter}()},
\code{\link{StackedBar}()},
\code{\link{Step}()},
\code{\link{WordCloud}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
