% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessoryFunctions.R
\name{DT_GDL_Feature_Engineering}
\alias{DT_GDL_Feature_Engineering}
\title{DT_GDL_Feature_Engineering}
\usage{
DT_GDL_Feature_Engineering(
  data,
  lags = 1,
  periods = 0,
  SDperiods = 0,
  Skewperiods = 0,
  Kurtperiods = 0,
  Quantileperiods = 0,
  statsFUNs = c("mean"),
  targets = NULL,
  groupingVars = NULL,
  sortDateName = NULL,
  timeDiffTarget = NULL,
  timeAgg = c("days"),
  WindowingLag = 0,
  ShortName = TRUE,
  Type = c("Lag"),
  SimpleImpute = TRUE
)
}
\arguments{
\item{data}{A data.table you want to run the function on}

\item{lags}{A numeric vector of the specific lags you want to have generated. You must include 1 if WindowingLag = 1.}

\item{periods}{A numeric vector of the specific rolling statistics window sizes you want to utilize in the calculations.}

\item{SDperiods}{A numeric vector of Standard Deviation rolling statistics window sizes you want to utilize in the calculations.}

\item{Skewperiods}{A numeric vector of Skewness rolling statistics window sizes you want to utilize in the calculations.}

\item{Kurtperiods}{A numeric vector of Kurtosis rolling statistics window sizes you want to utilize in the calculations.}

\item{Quantileperiods}{A numeric vector of Quantile rolling statistics window sizes you want to utilize in the calculations.}

\item{statsFUNs}{Select from the following c("mean","sd","skew","kurt","q5","q10","q15","q20","q25","q30","q35","q40","q45","q50","q55","q60","q65","q70","q75","q80","q85","q90","q95")}

\item{targets}{A character vector of the column names for the reference column in which you will build your lags and rolling stats}

\item{groupingVars}{A character vector of categorical variable names you will build your lags and rolling stats by}

\item{sortDateName}{The column name of your date column used to sort events over time}

\item{timeDiffTarget}{Specify a desired name for features created for time between events. Set to NULL if you don't want time between events features created.}

\item{timeAgg}{List the time aggregation level for the time between events features, such as "hour", "day", "week", "month", "quarter", or "year"}

\item{WindowingLag}{Set to 0 to build rolling stats off of target columns directly or set to 1 to build the rolling stats off of the lag-1 target}

\item{ShortName}{Default TRUE. If FALSE, Group Variable names will be added to the rolling stat and lag names. If you plan on have multiple versions of lags and rollings stats by different group variables then set this to FALSE.}

\item{Type}{List either "Lag" if you want features built on historical values or "Lead" if you want features built on future values}

\item{SimpleImpute}{Set to TRUE for factor level imputation of "0" and numeric imputation of -1}
}
\value{
data.table of original data plus created lags, rolling stats, and time between event lags and rolling stats
}
\description{
Builds autoregressive and moving average from target columns and distributed lags and distributed moving average for independent features distributed across time. On top of that, you can also create time between instances along with their associated lags and moving averages. This function works for data with groups and without groups.
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoLagRollStats}()},
\code{\link{DummifyDT}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
