% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_area_full}
\alias{e_area_full}
\title{Enhanced area Setter for echarts4r}
\usage{
e_area_full(
  e = NULL,
  serie = NULL,
  smooth = NULL,
  showSymbol = NULL,
  areaStyle.color = NULL,
  areaStyle.opacity = NULL
)
}
\arguments{
\item{e}{plot object}

\item{serie}{Variable}

\item{smooth}{Smooth line}

\item{showSymbol}{Logical}

\item{areaStyle.color}{Fill color. Can be a single color or vector of multiple colors for gradient.}

\item{areaStyle.opacity}{transparency}
}
\value{
The modified echarts4r object
}
\description{
Exposes every area* option so you don't have to hand-craft the JSON.
}
