% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_boxplot_full}
\alias{e_boxplot_full}
\title{Enhanced area Setter for echarts4r}
\usage{
e_boxplot_full(
  e = NULL,
  serie = NULL,
  label = FALSE,
  itemStyle.color = NULL,
  itemStyle.opacity = NULL,
  itemStyle.borderColor = NULL,
  itemStyle.borderWidth = NULL,
  itemStyle.borderType = NULL,
  itemStyle.borderCap = NULL,
  itemStyle.shadowBlur = NULL,
  itemStyle.shadowColor = NULL,
  itemStyle.shadowOffsetX = NULL,
  itemStyle.shadowOffsetY = NULL
)
}
\arguments{
\item{e}{plot object}

\item{serie}{Variable}

\item{label}{Logical}

\item{itemStyle.color}{Fill color. Can be a single color or vector of multiple colors for gradient.}

\item{itemStyle.opacity}{transparency}

\item{itemStyle.borderColor}{hex or color name}

\item{itemStyle.borderWidth}{numeric}

\item{itemStyle.borderType}{'solid' 'dashed' 'dotted'}

\item{itemStyle.borderCap}{'butt' 'round' 'square'}

\item{itemStyle.shadowBlur}{numeric}

\item{itemStyle.shadowColor}{hex or name}

\item{itemStyle.shadowOffsetX}{numeric}

\item{itemStyle.shadowOffsetY}{numeric}
}
\value{
The modified echarts4r object
}
\description{
Exposes every area* option so you don't have to hand-craft the JSON.
}
