% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_data_zoom_slider_full}
\alias{e_data_zoom_slider_full}
\title{Enhanced DataZoom Slider Setter for echarts4r}
\usage{
e_data_zoom_slider_full(
  e,
  index = NULL,
  dataZoom.id = NULL,
  dataZoom.show = TRUE,
  dataZoom.backgroundColor = NULL,
  dataZoom.fillerColor = NULL,
  dataZoom.borderColor = NULL,
  dataZoom.borderRadius = NULL,
  dataZoom.handleIcon = NULL,
  dataZoom.handleSize = NULL,
  dataZoom.handleLabel.show = NULL,
  dataZoom.moveHandleIcon = NULL,
  dataZoom.moveHandleSize = NULL,
  dataZoom.labelPrecision = NULL,
  dataZoom.labelFormatter = NULL,
  dataZoom.showDetail = NULL,
  dataZoom.showDataShadow = NULL,
  dataZoom.realtime = NULL,
  dataZoom.xAxisIndex = 0,
  dataZoom.yAxisIndex = NULL,
  dataZoom.radiusAxisIndex = NULL,
  dataZoom.angleAxisIndex = NULL,
  dataZoom.filterMode = NULL,
  dataZoom.start = NULL,
  dataZoom.end = NULL,
  dataZoom.startValue = NULL,
  dataZoom.endValue = NULL,
  dataZoom.minSpan = NULL,
  dataZoom.maxSpan = NULL,
  dataZoom.minValueSpan = NULL,
  dataZoom.maxValueSpan = NULL,
  dataZoom.orient = "horizontal",
  dataZoom.zoomLock = NULL,
  dataZoom.throttle = NULL,
  dataZoom.rangeMode = NULL,
  dataZoom.zlevel = NULL,
  dataZoom.z = NULL,
  dataZoom.left = NULL,
  dataZoom.top = NULL,
  dataZoom.right = NULL,
  dataZoom.bottom = NULL,
  dataZoom.width = NULL,
  dataZoom.height = NULL,
  dataZoom.brushSelect = NULL,
  dataZoom.dataBackground.lineStyle.color = NULL,
  dataZoom.dataBackground.lineStyle.width = NULL,
  dataZoom.dataBackground.lineStyle.type = NULL,
  dataZoom.dataBackground.lineStyle.dashOffset = NULL,
  dataZoom.dataBackground.lineStyle.cap = NULL,
  dataZoom.dataBackground.lineStyle.join = NULL,
  dataZoom.dataBackground.lineStyle.miterLimit = NULL,
  dataZoom.dataBackground.lineStyle.shadowBlur = NULL,
  dataZoom.dataBackground.lineStyle.shadowColor = NULL,
  dataZoom.dataBackground.lineStyle.shadowOffsetX = NULL,
  dataZoom.dataBackground.lineStyle.shadowOffsetY = NULL,
  dataZoom.dataBackground.lineStyle.opacity = NULL,
  dataZoom.dataBackground.areaStyle.color = NULL,
  dataZoom.dataBackground.areaStyle.shadowBlur = NULL,
  dataZoom.dataBackground.areaStyle.shadowColor = NULL,
  dataZoom.dataBackground.areaStyle.shadowOffsetX = NULL,
  dataZoom.dataBackground.areaStyle.shadowOffsetY = NULL,
  dataZoom.dataBackground.areaStyle.opacity = NULL,
  dataZoom.selectedDataBackground.lineStyle.color = NULL,
  dataZoom.selectedDataBackground.lineStyle.width = NULL,
  dataZoom.selectedDataBackground.lineStyle.type = NULL,
  dataZoom.selectedDataBackground.lineStyle.dashOffset = NULL,
  dataZoom.selectedDataBackground.lineStyle.cap = NULL,
  dataZoom.selectedDataBackground.lineStyle.join = NULL,
  dataZoom.selectedDataBackground.lineStyle.miterLimit = NULL,
  dataZoom.selectedDataBackground.lineStyle.shadowBlur = NULL,
  dataZoom.selectedDataBackground.lineStyle.shadowColor = NULL,
  dataZoom.selectedDataBackground.lineStyle.shadowOffsetX = NULL,
  dataZoom.selectedDataBackground.lineStyle.shadowOffsetY = NULL,
  dataZoom.selectedDataBackground.lineStyle.opacity = NULL,
  dataZoom.selectedDataBackground.areaStyle.color = NULL,
  dataZoom.selectedDataBackground.areaStyle.shadowBlur = NULL,
  dataZoom.selectedDataBackground.areaStyle.shadowColor = NULL,
  dataZoom.selectedDataBackground.areaStyle.shadowOffsetX = NULL,
  dataZoom.selectedDataBackground.areaStyle.shadowOffsetY = NULL,
  dataZoom.selectedDataBackground.areaStyle.opacity = NULL,
  dataZoom.handleStyle.color = NULL,
  dataZoom.handleStyle.borderColor = NULL,
  dataZoom.handleStyle.borderWidth = NULL,
  dataZoom.handleStyle.borderType = NULL,
  dataZoom.handleStyle.borderDashOffset = NULL,
  dataZoom.handleStyle.borderCap = NULL,
  dataZoom.handleStyle.borderJoin = NULL,
  dataZoom.handleStyle.borderMiterLimit = NULL,
  dataZoom.handleStyle.shadowBlur = NULL,
  dataZoom.handleStyle.shadowColor = NULL,
  dataZoom.handleStyle.shadowOffsetX = NULL,
  dataZoom.handleStyle.shadowOffsetY = NULL,
  dataZoom.handleStyle.opacity = NULL,
  dataZoom.moveHandleStyle.color = NULL,
  dataZoom.moveHandleStyle.borderColor = NULL,
  dataZoom.moveHandleStyle.borderWidth = NULL,
  dataZoom.moveHandleStyle.borderType = NULL,
  dataZoom.moveHandleStyle.borderDashOffset = NULL,
  dataZoom.moveHandleStyle.borderCap = NULL,
  dataZoom.moveHandleStyle.borderJoin = NULL,
  dataZoom.moveHandleStyle.borderMiterLimit = NULL,
  dataZoom.moveHandleStyle.shadowBlur = NULL,
  dataZoom.moveHandleStyle.shadowColor = NULL,
  dataZoom.moveHandleStyle.shadowOffsetX = NULL,
  dataZoom.moveHandleStyle.shadowOffsetY = NULL,
  dataZoom.moveHandleStyle.opacity = NULL,
  dataZoom.textStyle.color = NULL,
  dataZoom.textStyle.fontStyle = NULL,
  dataZoom.textStyle.fontWeight = NULL,
  dataZoom.textStyle.fontFamily = NULL,
  dataZoom.textStyle.fontSize = NULL,
  dataZoom.textStyle.lineHeight = NULL,
  dataZoom.textStyle.width = NULL,
  dataZoom.textStyle.height = NULL,
  dataZoom.textStyle.textBorderColor = NULL,
  dataZoom.textStyle.textBorderWidth = NULL,
  dataZoom.textStyle.textBorderType = NULL,
  dataZoom.textStyle.textBorderDashOffset = NULL,
  dataZoom.textStyle.textShadowColor = NULL,
  dataZoom.textStyle.textShadowBlur = NULL,
  dataZoom.textStyle.textShadowOffsetX = NULL,
  dataZoom.textStyle.textShadowOffsetY = NULL,
  dataZoom.textStyle.overflow = NULL,
  dataZoom.textStyle.ellipsis = NULL,
  dataZoom.brushStyle.color = NULL,
  dataZoom.brushStyle.borderColor = NULL,
  dataZoom.brushStyle.borderWidth = NULL,
  dataZoom.brushStyle.borderType = NULL,
  dataZoom.brushStyle.borderDashOffset = NULL,
  dataZoom.brushStyle.borderCap = NULL,
  dataZoom.brushStyle.borderJoin = NULL,
  dataZoom.brushStyle.borderMiterLimit = NULL,
  dataZoom.brushStyle.shadowBlur = NULL,
  dataZoom.brushStyle.shadowColor = NULL,
  dataZoom.brushStyle.shadowOffsetX = NULL,
  dataZoom.brushStyle.shadowOffsetY = NULL,
  dataZoom.brushStyle.opacity = NULL,
  dataZoom.emphasis.handleStyle.color = NULL,
  dataZoom.emphasis.handleStyle.borderColor = NULL,
  dataZoom.emphasis.handleStyle.borderWidth = NULL,
  dataZoom.emphasis.handleStyle.borderType = NULL,
  dataZoom.emphasis.handleStyle.opacity = NULL,
  dataZoom.emphasis.handleLabel.show = NULL,
  dataZoom.emphasis.moveHandleStyle.color = NULL,
  dataZoom.emphasis.moveHandleStyle.borderColor = NULL,
  dataZoom.emphasis.moveHandleStyle.borderWidth = NULL,
  dataZoom.emphasis.moveHandleStyle.borderType = NULL,
  dataZoom.emphasis.moveHandleStyle.opacity = NULL
)
}
\arguments{
\item{e}{An echarts4r object}

\item{index}{Which dataZoom component to target, zero-based. If NULL, append new slider.}

\item{dataZoom.id}{Component ID}

\item{dataZoom.show}{Whether to show the slider}

\item{dataZoom.backgroundColor}{Background color}

\item{dataZoom.fillerColor}{Color to fill selected area}

\item{dataZoom.borderColor}{Border color}

\item{dataZoom.borderRadius}{Border radius, numeric or array}

\item{dataZoom.handleIcon}{Handle icon path/image}

\item{dataZoom.handleSize}{Handle size}

\item{dataZoom.handleLabel.show}{Whether to show handle label}

\item{dataZoom.moveHandleIcon}{Move handle icon path/image}

\item{dataZoom.moveHandleSize}{Move handle size}

\item{dataZoom.labelPrecision}{Label precision, numeric or 'auto'}

\item{dataZoom.labelFormatter}{Label formatter, string or JS function}

\item{dataZoom.showDetail}{Whether to show detail while dragging}

\item{dataZoom.showDataShadow}{Whether to show data shadow}

\item{dataZoom.realtime}{Whether to update view while dragging}

\item{dataZoom.xAxisIndex}{xAxis index or array of indices controlled by this slider}

\item{dataZoom.yAxisIndex}{yAxis index or array of indices controlled by this slider}

\item{dataZoom.radiusAxisIndex}{radiusAxis index or array of indices}

\item{dataZoom.angleAxisIndex}{angleAxis index or array of indices}

\item{dataZoom.filterMode}{'filter', 'weakFilter', 'empty', 'none'}

\item{dataZoom.start}{Start percentage, 0 to 100}

\item{dataZoom.end}{End percentage, 0 to 100}

\item{dataZoom.startValue}{Start absolute value}

\item{dataZoom.endValue}{End absolute value}

\item{dataZoom.minSpan}{Minimum window size in percent}

\item{dataZoom.maxSpan}{Maximum window size in percent}

\item{dataZoom.minValueSpan}{Minimum window size in value units}

\item{dataZoom.maxValueSpan}{Maximum window size in value units}

\item{dataZoom.orient}{'horizontal' or 'vertical'}

\item{dataZoom.zoomLock}{Whether to lock window size}

\item{dataZoom.throttle}{Refresh throttle in milliseconds}

\item{dataZoom.rangeMode}{Array such as c('value', 'percent')}

\item{dataZoom.zlevel}{Canvas layer zlevel}

\item{dataZoom.z}{Draw order z}

\item{dataZoom.left}{Left position}

\item{dataZoom.top}{Top position}

\item{dataZoom.right}{Right position}

\item{dataZoom.bottom}{Bottom position}

\item{dataZoom.width}{Width}

\item{dataZoom.height}{Height}

\item{dataZoom.brushSelect}{Whether to enable brush selection}

\item{dataZoom.dataBackground.lineStyle.color}{Data shadow line color}

\item{dataZoom.dataBackground.lineStyle.width}{Data shadow line width}

\item{dataZoom.dataBackground.lineStyle.type}{Data shadow line type}

\item{dataZoom.dataBackground.areaStyle.color}{Data shadow area color}

\item{dataZoom.selectedDataBackground.lineStyle.color}{Selected shadow line color}

\item{dataZoom.selectedDataBackground.lineStyle.width}{Selected shadow line width}

\item{dataZoom.selectedDataBackground.lineStyle.type}{Selected shadow line type}

\item{dataZoom.selectedDataBackground.areaStyle.color}{Selected shadow area color}

\item{dataZoom.handleStyle.color}{Handle color}

\item{dataZoom.handleStyle.borderColor}{Handle border color}

\item{dataZoom.handleStyle.borderWidth}{Handle border width}

\item{dataZoom.handleStyle.borderType}{Handle border type}

\item{dataZoom.handleStyle.shadowBlur}{Handle shadow blur}

\item{dataZoom.handleStyle.shadowColor}{Handle shadow color}

\item{dataZoom.handleStyle.shadowOffsetX}{Handle shadow X offset}

\item{dataZoom.handleStyle.shadowOffsetY}{Handle shadow Y offset}

\item{dataZoom.handleStyle.opacity}{Handle opacity}

\item{dataZoom.moveHandleStyle.color}{Move handle color}

\item{dataZoom.moveHandleStyle.borderColor}{Move handle border color}

\item{dataZoom.moveHandleStyle.borderWidth}{Move handle border width}

\item{dataZoom.moveHandleStyle.borderType}{Move handle border type}

\item{dataZoom.moveHandleStyle.opacity}{Move handle opacity}

\item{dataZoom.textStyle.color}{Text color}

\item{dataZoom.textStyle.fontStyle}{Text font style}

\item{dataZoom.textStyle.fontWeight}{Text font weight}

\item{dataZoom.textStyle.fontFamily}{Text font family}

\item{dataZoom.textStyle.fontSize}{Text font size}

\item{dataZoom.textStyle.lineHeight}{Text line height}

\item{dataZoom.textStyle.width}{Text block width}

\item{dataZoom.textStyle.height}{Text block height}

\item{dataZoom.textStyle.overflow}{Text overflow mode}

\item{dataZoom.textStyle.ellipsis}{Text ellipsis}

\item{dataZoom.brushStyle.color}{Brush color}

\item{dataZoom.brushStyle.borderColor}{Brush border color}

\item{dataZoom.brushStyle.borderWidth}{Brush border width}

\item{dataZoom.brushStyle.borderType}{Brush border type}

\item{dataZoom.brushStyle.opacity}{Brush opacity}
}
\value{
The modified echarts4r object
}
\description{
Exposes dataZoom-slider.* options so you don't have to hand-craft the JSON.
}
