% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_heatmap_full}
\alias{e_heatmap_full}
\title{Heatmap customization}
\usage{
e_heatmap_full(
  e,
  y,
  z,
  itemStyle.borderColor = NULL,
  itemStyle.borderWidth = NULL,
  itemStyle.shadowColor = NULL,
  itemStyle.shadowBlur = NULL,
  itemStyle.shadowOffsetY = NULL,
  itemStyle.shadowOffsetX = NULL,
  itemStyle.opacity = NULL,
  itemStyle.borderRadius = NULL,
  emphasis.shadowColor = NULL,
  emphasis.shadowBlur = NULL,
  itemStyle.emphasis.label.show = NULL,
  itemStyle.emphasis.label.color = NULL,
  itemStyle.emphasis.label.fontStyle = NULL,
  itemStyle.emphasis.label.fontWeight = NULL,
  itemStyle.emphasis.label.fontFamily = NULL,
  itemStyle.emphasis.label.fontSize = NULL,
  itemStyle.emphasis.label.align = NULL,
  itemStyle.emphasis.label.verticalAlign = NULL,
  itemStyle.emphasis.label.backgroundColor = NULL,
  itemStyle.emphasis.label.borderColor = NULL,
  itemStyle.emphasis.label.borderWidth = NULL,
  itemStyle.emphasis.label.borderType = NULL,
  itemStyle.emphasis.label.borderRadius = NULL,
  itemStyle.emphasis.label.shadowColor = NULL,
  itemStyle.emphasis.label.shadowBlur = NULL,
  itemStyle.emphasis.label.shadowOffsetY = NULL,
  itemStyle.emphasis.label.shadowOffsetX = NULL,
  itemStyle.emphasis.label.width = NULL,
  itemStyle.emphasis.label.height = NULL,
  itemStyle.emphasis.label.textBorderColor = NULL,
  itemStyle.emphasis.label.textBorderWidth = NULL,
  itemStyle.emphasis.label.textShadowColor = NULL,
  itemStyle.emphasis.label.textShadowBlur = NULL,
  itemStyle.emphasis.label.textShadowOffsetY = NULL,
  itemStyle.emphasis.label.textShadowOffsetX = NULL,
  label.show = NULL,
  label.color = NULL,
  label.fontStyle = NULL,
  label.fontWeight = NULL,
  label.fontFamily = NULL,
  label.fontSize = NULL,
  label.align = NULL,
  label.verticalAlign = NULL,
  label.backgroundColor = NULL,
  label.borderColor = NULL,
  label.borderWidth = NULL,
  label.borderType = NULL,
  label.borderRadius = NULL,
  label.shadowColor = NULL,
  label.shadowBlur = NULL,
  label.shadowOffsetY = NULL,
  label.shadowOffsetX = NULL,
  label.width = NULL,
  label.height = NULL,
  label.textBorderColor = NULL,
  label.textBorderWidth = NULL,
  label.textShadowColor = NULL,
  label.textShadowBlur = NULL,
  label.textShadowOffsetY = NULL,
  label.textShadowOffsetX = NULL
)
}
\arguments{
\item{e}{existing plot variable}

\item{y}{YVar}

\item{z}{ZVar}

\item{itemStyle.borderColor}{hex or name}

\item{itemStyle.borderWidth}{number}

\item{itemStyle.shadowColor}{hex or name}

\item{itemStyle.shadowBlur}{number}

\item{itemStyle.shadowOffsetY}{number}

\item{itemStyle.shadowOffsetX}{number}

\item{itemStyle.opacity}{decimal}

\item{itemStyle.borderRadius}{number}

\item{emphasis.shadowColor}{hex or name}

\item{emphasis.shadowBlur}{number}

\item{itemStyle.emphasis.label.show}{logical}

\item{itemStyle.emphasis.label.color}{hex or name}

\item{itemStyle.emphasis.label.fontStyle}{'normal' 'italic' 'oblique'}

\item{itemStyle.emphasis.label.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{itemStyle.emphasis.label.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{itemStyle.emphasis.label.fontSize}{number}

\item{itemStyle.emphasis.label.align}{'left' 'center' 'right'}

\item{itemStyle.emphasis.label.verticalAlign}{'top' 'middle' 'bottom'}

\item{itemStyle.emphasis.label.backgroundColor}{hex or name}

\item{itemStyle.emphasis.label.borderColor}{hex or name}

\item{itemStyle.emphasis.label.borderWidth}{number}

\item{itemStyle.emphasis.label.borderType}{'solid' 'dashed' 'dotted'}

\item{itemStyle.emphasis.label.borderRadius}{number}

\item{itemStyle.emphasis.label.shadowColor}{hex or name}

\item{itemStyle.emphasis.label.shadowBlur}{number}

\item{itemStyle.emphasis.label.shadowOffsetY}{number}

\item{itemStyle.emphasis.label.shadowOffsetX}{number}

\item{itemStyle.emphasis.label.width}{number}

\item{itemStyle.emphasis.label.height}{number}

\item{itemStyle.emphasis.label.textBorderColor}{hex or name}

\item{itemStyle.emphasis.label.textBorderWidth}{number}

\item{itemStyle.emphasis.label.textShadowColor}{hex or name}

\item{itemStyle.emphasis.label.textShadowBlur}{number}

\item{itemStyle.emphasis.label.textShadowOffsetY}{number}

\item{itemStyle.emphasis.label.textShadowOffsetX}{number}

\item{label.show}{logical}

\item{label.color}{hex or name}

\item{label.fontStyle}{'normal' 'italic' 'oblique'}

\item{label.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{label.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{label.fontSize}{number}

\item{label.align}{'left' 'center' 'right'}

\item{label.verticalAlign}{'top' 'middle' 'bottom'}

\item{label.backgroundColor}{hex or name}

\item{label.borderColor}{hex or name}

\item{label.borderWidth}{number}

\item{label.borderType}{'solid' 'dashed' 'dotted'}

\item{label.borderRadius}{number}

\item{label.shadowColor}{hex or name}

\item{label.shadowBlur}{number}

\item{label.shadowOffsetY}{number}

\item{label.shadowOffsetX}{number}

\item{label.width}{number}

\item{label.height}{number}

\item{label.textBorderColor}{hex or name}

\item{label.textBorderWidth}{number}

\item{label.textShadowColor}{hex or name}

\item{label.textShadowBlur}{number}

\item{label.textShadowOffsetY}{number}

\item{label.textShadowOffsetX}{number}
}
\value{
The modified echarts4r object
}
\description{
Expose most heatmap options
}
