% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_line_full}
\alias{e_line_full}
\title{Enhanced line Setter for echarts4r}
\usage{
e_line_full(
  e = NULL,
  serie = NULL,
  y_index = 0,
  x_index = 0,
  smooth = NULL,
  showSymbol = NULL,
  label = NULL,
  lineStyle.width = NULL,
  lineStyle.type = NULL,
  lineStyle.shadowColor = NULL,
  lineStyle.shadowBlur = NULL,
  lineStyle.shadowOffsetX = NULL,
  lineStyle.shadowOffsetY = NULL
)
}
\arguments{
\item{e}{plot object}

\item{serie}{Variable}

\item{y_index}{default 0}

\item{x_index}{default 0}

\item{smooth}{Smooth line}

\item{showSymbol}{Logical}

\item{label}{logical}

\item{lineStyle.width}{numeric}

\item{lineStyle.type}{'solid' 'dashed' 'dotted'}

\item{lineStyle.shadowColor}{hex or color}

\item{lineStyle.shadowBlur}{numeric}

\item{lineStyle.shadowOffsetX}{numeric}

\item{lineStyle.shadowOffsetY}{numeric}
}
\value{
The modified echarts4r object
}
\description{
Exposes line* option so you don't have to hand-craft the JSON.
}
