% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_parallel_full}
\alias{e_parallel_full}
\title{Parallel customization}
\usage{
e_parallel_full(
  e,
  vars,
  rm_x = TRUE,
  rm_y = TRUE,
  lineStyle.color = NULL,
  lineStyle.width = NULL,
  lineStyle.type = NULL,
  lineStyle.shadowBlur = NULL,
  lineStyle.shadowColor = NULL,
  lineStyle.shadowOffsetX = NULL,
  lineStyle.shadowOffsetY = NULL,
  lineStyle.opacity = NULL,
  lineStyle.emphasis.color = NULL,
  lineStyle.emphasis.width = NULL,
  lineStyle.emphasis.type = NULL,
  lineStyle.emphasis.shadowBlur = NULL,
  lineStyle.emphasis.shadowColor = NULL,
  lineStyle.emphasis.shadowOffsetX = NULL,
  lineStyle.emphasis.shadowOffsetY = NULL,
  lineStyle.emphasis.opacity = NULL
)
}
\arguments{
\item{e}{plot}

\item{vars}{character vector}

\item{rm_x}{logical}

\item{rm_y}{logical}

\item{lineStyle.color}{hex or color}

\item{lineStyle.width}{number}

\item{lineStyle.type}{'solid' 'dashed' 'dotted'}

\item{lineStyle.shadowBlur}{number}

\item{lineStyle.shadowColor}{hex or color}

\item{lineStyle.shadowOffsetX}{number}

\item{lineStyle.shadowOffsetY}{number}

\item{lineStyle.opacity}{decimal}

\item{lineStyle.emphasis.color}{hex or color}

\item{lineStyle.emphasis.width}{number}

\item{lineStyle.emphasis.type}{'solid' 'dashed' 'dotted'}

\item{lineStyle.emphasis.shadowBlur}{number}

\item{lineStyle.emphasis.shadowColor}{hex or color}

\item{lineStyle.emphasis.shadowOffsetX}{number}

\item{lineStyle.emphasis.shadowOffsetY}{number}

\item{lineStyle.emphasis.opacity}{decimal}
}
\value{
The modified echarts4r object
}
\description{
Expose most parallel options
}
