% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_visual_map_full}
\alias{e_visual_map_full}
\title{Enhanced visual map}
\usage{
e_visual_map_full(
  e,
  serie,
  visualMap.show = TRUE,
  visualMap.min = NULL,
  visualMap.max = NULL,
  visualMap.orient = NULL,
  visualMap.right = NULL,
  visualMap.top = NULL,
  visualMap.bottom = NULL,
  visualMap.left = NULL,
  visualMap.backgroundColor = NULL,
  visualMap.borderColor = NULL,
  visualMap.borderWidth = NULL,
  visualMap.InRange.color = NULL,
  visualMap.InRange.opacity = NULL,
  visualMap.InRange.symbol = NULL,
  visualMap.InRange.symbolSize = NULL
)
}
\arguments{
\item{e}{An echarts4r object}

\item{serie}{ZVar XVar YVar}

\item{visualMap.show}{logical}

\item{visualMap.min}{min value}

\item{visualMap.max}{max value}

\item{visualMap.orient}{'vertical' 'horizontal'}

\item{visualMap.right}{number}

\item{visualMap.top}{number}

\item{visualMap.bottom}{number}

\item{visualMap.left}{number}

\item{visualMap.backgroundColor}{hex or name}

\item{visualMap.borderColor}{hex or name}

\item{visualMap.borderWidth}{number}

\item{visualMap.InRange.color}{hex or name}

\item{visualMap.InRange.opacity}{number}

\item{visualMap.InRange.symbol}{'circle', 'rect', 'roundRect', 'triangle', 'diamond', 'pin', 'arrow', 'none'}

\item{visualMap.InRange.symbolSize}{number}
}
\value{
plot
}
\description{
Exposes most visual map options
}
