% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_x_axis_full}
\alias{e_x_axis_full}
\title{Enhanced X-Axis Setter for echarts4r}
\usage{
e_x_axis_full(
  e,
  index = 0,
  serie = NULL,
  axis = "x",
  name = NULL,
  xAxis.position = "bottom",
  xAxis.nameLocation = "center",
  xAxis.axisTick.customValues = NULL,
  xAxis.nameTextStyle.color = NULL,
  xAxis.nameTextStyle.backgroundColor = NULL,
  xAxis.nameTextStyle.padding = 15,
  xAxis.nameTextStyle.align = NULL,
  xAxis.nameTextStyle.fontStyle = NULL,
  xAxis.nameTextStyle.fontWeight = NULL,
  xAxis.nameTextStyle.fontSize = NULL,
  xAxis.nameTextStyle.fontFamily = NULL,
  xAxis.min = NULL,
  xAxis.max = NULL,
  xAxis.splitNumber = NULL,
  xAxis.axisLabel.rotate = NULL,
  xAxis.axisLabel.margin = NULL,
  xAxis.axisLabel.color = NULL,
  xAxis.axisLabel.fontStyle = NULL,
  xAxis.axisLabel.fontWeight = NULL,
  xAxis.axisLabel.fontFamily = NULL,
  xAxis.axisLabel.fontSize = NULL,
  xAxis.axisLabel.align = NULL,
  xAxis.axisLabel.verticalAlign = NULL,
  xAxis.axisLabel.backgroundColor = NULL,
  xAxis.axisLabel.borderColor = NULL,
  xAxis.axisLabel.borderWidth = NULL,
  xAxis.axisLabel.borderType = NULL,
  xAxis.axisLabel.borderRadius = NULL,
  xAxis.axisLabel.padding = NULL,
  xAxis.axisLabel.shadowColor = NULL,
  xAxis.axisLabel.shadowBlur = NULL,
  xAxis.axisLabel.shadowOffsetX = NULL,
  xAxis.axisLabel.shadowOffsetY = NULL,
  xAxis.axisLabel.textBorderColor = NULL,
  xAxis.axisLabel.textBorderWidth = NULL,
  xAxis.axisLabel.textBorderType = NULL,
  xAxis.axisLabel.textShadowColor = NULL,
  xAxis.axisLabel.textShadowBlur = NULL,
  xAxis.axisLabel.textShadowOffsetX = NULL,
  xAxis.axisLabel.textShadowOffsetY = NULL,
  xAxis.axisLabel.overflow = NULL,
  xAxis.axisPointer.axis = NULL,
  xAxis.axisPointer.type = NULL,
  xAxis.axisPointer.triggerEmphasis = NULL,
  xAxis.axisPointer.triggerTooltip = NULL
)
}
\arguments{
\item{e}{An echarts4r object}

\item{index}{Which x axis to target (zero-based). Default: 0}

\item{xAxis.nameTextStyle.fontStyle, fontWeight, fontSize}{Font settings for the axis name}

\item{xAxis.min, xAxis.max}{Numeric data min/max}

\item{xAxis.splitNumber}{Numeric: how many segments between min & max}

\item{xAxis.axisLabel.*}{See list below}

\item{xAxis.axisPointer.*}{See list below}
}
\value{
The modified echarts4r object
}
\description{
Exposes every xAxis.* option so you don't have to hand-craft the JSON.
}
