% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessoryFunctions.R
\name{FakeDataGenerator}
\alias{FakeDataGenerator}
\title{FakeDataGenerator}
\usage{
FakeDataGenerator(
  Correlation = 0.7,
  N = 1000L,
  ID = 5L,
  FactorCount = 2L,
  AddDate = TRUE,
  AddComment = FALSE,
  AddWeightsColumn = FALSE,
  ZIP = 5L,
  ChainLadderData = FALSE,
  Classification = FALSE,
  MultiClass = FALSE
)
}
\arguments{
\item{Correlation}{Set the correlation value for simulated data}

\item{N}{Number of records}

\item{ID}{Number of IDcols to include}

\item{FactorCount}{Number of factor type columns to create}

\item{AddDate}{Set to TRUE to include a date column}

\item{AddComment}{Set to TRUE to add a comment column}

\item{AddWeightsColumn}{Add a weights column for ML}

\item{ZIP}{Zero Inflation Model target variable creation. Select from 0 to 5 to create that number of distinctly distributed data, stratifed from small to large}

\item{ChainLadderData}{Set to TRUE to return Chain Ladder Data for using AutoMLChainLadderTrainer}

\item{Classification}{Set to TRUE to build classification data}

\item{MultiClass}{Set to TRUE to build MultiClass data}
}
\value{
data.table of data
}
\description{
Create fake data for examples
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
