% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Area}
\alias{Plot.Area}
\title{Plot.Area}
\usage{
Plot.Area(
  dt = NULL,
  AggMethod = "mean",
  PreAgg = TRUE,
  Engine = "Echarts",
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Line Plot",
  EchartsTheme = "macaron",
  X_Scroll = FALSE,
  Y_Scroll = FALSE,
  TimeLine = TRUE,
  Alpha = 0.5,
  Smooth = TRUE,
  ShowSymbol = FALSE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{AggMethod}{character}

\item{PreAgg}{logical}

\item{Engine}{"Echarts" or "Plotly"}

\item{XVar}{Column name of the predicted values from your model}

\item{YVar}{Column name of the target variable from your model}

\item{GroupVar}{One Grouping Variable}

\item{Title}{"Title"}

\item{EchartsTheme}{Provide an "Echarts" theme}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{TimeLine}{Logical}

\item{Alpha}{0 to 1 for setting transparency}

\item{Smooth}{= TRUE}

\item{ShowSymbol}{= FALSE}

\item{BackGroundColor}{color}

\item{ChartColor}{color}

\item{FillColor}{color}

\item{FillColorReverse}{character}

\item{GridColor}{color}

\item{TextColor}{"Not Implemented"}

\item{ZeroLineColor}{color}

\item{ZeroLineWidth}{1}

\item{Debug}{FALSE}

\item{Area}{logical}
}
\value{
Calibration plot or boxplot
}
\description{
This function automatically builds calibration plots and calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
\dontrun{
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.Area(
  dt = data,
  PreAgg = TRUE,
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Title = 'Bar Plot',
  FillColor = "#0066ff",
  BackGroundColor = "#6a6969", #"#1b1959", #'#00060b',
  ChartColor = '#001534',
  GridColor = 'white',
  TextColor = 'white',
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# Y_Scroll = FALSE
# X_Scroll = FALSE
# Engine = "Echarts" # "Plotly"
# TimeLine = TRUE
# EchartsTheme = "macaron"
# Area = TRUE
# Alpha = 0.50
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# FillColor = "#0066ff"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# PreAgg = TRUE
# Debug = FALSE
}
}
\seealso{
Other Standard Plots: 
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
