% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Bar}
\alias{Plot.Bar}
\title{Plot.Bar}
\usage{
Plot.Bar(
  data = NULL,
  PreAgg = FALSE,
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  X_Index = TRUE,
  Y_Index = TRUE,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  AggMethod = "mean",
  Title = "Bar Plot",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  SubTitleColor = "white",
  BorderColor = "white",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{PreAgg}{logical}

\item{Engine}{'Plotly' or "Echarts"}

\item{EchartsTheme}{"macaron"}

\item{TimeLine}{logical}

\item{X_Index}{logical}

\item{Y_Index}{logical}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{= NULL}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{Title}{title}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff'}

\item{FacetVar1}{Column name of facet variable 1. If NULL then ignored}

\item{FacetVar2}{Column name of facet variable 2. If NULL then ignored}

\item{TextSize}{14}

\item{AngleX}{90}

\item{AngleY}{0}

\item{SubTitleColor}{'darkblue'}

\item{BorderColor}{'darkblue'}

\item{LegendPosition}{'bottom'}

\item{LegendBorderSize}{0.50}

\item{LegendLineType}{'solid'}

\item{Debug}{FALSE}
}
\description{
Build a bar plot by simply passing arguments to a single function
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Bar(
  Engine = 'Plotly', # "Echarts"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  data = data,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  AggMethod = 'mean',
  FacetVar1 = 'Store',
  FacetVar2 = 'Dept',
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = 'lightsteelblue1',
  BorderColor = 'darkblue',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  SubTitleColor = 'blue',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  Debug = FALSE)

# # Step through function
# dt <- data
# Engine = 'Plotly' # "Echarts"
# EchartsTheme = "macaron"
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# AggMethod = 'mean'
# GroupVar = NULL
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# FacetVar1 = NULL
# FacetVar2 = NULL
# FillColor = "#0066ff"
# TextSize = 12
# AngleX = 90
# AngleY = 0
# BackGroundColor = "#6a6969" #, #"#1b1959", #'#00060b',
# ChartColor = '#001534'
# BorderColor = 'darkblue'
# TextColor = 'darkblue'
# GridColor = 'white'
# SubTitleColor = 'blue'
# LegendPosition = 'bottom'
# LegendBorderSize = 0.50
# LegendLineType = 'solid'
# Debug = FALSE
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
