% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.BarPlot3D}
\alias{Plot.BarPlot3D}
\title{Plot.BarPlot3D}
\usage{
Plot.BarPlot3D(
  dt,
  PreAgg = FALSE,
  AggMethod = "mean",
  XVar = NULL,
  YVar = NULL,
  ZVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  ZVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  NumberBins = 21,
  NumLevels_Y = 33,
  NumLevels_X = 33,
  Height = NULL,
  Width = NULL,
  Title = "3D Bar Plot",
  ShowLabels = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  Engine = "Plotly",
  EchartsTheme = "dark",
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = "#63aeff",
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  yaxis.fontSize = 14,
  xaxis.fontSize = 14,
  zaxis.fontSize = 14,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{AggMethod}{'mean', 'median', 'sum', 'sd', 'coeffvar', 'count'}

\item{XVar}{X-Axis variable name}

\item{YVar}{Y-Axis variable name}

\item{ZVar}{Z-Axis variable name}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{ZVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{NumberBins}{= 21}

\item{NumLevels_Y}{= 20}

\item{NumLevels_X}{= 20}

\item{Height}{= NULL,}

\item{Width}{= NULL,}

\item{Title}{"Heatmap"}

\item{ShowLabels}{character}

\item{Title.YAxis}{character}

\item{Title.XAxis}{character}

\item{Engine}{"plotly", "echarts4r"}

\item{EchartsTheme}{"dark-blue"}

\item{BackGroundColor}{color outside of plot window. Rcolors and hex outside of plot window. Rcolors and hex character}

\item{ChartColor}{= '#001534'}

\item{FillColor}{= "#0066ff"}

\item{FillColorReverse}{character}

\item{GridColor}{= '#ffffff'}

\item{Debug}{Debugging purposes}
}
\description{
Build a 3D Bar Plot
}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Violin}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
