% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Box}
\alias{Plot.Box}
\title{Plot.Box}
\usage{
Plot.Box(
  dt = NULL,
  SampleSize = 100000L,
  XVar = NULL,
  YVar = NULL,
  GroupVars = NULL,
  Title = "Box Plot",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TimeLine,
  X_Index = TRUE,
  Y_Index = TRUE,
  FillColor = "#0066ff",
  FillColorReverse = "#ff9900",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  GridColor = "#ffff",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVars}{Group Variable}

\item{Title}{'Box Plot'}

\item{Engine}{= "Plotly"}

\item{EchartsTheme}{= "macaron"}

\item{TimeLine}{Logical}

\item{X_Index}{logical}

\item{Y_Index}{logical}

\item{FillColor}{'gray'}

\item{FillColorReverse}{= "#ff9900"}

\item{BackGroundColor}{'#17108db8'}

\item{ChartColor}{'#001534'}

\item{GridColor}{'white'}

\item{ZeroLineColor}{'#ffff'}

\item{ZeroLineWidth}{1.25}

\item{TextColor}{'darkblue'}

\item{Debug}{FALSE}

\item{FacetVar1}{Column name of facet variable 1. If NULL then ignored}

\item{FacetVar2}{Column name of facet variable 2. If NULL then ignored}

\item{TextSize}{14}

\item{AngleX}{90}

\item{AngleY}{0}

\item{BorderColor}{'darkblue'}

\item{SubTitleColor}{'darkblue'}

\item{LegendPosition}{'bottom'}

\item{LegendBorderSize}{0.50}

\item{LegendLineType}{'solid'}
}
\description{
Build a box plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Box(
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TimeLine,
  dt = data,
  XVar = 'Predict',
  YVar = 'CHILLED_Margin_PerDay',
  GroupVars = 'BRAND',
  Title = 'Box Plot',
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  SampleSize = 100000,
  FillColor = "#0066ff",
  FillColorReverse = "#ff9900",
  BackGroundColor = "#6a6969",
  ChartColor = '#001534',
  GridColor = '#ffff',
  TextColor = 'white',
  Debug = FALSE)

# # Step through function
# Engine = "Plotly"
# EchartsTheme = "macaron"
# TimeLine = TimeLine
# data = data
# XVar = 'Predict'
# YVar = 'CHILLED_Margin_PerDay'
# GroupVars = 'BRAND'
# Title = 'Box Plot'
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# SampleSize = 100000
# FillColor = "#0066ff"
# FillColorReverse = "#ff9900"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = '#ffff'
# TextColor = 'white'
# Debug = FALSE
}
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
