% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.CorrMatrix}
\alias{Plot.CorrMatrix}
\title{Plot.CorrMatrix}
\usage{
Plot.CorrMatrix(
  dt = NULL,
  CorrVars = NULL,
  CorrVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  Method = "spearman",
  PreAgg = FALSE,
  Height = "600px",
  Width = "1135px",
  Title = "Correlation Matrix",
  Engine = "Plotly",
  EchartsTheme = "macarons",
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = "#63aeff",
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  yaxis.fontSize = 14,
  xaxis.fontSize = 14,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{CorrVars}{character}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{Method}{character}

\item{PreAgg}{logical}

\item{Height}{= NULL,}

\item{Width}{= NULL,}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{= "macaron"}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{FillColorReverse}{character}

\item{GridColor}{character hex}

\item{TextColor}{character hex}

\item{Debug}{logical}

\item{CorrVarsTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Violin}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
