% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Density}
\alias{Plot.Density}
\title{Plot.Density}
\usage{
Plot.Density(
  dt = NULL,
  YVar = NULL,
  XVar = NULL,
  GroupVar = NULL,
  Title = "Density Plot",
  Engine = "Plotly",
  EchartsTheme = "macarons",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#17108db8",
  ChartColor = "#001534",
  GridColor = "white",
  TextColor = "white",
  SubTitleColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{From App}

\item{YVar}{From App}

\item{XVar}{character}

\item{GroupVar}{character}

\item{Title}{= "Density Plot"}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{From App}

\item{ChartColor}{From App}

\item{GridColor}{From App}

\item{TextColor}{From App}

\item{SubTitleColor}{From App}

\item{Debug}{From App}

\item{BorderColor}{From App}

\item{LegendPosition}{From App}

\item{LegendBorderSize}{From App}

\item{LegendLineType}{From App}
}
\description{
Density plots, by groups, with transparent continuous plots
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
p1 <- AutoPlots:::Plot.Density(
  TimeLine = FALSE,
  data = data,
  YVar = NULL,
  GroupVar = 'Weekly_Sales',
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = 'lightsteelblue1',
  BorderColor = 'darkblue',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  SubTitleColor = 'blue',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  Debug = FALSE)

# Step through function
# # plotly::ggplotly(p1)
# TimeLine = FALSE
# Engine = "Plotly"
# EchartsTheme = "macarons"
# data = data
# YVar = 'Weekly_Sales'
# GroupVar = c('Store','Dept')
# TextSize = 12
# AngleX = 90
# AngleY = 0
# ChartColor = 'lightsteelblue1'
# BorderColor = 'darkblue'
# TextColor = 'darkblue'
# GridColor = 'white'
# BackGroundColor = 'gray95'
# SubTitleColor = 'blue'
# LegendPosition = 'bottom'
# LegendBorderSize = 0.50
# LegendLineType = 'solid'
# Debug = FALSE
# NumberBins
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\concept{Standard Plots}
