% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.HeatMap}
\alias{Plot.HeatMap}
\title{Plot.HeatMap}
\usage{
Plot.HeatMap(
  dt,
  PreAgg = FALSE,
  AggMethod = "mean",
  XVar = NULL,
  YVar = NULL,
  ZVar = NULL,
  PercentileBuckets_X = 0.1,
  PercentileBuckets_Y = 0.1,
  NLevels_X = 33,
  NLevels_Y = 33,
  Title = "Heatmap",
  Engine = "Plotly",
  EchartsTheme = "dark",
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "#ffffff"
)
}
\arguments{
\item{dt}{Source data.table}

\item{AggMethod}{'mean', 'median', 'sum', 'sd', 'count'}

\item{XVar}{X-Axis variable}

\item{YVar}{Y-Axis variable}

\item{ZVar}{Z-Axis variable}

\item{PercentileBuckets_X}{= 0.10}

\item{PercentileBuckets_Y}{= 0.10}

\item{NLevels_X}{= 20}

\item{NLevels_Y}{= 20}

\item{Title}{"Heatmap"}

\item{Engine}{"plotly", "echarts4r"}

\item{EchartsTheme}{"dark-blue"}

\item{BackGroundColor}{= "#6a6969"}

\item{ChartColor}{= '#001534'}

\item{FillColor}{= "#0066ff"}

\item{GridColor}{= '#ffffff'}
}
\description{
Create heat maps with numeric or categorical dt
}
\examples{
\dontrun{
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")
Columns <- c(
  "DATE_ISO",
  "ARTICLE",
  "BRAND",
  "CUSTOMER_COD_char",
  "CHILLED_Margin_PerDay",
  "CHILLED_Liters_PerDay",
  "CHILLED_Units_PerDay")
data <- AutoPlots::DM.pgQuery(
  Query = NULL,
  DataBase = "KompsProcessed",
  SELECT = Columns,
  AggStat = "AVG",
  FROM = "POS_Processed_Long_Daily_backward",
  GroupBy = NULL,
  SamplePercent = 1,
  Host = 'localhost',
  User = 'postgres',
  Port = 5432,
  Password = "Aa...")

AutoPlots:::Plot.HeatMap(
  data,
  PreAgg = FALSE,
  Engine = 'echarts4r',
  XVar = 'ARTICLE',
  YVar = 'BRAND',
  ZVar = 'CHILLED_Margin_PerDay',
  AggMethod = 'mean',
  PercentileBuckets_X = 0.10,
  PercentileBuckets_Y = 0.10,
  NLevels_X = 33,
  NLevels_Y = 33)

# QA
data
PreAgg = FALSE
Engine = 'echarts4r'
XVar = 'ARTICLE'
YVar = 'BRAND'
ZVar = 'CHILLED_Margin_PerDay'
AggMethod = 'mean'
PercentileBuckets_X = 0.10
PercentileBuckets_Y = 0.10
NLevels_X = 33
NLevels_Y = 33
gridcolor = GridColor
plot_bgcolor = ChartColor
paper_bgcolor = BackGroundColor

}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
