% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Histogram}
\alias{Plot.Histogram}
\title{Plot.Histogram}
\usage{
Plot.Histogram(
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  data = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Histogram",
  SampleSize = 1e+05,
  NumberBins = 30,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineWidth = 1.25,
  ZeroLineColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{Engine}{= PlotEngineType,}

\item{EchartsTheme}{= EchartsTheme,}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{data}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Column name of Group Variable for distinct colored histograms by group levels}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{NumberBins}{= 30}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineWidth}{= 1.25,}

\item{ZeroLineColor}{= "white",}

\item{Debug}{FALSE}
}
\description{
Build a histogram plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
p1 <- AutoPlots:::Plot.Histogram(
  data = data,
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  FacetVar1 = 'Store',
  FacetVar2 = 'Dept',
  SampleSize = 100000,
  NumberBins = 20,
  FillColor = "#0066ff",
  ChartColor = 'lightsteelblue1',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  Debug = FALSE)

# # Step through function
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# TimeLine = FALSE
# AggMethod = 'mean'
# ZeroLineWidth = 1.25
# Title = 'Histogram'
# FacetVar1 = NULL
# FacetVar2 = NULL
# NumberBins = 20
# SampleSize = 100000
# FillColor = "#0066ff"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'darkblue'
# Debug = FALSE
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
