% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Line}
\alias{Plot.Line}
\title{Plot.Line}
\usage{
Plot.Line(
  dt = NULL,
  PreAgg = TRUE,
  Engine = "Echarts",
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Line Plot",
  EchartsTheme = "macaron",
  X_Scroll = FALSE,
  Y_Scroll = FALSE,
  TimeLine = TRUE,
  Area = FALSE,
  Alpha = 0.5,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  GridColor = "white",
  TextColor = "white",
  FillColor = "#0066ff",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{PreAgg}{logical}

\item{Engine}{"Echarts" or "Plotly"}

\item{XVar}{Column name of the predicted values from your model}

\item{YVar}{Column name of the target variable from your model}

\item{GroupVar}{One Grouping Variable}

\item{Title}{"Title"}

\item{EchartsTheme}{Provide an "Echarts" theme}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{TimeLine}{Logical}

\item{Area}{logical}

\item{Alpha}{0 to 1 for setting transparency}

\item{BackGroundColor}{color}

\item{ChartColor}{color}

\item{GridColor}{color}

\item{TextColor}{"Not Implemented"}

\item{FillColor}{color}

\item{ZeroLineColor}{color}

\item{ZeroLineWidth}{1}

\item{Debug}{FALSE}

\item{TextSize}{1.25}
}
\value{
Calibration plot or boxplot
}
\description{
This function automatically builds calibration plots and calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
\dontrun{
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.Line(
  dt = data,
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Title = 'Bar Plot',
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  BackGroundColor = "#6a6969", #"#1b1959", #'#00060b',
  ChartColor = '#001534',
  GridColor = 'white',
  TextColor = 'white',
  SubTitleColor = 'white',
  BorderColor = 'white',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  PreAgg = TRUE,
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# Y_Scroll = FALSE
# X_Scroll = FALSE
# Engine = "Echarts" # "Plotly"
# TimeLine = TRUE
# EchartsTheme = "macaron"
# Area = TRUE
# Alpha = 0.50
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# FillColor = "#0066ff"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# PreAgg = TRUE
# Debug = FALSE
}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
