% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Parallel}
\alias{Plot.Parallel}
\title{Plot.Parallel}
\usage{
Plot.Parallel(
  dt = NULL,
  SampleSize = 50000,
  CorrVars = NULL,
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  PreAgg = FALSE,
  Height = NULL,
  Width = NULL,
  Title = "Parallel Plot",
  ShowLabels = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  EchartsTheme = "macarons",
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  TextColor = "white",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = "#63aeff",
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  yaxis.fontSize = 14,
  xaxis.fontSize = 14,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{CorrVars}{vector of variable names}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{PreAgg}{logical}

\item{Height}{= NULL,}

\item{Width}{= NULL,}

\item{Title}{character}

\item{ShowLabels}{character}

\item{Title.YAxis}{character}

\item{Title.XAxis}{character}

\item{EchartsTheme}{"auritus","azul","bee-inspired","blue","caravan","carp","chalk","cool","dark-bold","dark","eduardo", #' "essos","forest","fresh-cut","fruit","gray","green","halloween","helianthus","infographic","inspired", #' "jazz","london","dark","macarons","macarons2","mint","purple-passion","red-velvet","red","roma","royal", #' "sakura","shine","tech-blue","vintage","walden","wef","weforum","westeros","wonderland"}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{TextColor}{character hex}

\item{Debug}{Debugging purposes}
}
\description{
Build a parallel plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
dt = data.table::fread(file.choose())
SampleSize = 5000
CorrVars = c("Customer","Brand","Category") #names(dt)[!names(dt) \%in\% "Date"]
CorrVarTrans = "Identity"
FacetRows = 1
FacetCols = 1
FacetLevels = NULL
PreAgg = FALSE
Height = "400px"
Width = "600px"
Title = "Parallel Plot"
ShowLabels = FALSE
Title.YAxis = "bla"
Title.XAxis = "bloke"
EchartsTheme = "macarons"
X_Scroll = TRUE
Y_Scroll = TRUE
TextColor = "white"
title.fontSize = 22
title.fontWeight = "bold"
title.textShadowColor = '#63aeff'
title.textShadowBlur = 3
title.textShadowOffsetY = 1
title.textShadowOffsetX = -1
yaxis.fontSize = 14
xaxis.fontSize = 14
Debug = FALSE

AutoPlots::Plot.Parallel(
  dt = dt,
  CorrVars = CorrVars,
  FacetRows = FacetRows,
  FacetCols = FacetCols,
  FacetLevels = FacetLevels,
  PreAgg = PreAgg,
  Height = Height,
  Width = Width,
  Title = Title,
  ShowLabels = ShowLabels,
  Title.YAxis = Title.YAxis,
  Title.XAxis = Title.XAxis,
  EchartsTheme = EchartsTheme,
  X_Scroll = X_Scroll,
  Y_Scroll = Y_Scroll,
  TextColor = TextColor,
  title.fontSize = title.fontSize,
  title.fontWeight = title.fontWeight,
  title.textShadowColor = title.textShadowColor,
  title.textShadowBlur = title.textShadowBlur,
  title.textShadowOffsetY = title.textShadowOffsetY,
  title.textShadowOffsetX = title.textShadowOffsetX,
  yaxis.fontSize = yaxis.fontSize,
  xaxis.fontSize = xaxis.fontSize,
  Debug = Debug
)
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.ACF}()},
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.Donut}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.PACF}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.ProbabilityPlot}()},
\code{\link{Plot.Radar}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Rosetype}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.WordCloud}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
