% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Violin}
\alias{Plot.Violin}
\title{Plot.Violin}
\usage{
Plot.Violin(
  dt = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Violin Plot",
  SampleSize = 1e+05,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Requires an XVar and YVar already be defined}

\item{Title}{'Violin Plot'}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{FillColorReverse}{character}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff',}

\item{ZeroLineWidth}{= 2,}

\item{Debug}{FALSE}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoQuant)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Violin(
  dt = data,
  SampleSize = 100000,
  XVar = 'CHILLED_Margin_PerDay',
  YVar = 'CHILLED_Units_PerDay',
  GroupVar = 'BRAND',
  BackGroundColor = 'gray95',
  ChartColor = 'lightsteelblue1',
  FillColor = "#0066ff",
  TextColor = 'darkblue',
  GridColor = 'white',
  Debug = FALSE)

# Step through function
# SampleSize = 100000
# XVar = 'Region'
# YVar = 'Weekly_Sales'
# GroupVar = 'Store'
# BackGroundColor = 'gray95'
# ChartColor = 'lightsteelblue1'
# FillColor = "#0066ff"
# GridColor = 'white'
# TextColor = 'darkblue'
# Debug = FALSE
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Stock}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
