% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{PlotREsim}
\alias{PlotREsim}
\title{PlotREsim}
\usage{
PlotREsim(
  data = NULL,
  level = 0.95,
  stat = "median",
  sd = TRUE,
  sigmaScale = NULL,
  oddsRatio = FALSE,
  labs = FALSE,
  facet = TRUE
)
}
\arguments{
\item{data}{= AutoPlots::REsim(model),}

\item{level}{= 0.95}

\item{stat}{= "median}

\item{sd}{= TRUE}

\item{sigmaScale}{= NULL}

\item{oddsRatio}{= FALSE}

\item{labs}{= FALSE}

\item{facet}{= TRUE}
}
\description{
Modified version of merTools::PlotREsim
}
\examples{
\dontrun{

# Data, lmer model, output, plot
gpa <- data.table::fread(file = "C:/Users/Bizon/Documents/GitHub/gpa.csv")
model = lme4::lmer(gpa ~ occasion + (1 | student), data = gpa)
x <- AutoPlots::REsim(model)
AutoPlots::PlotREsim(x)
}

}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
