% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions_NEW.R
\name{Radar}
\alias{Radar}
\title{Radar}
\usage{
Radar(
  dt = NULL,
  AggMethod = "mean",
  PreAgg = TRUE,
  YVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  Height = NULL,
  Width = NULL,
  Title = "Radar Plot",
  ShowLabels = FALSE,
  Theme = "dark",
  ShowSymbol = FALSE,
  TextColor = "white",
  ContainLabel = TRUE,
  DarkMode = FALSE,
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = "#63aeff",
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  tooltip.show = TRUE,
  tooltip.trigger = "cross",
  tooltip.backgroundColor = NULL,
  tooltip.borderColor = NULL,
  tooltip.borderWidth = NULL,
  tooltip.padding = NULL,
  tooltip.axisPointer.type = "cross",
  tooltip.axisPointer.lineStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.shadowBlur = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetX = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetY = NULL,
  tooltip.axisPointer.shadowStyle.opacity = NULL,
  tooltip.textStyle.color = NULL,
  tooltip.textStyle.fontStyle = NULL,
  tooltip.textStyle.fontWeight = NULL,
  tooltip.textStyle.fontFamily = NULL,
  tooltip.textStyle.lineHeight = NULL,
  tooltip.textStyle.width = NULL,
  tooltip.textStyle.height = NULL,
  tooltip.textStyle.textBorderColor = NULL,
  tooltip.textStyle.textBorderWidth = NULL,
  tooltip.textStyle.textBorderType = NULL,
  tooltip.textStyle.textShadowColor = NULL,
  tooltip.textStyle.textShadowBlur = NULL,
  tooltip.textStyle.textShadowOffsetX = NULL,
  tooltip.textStyle.textShadowOffsetY = NULL,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{AggMethod}{character}

\item{PreAgg}{logical}

\item{YVar}{Y-Axis variable name. You can supply multiple YVars}

\item{GroupVar}{One Grouping Variable}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{Height}{"400px"}

\item{Width}{"200px"}

\item{Title}{"Title"}

\item{ShowLabels}{character}

\item{Theme}{Provide an "Echarts" theme}

\item{ShowSymbol}{= FALSE}

\item{TextColor}{"Not Implemented"}

\item{ContainLabel}{TRUE}

\item{DarkMode}{FALSE}

\item{title.fontSize}{22}

\item{title.fontWeight}{"bold"}

\item{title.textShadowColor}{'#63aeff'}

\item{title.textShadowBlur}{3}

\item{title.textShadowOffsetY}{1}

\item{title.textShadowOffsetX}{-1}

\item{tooltip.show}{logical}

\item{tooltip.trigger}{"cross" "axis" "item" "none"}

\item{tooltip.backgroundColor}{hex or name}

\item{tooltip.borderColor}{numeric}

\item{tooltip.borderWidth}{numeric}

\item{tooltip.padding}{numeric}

\item{tooltip.axisPointer.type}{"line" or "shadow"}

\item{tooltip.axisPointer.lineStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.shadowBlur}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetX}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetY}{numeric}

\item{tooltip.axisPointer.shadowStyle.opacity}{numeric between 0 and 1}

\item{tooltip.textStyle.color}{hex or name}

\item{tooltip.textStyle.fontStyle}{"normal" "italic" "oblique"}

\item{tooltip.textStyle.fontWeight}{"normal" "bold" "bolder" "lighter"}

\item{tooltip.textStyle.fontFamily}{valid family name}

\item{tooltip.textStyle.lineHeight}{numeric}

\item{tooltip.textStyle.width}{numeric}

\item{tooltip.textStyle.height}{numeric}

\item{tooltip.textStyle.textBorderColor}{hex or name}

\item{tooltip.textStyle.textBorderWidth}{numeric}

\item{tooltip.textStyle.textBorderType}{"solid" "dashed" "dotted"}

\item{tooltip.textStyle.textShadowColor}{hex or name}

\item{tooltip.textStyle.textShadowBlur}{numeric}

\item{tooltip.textStyle.textShadowOffsetX}{numeric}

\item{tooltip.textStyle.textShadowOffsetY}{numeric}

\item{Debug}{Debugging purposes}
}
\value{
plot
}
\description{
Radar
}
\examples{
# Create Data
dt <- data.table::data.table(Y = pnorm(q = runif(8)), GV = sample(LETTERS[1:4], 8, TRUE))

# Create plot
AutoPlots::Radar(
  dt = dt,
  AggMethod = "mean",
  PreAgg = FALSE,
  YVar = "Y",
  GroupVar = "GV",
  YVarTrans = "Identity",
  Height = NULL,
  Width = NULL,
  Title = 'Radar Plot',
  ShowLabels = FALSE,
  Theme = "dark",
  ShowSymbol = FALSE,
  TextColor = "black",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = '#63aeff',
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  ContainLabel = TRUE,
  DarkMode = FALSE,
  Debug = FALSE)

}
\seealso{
Other Standard Plots: 
\code{\link{ACF}()},
\code{\link{Area}()},
\code{\link{BarPlot3D}()},
\code{\link{Bar}()},
\code{\link{Box}()},
\code{\link{Copula3D}()},
\code{\link{Copula}()},
\code{\link{CorrMatrix}()},
\code{\link{Density}()},
\code{\link{Donut}()},
\code{\link{HeatMap}()},
\code{\link{Histogram}()},
\code{\link{Line}()},
\code{\link{PACF}()},
\code{\link{Parallel}()},
\code{\link{Pie}()},
\code{\link{Plot.ACF}()},
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.Donut}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.PACF}()},
\code{\link{Plot.Parallel}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.ProbabilityPlot}()},
\code{\link{Plot.Radar}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Rosetype}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.WordCloud}()},
\code{\link{ProbabilityPlot}()},
\code{\link{River}()},
\code{\link{Rosetype}()},
\code{\link{Scatter3D}()},
\code{\link{Scatter}()},
\code{\link{StackedBar}()},
\code{\link{Step}()},
\code{\link{WordCloud}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
