% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions_NEW.R
\name{Scatter}
\alias{Scatter}
\title{Scatter}
\usage{
Scatter(
  dt = NULL,
  SampleSize = 30000L,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  Height = NULL,
  Width = NULL,
  ShowLabels = FALSE,
  AddGLM = FALSE,
  Theme = "dark",
  MouseScroll = FALSE,
  TimeLine = FALSE,
  TextColor = "white",
  ContainLabel = TRUE,
  title.text = "Scatter Plot",
  title.subtext = NULL,
  title.link = NULL,
  title.sublink = NULL,
  title.Align = NULL,
  title.top = NULL,
  title.left = NULL,
  title.right = NULL,
  title.bottom = NULL,
  title.padding = NULL,
  title.itemGap = NULL,
  title.backgroundColor = NULL,
  title.borderColor = NULL,
  title.borderWidth = NULL,
  title.borderRadius = NULL,
  title.shadowColor = NULL,
  title.shadowBlur = NULL,
  title.shadowOffsetX = NULL,
  title.shadowOffsetY = NULL,
  title.textStyle.color = NULL,
  title.textStyle.fontStyle = NULL,
  title.textStyle.fontWeight = NULL,
  title.textStyle.fontFamily = NULL,
  title.textStyle.fontSize = NULL,
  title.textStyle.lineHeight = NULL,
  title.textStyle.width = NULL,
  title.textStyle.height = NULL,
  title.textStyle.textBorderColor = NULL,
  title.textStyle.textBorderWidth = NULL,
  title.textStyle.textBorderType = NULL,
  title.textStyle.textBorderDashOffset = NULL,
  title.textStyle.textShadowColor = NULL,
  title.textStyle.textShadowBlur = NULL,
  title.textStyle.textShadowOffsetX = NULL,
  title.textStyle.textShadowOffsetY = NULL,
  title.subtextStyle.color = NULL,
  title.subtextStyle.align = NULL,
  title.subtextStyle.fontStyle = NULL,
  title.subtextStyle.fontWeight = NULL,
  title.subtextStyle.fontFamily = NULL,
  title.subtextStyle.fontSize = NULL,
  title.subtextStyle.lineHeight = NULL,
  title.subtextStyle.width = NULL,
  title.subtextStyle.height = NULL,
  title.subtextStyle.textBorderColor = NULL,
  title.subtextStyle.textBorderWidth = NULL,
  title.subtextStyle.textBorderType = NULL,
  title.subtextStyle.textBorderDashOffset = NULL,
  title.subtextStyle.textShadowColor = NULL,
  title.subtextStyle.textShadowBlur = NULL,
  title.subtextStyle.textShadowOffsetX = NULL,
  title.subtextStyle.textShadowOffsetY = NULL,
  xAxis.title = NULL,
  xAxis.min = NULL,
  xAxis.max = NULL,
  xAxis.position = "bottom",
  xAxis.nameLocation = "center",
  xAxis.axisTick.customValues = NULL,
  xAxis.nameTextStyle.color = NULL,
  xAxis.nameTextStyle.padding = 25,
  xAxis.nameTextStyle.align = NULL,
  xAxis.nameTextStyle.fontStyle = NULL,
  xAxis.nameTextStyle.fontWeight = NULL,
  xAxis.nameTextStyle.fontSize = NULL,
  xAxis.nameTextStyle.fontFamily = NULL,
  xAxis.splitNumber = NULL,
  xAxis.axisLabel.rotate = NULL,
  xAxis.axisLabel.margin = NULL,
  xAxis.axisLabel.color = NULL,
  xAxis.axisLabel.fontStyle = NULL,
  xAxis.axisLabel.fontWeight = NULL,
  xAxis.axisLabel.fontFamily = NULL,
  xAxis.axisLabel.fontSize = 14,
  xAxis.axisLabel.align = NULL,
  xAxis.axisLabel.verticalAlign = NULL,
  xAxis.axisLabel.backgroundColor = NULL,
  xAxis.axisLabel.borderColor = NULL,
  xAxis.axisLabel.borderWidth = NULL,
  xAxis.axisLabel.borderType = NULL,
  xAxis.axisLabel.borderRadius = NULL,
  xAxis.axisLabel.padding = NULL,
  xAxis.axisLabel.shadowColor = NULL,
  xAxis.axisLabel.shadowBlur = NULL,
  xAxis.axisLabel.shadowOffsetX = NULL,
  xAxis.axisLabel.shadowOffsetY = NULL,
  xAxis.axisLabel.textBorderColor = NULL,
  xAxis.axisLabel.textBorderWidth = NULL,
  xAxis.axisLabel.textBorderType = NULL,
  xAxis.axisLabel.textShadowColor = NULL,
  xAxis.axisLabel.textShadowBlur = NULL,
  xAxis.axisLabel.textShadowOffsetX = NULL,
  xAxis.axisLabel.textShadowOffsetY = NULL,
  xAxis.axisLabel.overflow = NULL,
  yAxis.title = NULL,
  yAxis.min = NULL,
  yAxis.max = NULL,
  yAxis.position = "center",
  yAxis.nameLocation = "center",
  yAxis.axisTick.customValues = NULL,
  yAxis.nameTextStyle.color = NULL,
  yAxis.nameTextStyle.padding = 25,
  yAxis.nameTextStyle.align = NULL,
  yAxis.nameTextStyle.fontStyle = NULL,
  yAxis.nameTextStyle.fontWeight = NULL,
  yAxis.nameTextStyle.fontSize = NULL,
  yAxis.nameTextStyle.fontFamily = NULL,
  yAxis.splitNumber = NULL,
  yAxis.axisLabel.rotate = NULL,
  yAxis.axisLabel.margin = NULL,
  yAxis.axisLabel.color = NULL,
  yAxis.axisLabel.fontStyle = NULL,
  yAxis.axisLabel.fontWeight = NULL,
  yAxis.axisLabel.fontFamily = NULL,
  yAxis.axisLabel.fontSize = 14,
  yAxis.axisLabel.align = NULL,
  yAxis.axisLabel.verticalAlign = NULL,
  yAxis.axisLabel.backgroundColor = NULL,
  yAxis.axisLabel.borderColor = NULL,
  yAxis.axisLabel.borderWidth = NULL,
  yAxis.axisLabel.borderType = NULL,
  yAxis.axisLabel.borderRadius = NULL,
  yAxis.axisLabel.padding = NULL,
  yAxis.axisLabel.shadowColor = NULL,
  yAxis.axisLabel.shadowBlur = NULL,
  yAxis.axisLabel.shadowOffsetX = NULL,
  yAxis.axisLabel.shadowOffsetY = NULL,
  yAxis.axisLabel.textBorderColor = NULL,
  yAxis.axisLabel.textBorderWidth = NULL,
  yAxis.axisLabel.textBorderType = NULL,
  yAxis.axisLabel.textShadowColor = NULL,
  yAxis.axisLabel.textShadowBlur = NULL,
  yAxis.axisLabel.textShadowOffsetX = NULL,
  yAxis.axisLabel.textShadowOffsetY = NULL,
  yAxis.axisLabel.overflow = NULL,
  tooltip.show = TRUE,
  tooltip.trigger = "cross",
  tooltip.backgroundColor = NULL,
  tooltip.borderColor = NULL,
  tooltip.borderWidth = NULL,
  tooltip.padding = NULL,
  tooltip.axisPointer.type = "cross",
  tooltip.axisPointer.lineStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.color = NULL,
  tooltip.axisPointer.shadowStyle.shadowBlur = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetX = NULL,
  tooltip.axisPointer.shadowStyle.shadowOffsetY = NULL,
  tooltip.axisPointer.shadowStyle.opacity = NULL,
  tooltip.textStyle.color = NULL,
  tooltip.textStyle.fontStyle = NULL,
  tooltip.textStyle.fontWeight = NULL,
  tooltip.textStyle.fontFamily = NULL,
  tooltip.textStyle.lineHeight = NULL,
  tooltip.textStyle.width = NULL,
  tooltip.textStyle.height = NULL,
  tooltip.textStyle.textBorderColor = NULL,
  tooltip.textStyle.textBorderWidth = NULL,
  tooltip.textStyle.textBorderType = NULL,
  tooltip.textStyle.textShadowColor = NULL,
  tooltip.textStyle.textShadowBlur = NULL,
  tooltip.textStyle.textShadowOffsetX = NULL,
  tooltip.textStyle.textShadowOffsetY = NULL,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{SampleSize}{numeric}

\item{XVar}{X-Axis variable name}

\item{YVar}{Y-Axis variable name}

\item{GroupVar}{Character variable}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{Height}{"400px"}

\item{Width}{"200px"}

\item{ShowLabels}{character}

\item{AddGLM}{logical}

\item{Theme}{"auritus","azul","bee-inspired","blue","caravan","carp","chalk","cool","dark-bold","dark","eduardo", #' "essos","forest","fresh-cut","fruit","gray","green","halloween","helianthus","infographic","inspired", #' "jazz","london","dark","macarons","macarons2","mint","purple-passion","red-velvet","red","roma","royal", #' "sakura","shine","tech-blue","vintage","walden","wef","weforum","westeros","wonderland"}

\item{MouseScroll}{logical, zoom via mouse scroll}

\item{TimeLine}{logical}

\item{TextColor}{character hex}

\item{ContainLabel}{TRUE}

\item{title.text}{Title name}

\item{title.subtext}{Subtitle name}

\item{title.link}{Title as a link}

\item{title.sublink}{Subtitle as a link}

\item{title.Align}{'auto' 'left' 'right' 'center'}

\item{title.top}{'auto' '20' 'top' 'middle' 'bottom'}

\item{title.left}{distance between title and left side of container}

\item{title.right}{distance between title and right side of container}

\item{title.bottom}{'auto' '20' 'top' 'middle' 'bottom'}

\item{title.padding}{numeric}

\item{title.itemGap}{space between title and subtitle}

\item{title.backgroundColor}{hex or name}

\item{title.borderColor}{hex or name}

\item{title.borderWidth}{numeric}

\item{title.borderRadius}{numeric}

\item{title.shadowColor}{hex or name}

\item{title.shadowBlur}{numeric}

\item{title.shadowOffsetX}{numeric}

\item{title.shadowOffsetY}{numeric}

\item{title.textStyle.color}{hex or name}

\item{title.textStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{title.textStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{title.textStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{title.textStyle.fontSize}{numeric}

\item{title.textStyle.lineHeight}{numeric}

\item{title.textStyle.width}{numeric}

\item{title.textStyle.height}{numeric}

\item{title.textStyle.textBorderColor}{hex or name}

\item{title.textStyle.textBorderWidth}{numeric}

\item{title.textStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{title.textStyle.textBorderDashOffset}{numeric}

\item{title.textStyle.textShadowColor}{hex or name}

\item{title.textStyle.textShadowBlur}{numeric}

\item{title.textStyle.textShadowOffsetX}{numeric}

\item{title.textStyle.textShadowOffsetY}{numeric}

\item{title.subtextStyle.color}{hex or name}

\item{title.subtextStyle.align}{'auto' 'left' 'right' 'center'}

\item{title.subtextStyle.fontStyle}{'normal' 'italic' 'oblique'}

\item{title.subtextStyle.fontWeight}{'normal' 'bold' 'bolder' 'lighter'}

\item{title.subtextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{title.subtextStyle.fontSize}{numeric}

\item{title.subtextStyle.lineHeight}{numeric}

\item{title.subtextStyle.width}{numeric}

\item{title.subtextStyle.height}{numeric}

\item{title.subtextStyle.textBorderColor}{hex or name}

\item{title.subtextStyle.textBorderWidth}{numeric}

\item{title.subtextStyle.textBorderType}{'solid' 'dashed' 'dotted'}

\item{title.subtextStyle.textBorderDashOffset}{numeric}

\item{title.subtextStyle.textShadowColor}{numeric}

\item{title.subtextStyle.textShadowBlur}{numeric}

\item{title.subtextStyle.textShadowOffsetX}{numeric}

\item{title.subtextStyle.textShadowOffsetY}{numeric}

\item{xAxis.title}{Axis title}

\item{xAxis.min}{Min value}

\item{xAxis.max}{Max value}

\item{xAxis.position}{'bottom', 'top'}

\item{xAxis.nameLocation}{'center', 'start', 'middle'}

\item{xAxis.axisTick.customValues}{Custom values for the x-axis}

\item{xAxis.nameTextStyle.color}{hex or named color}

\item{xAxis.nameTextStyle.padding}{numeric}

\item{xAxis.nameTextStyle.align}{'center', 'left', 'right'}

\item{xAxis.nameTextStyle.fontStyle}{'italic', 'normal' 'oblique'}

\item{xAxis.nameTextStyle.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{xAxis.nameTextStyle.fontSize}{numeric}

\item{xAxis.nameTextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{xAxis.splitNumber}{numeric. Increments for numeric axis labels}

\item{xAxis.axisLabel.rotate}{numeric}

\item{xAxis.axisLabel.margin}{numeric}

\item{xAxis.axisLabel.color}{hex or named}

\item{xAxis.axisLabel.fontStyle}{'italic', 'normal' 'oblique'}

\item{xAxis.axisLabel.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{xAxis.axisLabel.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{xAxis.axisLabel.fontSize}{numeric}

\item{xAxis.axisLabel.align}{'center', 'left', 'right'}

\item{xAxis.axisLabel.verticalAlign}{'top', 'bottom'}

\item{xAxis.axisLabel.backgroundColor}{hex or named}

\item{xAxis.axisLabel.borderColor}{hex or named}

\item{xAxis.axisLabel.borderWidth}{numeric}

\item{xAxis.axisLabel.borderType}{'solid', 'dashed', 'dotted'}

\item{xAxis.axisLabel.borderRadius}{numeric}

\item{xAxis.axisLabel.padding}{numeric}

\item{xAxis.axisLabel.shadowColor}{hex or named}

\item{xAxis.axisLabel.shadowBlur}{numeric}

\item{xAxis.axisLabel.shadowOffsetX}{numeric}

\item{xAxis.axisLabel.shadowOffsetY}{numeric}

\item{xAxis.axisLabel.textBorderColor}{hex or named}

\item{xAxis.axisLabel.textBorderWidth}{numeric}

\item{xAxis.axisLabel.textBorderType}{'solid', 'dashed', 'dotted'}

\item{xAxis.axisLabel.textShadowColor}{hex or named}

\item{xAxis.axisLabel.textShadowBlur}{numeric}

\item{xAxis.axisLabel.textShadowOffsetX}{numeric}

\item{xAxis.axisLabel.textShadowOffsetY}{numeric}

\item{xAxis.axisLabel.overflow}{'truncate'}

\item{yAxis.title}{Axis title}

\item{yAxis.min}{Min value}

\item{yAxis.max}{Max value}

\item{yAxis.position}{'bottom', 'top'}

\item{yAxis.nameLocation}{'center', 'start', 'middle'}

\item{yAxis.axisTick.customValues}{Custom values for the y-axis}

\item{yAxis.nameTextStyle.color}{hex or named color}

\item{yAxis.nameTextStyle.padding}{numeric}

\item{yAxis.nameTextStyle.align}{'center', 'left', 'right'}

\item{yAxis.nameTextStyle.fontStyle}{'italic', 'normal' 'oblique'}

\item{yAxis.nameTextStyle.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{yAxis.nameTextStyle.fontSize}{numeric}

\item{yAxis.nameTextStyle.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{yAxis.splitNumber}{numeric. Increments for numeric axis labels}

\item{yAxis.axisLabel.rotate}{numeric}

\item{yAxis.axisLabel.margin}{numeric}

\item{yAxis.axisLabel.color}{hex or named}

\item{yAxis.axisLabel.fontStyle}{'italic', 'normal' 'oblique'}

\item{yAxis.axisLabel.fontWeight}{'normal', 'bold', 'bolder', 'lighter'}

\item{yAxis.axisLabel.fontFamily}{'sans-serif', 'serif', 'monospace', 'Arial', 'Times New Roman', 'Roboto', 'Open Sans', 'Lato', 'Helvetica', 'Georgia', 'Verdana', 'Arial', 'Tahoma', 'Courier New'}

\item{yAxis.axisLabel.fontSize}{numeric}

\item{yAxis.axisLabel.align}{'center', 'left', 'right'}

\item{yAxis.axisLabel.verticalAlign}{'top', 'bottom'}

\item{yAxis.axisLabel.backgroundColor}{hex or named}

\item{yAxis.axisLabel.borderColor}{hex or named}

\item{yAxis.axisLabel.borderWidth}{numeric}

\item{yAxis.axisLabel.borderType}{'solid', 'dashed', 'dotted'}

\item{yAxis.axisLabel.borderRadius}{numeric}

\item{yAxis.axisLabel.padding}{numeric}

\item{yAxis.axisLabel.shadowColor}{hex or named}

\item{yAxis.axisLabel.shadowBlur}{numeric}

\item{yAxis.axisLabel.shadowOffsetX}{numeric}

\item{yAxis.axisLabel.shadowOffsetY}{numeric}

\item{yAxis.axisLabel.textBorderColor}{hex or named}

\item{yAxis.axisLabel.textBorderWidth}{numeric}

\item{yAxis.axisLabel.textBorderType}{'solid', 'dashed', 'dotted'}

\item{yAxis.axisLabel.textShadowColor}{hex or named}

\item{yAxis.axisLabel.textShadowBlur}{numeric}

\item{yAxis.axisLabel.textShadowOffsetX}{numeric}

\item{yAxis.axisLabel.textShadowOffsetY}{numeric}

\item{yAxis.axisLabel.overflow}{'truncate'}

\item{tooltip.show}{logical}

\item{tooltip.trigger}{"cross" "axis" "item" "none"}

\item{tooltip.backgroundColor}{hex or name}

\item{tooltip.borderColor}{numeric}

\item{tooltip.borderWidth}{numeric}

\item{tooltip.padding}{numeric}

\item{tooltip.axisPointer.type}{"line" or "shadow"}

\item{tooltip.axisPointer.lineStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.color}{hex or name}

\item{tooltip.axisPointer.shadowStyle.shadowBlur}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetX}{numeric}

\item{tooltip.axisPointer.shadowStyle.shadowOffsetY}{numeric}

\item{tooltip.axisPointer.shadowStyle.opacity}{numeric between 0 and 1}

\item{tooltip.textStyle.color}{hex or name}

\item{tooltip.textStyle.fontStyle}{"normal" "italic" "oblique"}

\item{tooltip.textStyle.fontWeight}{"normal" "bold" "bolder" "lighter"}

\item{tooltip.textStyle.fontFamily}{valid family name}

\item{tooltip.textStyle.lineHeight}{numeric}

\item{tooltip.textStyle.width}{numeric}

\item{tooltip.textStyle.height}{numeric}

\item{tooltip.textStyle.textBorderColor}{hex or name}

\item{tooltip.textStyle.textBorderWidth}{numeric}

\item{tooltip.textStyle.textBorderType}{"solid" "dashed" "dotted"}

\item{tooltip.textStyle.textShadowColor}{hex or name}

\item{tooltip.textStyle.textShadowBlur}{numeric}

\item{tooltip.textStyle.textShadowOffsetX}{numeric}

\item{tooltip.textStyle.textShadowOffsetY}{numeric}

\item{Debug}{Debugging purposes}
}
\value{
plot
}
\description{
Build a copula plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
# Create fake data
data <- AutoPlots::FakeDataGenerator(N = 100000)
data[, Independent_Variable8 := Independent_Variable8 * runif(.N)]

# Echarts Scatter Plot Chart
AutoPlots::Scatter(
  dt = data,
  SampleSize = 10000,
  XVar = "Independent_Variable10",
  YVar = "Independent_Variable8",
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  Height = NULL,
  Width = NULL,
  Title = 'Scatter Plot',
  ShowLabels = FALSE,
  AddGLM = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  Theme = "blue",
  MouseScroll = FALSE,
  TimeLine = FALSE,
  TextColor = "black",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = '#63aeff',
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  yaxis.fontSize = 14,
  xaxis.fontSize = 14,
  xaxis.rotate = 0,
  yaxis.rotate = 0,
  ContainLabel = TRUE,
  tooltip.trigger = "axis",
  Debug = FALSE)

}
\seealso{
Other Standard Plots: 
\code{\link{ACF}()},
\code{\link{Area}()},
\code{\link{BarPlot3D}()},
\code{\link{Bar}()},
\code{\link{Box}()},
\code{\link{Copula3D}()},
\code{\link{Copula}()},
\code{\link{CorrMatrix}()},
\code{\link{Density}()},
\code{\link{Donut}()},
\code{\link{HeatMap}()},
\code{\link{Histogram}()},
\code{\link{Line}()},
\code{\link{PACF}()},
\code{\link{Parallel}()},
\code{\link{Pie}()},
\code{\link{Plot.ACF}()},
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.Donut}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.PACF}()},
\code{\link{Plot.Parallel}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.ProbabilityPlot}()},
\code{\link{Plot.Radar}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Rosetype}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.WordCloud}()},
\code{\link{ProbabilityPlot}()},
\code{\link{Radar}()},
\code{\link{River}()},
\code{\link{Rosetype}()},
\code{\link{Scatter3D}()},
\code{\link{StackedBar}()},
\code{\link{Step}()},
\code{\link{WordCloud}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
