% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{VI_Plot}
\alias{VI_Plot}
\title{VI_Plot}
\usage{
VI_Plot(
  Algo = "CatBoost",
  dt = NULL,
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  AggMethod = "mean",
  Title = "Variable Importance Plot",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  SubTitleColor = "white",
  BorderColor = "white",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{Algo}{'catboost', 'xgboost', 'h2o'}

\item{Engine}{'Plotly' or "Echarts"}

\item{EchartsTheme}{"macaron"}

\item{TimeLine}{logical}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{= NULL}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{Title}{title}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff'}

\item{Debug}{FALSE}

\item{data}{Source data.table}
}
\value{
ROC Plot for classification models
}
\description{
Generate variable importance plots
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML2_ScoringData.csv")
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_Test_VI_Data.csv")

# Step Through Function
library(AutoPlots)
library(data.table)
dt = data
Engine =  "Plotly" #@ "Echarts"
TimeLine = FALSE
EchartsTheme = "purple-passion"
XVar = "Importance" # "Predict"
YVar = "Variable" # "CHILLED_Margin_PerDay"
AggMethod = 'mean'
GroupVar = NULL # "BRAND"
NumBuckets = 20
AggStat = "mean"
ZeroLineColor = '#ffff'
ZeroLineWidth = 1.25
Title = 'Bar Plot'
FacetVar1 = NULL
FacetVar2 = NULL
FillColor = "#0066ff"
TextSize = 12
AngleX = 90
AngleY = 0
BackGroundColor = "#6a6969"
ChartColor = '#001534'
GridColor = 'white'
TextColor = 'white'
SubTitleColor = 'white'
BorderColor = 'white'
LegendPosition = 'bottom'
LegendBorderSize = 0.50
LegendLineType = 'solid'
Debug = FALSE

AutoPlots:::VI_Plot(
# Algo = "catboost",
# dt = data,
# Engine = 'Plotly',
# EchartsTheme = "macaron",
# TimeLine = TRUE,
# XVar = "Importance",
# YVar = "Variable",
# GroupVar = NULL,
# AggMethod = 'mean',
# Title = 'Variable Importance Plot',
# BackGroundColor = "#6a6969",
# ChartColor = '#001534',
# FillColor = "#0066ff",
# GridColor = 'white',
# TextColor = 'white',
# ZeroLineColor = '#ffff',
# ZeroLineWidth = 1.25,
# FacetVar1 = NULL,
# FacetVar2 = NULL,
# TextSize = 12,
# AngleX = 90,
# AngleY = 0,
# SubTitleColor = 'white',
# BorderColor = 'white',
# LegendPosition = 'bottom',
# LegendBorderSize = 0.50,
# LegendLineType = 'solid',
# Debug = FALSE)
}

}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
