% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_area_full}
\alias{e_area_full}
\title{Enhanced area Setter for echarts4r}
\usage{
e_area_full(
  e = NULL,
  serie = NULL,
  smooth = NULL,
  showSymbol = NULL,
  areaStyle.color = NULL,
  areaStyle.shadowBlur = NULL,
  areaStyle.shadowColor = NULL,
  areaStyle.shadowOffsetX = NULL,
  areaStyle.shadowOffsetY = NULL,
  areaStyle.opacity = NULL
)
}
\arguments{
\item{e}{plot object}

\item{serie}{Variable}

\item{smooth}{Smooth line}

\item{showSymbol}{Logical}

\item{areaStyle.color}{Fill color. Can be a single color or vector of multiple colors for gradient.}

\item{areaStyle.shadowBlur}{blur effect}

\item{areaStyle.shadowColor}{shadow color}

\item{areaStyle.shadowOffsetX}{shadow position along x-axis}

\item{areaStyle.shadowOffsetY}{shadow position along y-axis}

\item{areaStyle.opacity}{transparency}

\item{y_index}{Index of y axis}

\item{x_index}{Index of x axis}

\item{coord_system}{Coordinate system to plot against}
}
\value{
The modified echarts4r object
}
\description{
Exposes every area* option so you don't have to hand-craft the JSON.
}
