% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{e_legend_full}
\alias{e_legend_full}
\title{Enhanced Title Setter for echarts4r}
\usage{
e_legend_full(
  e = NULL,
  legend.show = FALSE,
  legend.type = "scroll",
  legend.selector = "all",
  legend.icon = NULL,
  legend.align = "auto",
  legend.padding = c(5, 10, 5, 10),
  legend.itemGap = 10,
  legend.itemWidth = 25,
  legend.orient = "horizontal",
  legend.width = NULL,
  legend.height = "240px",
  legend.left = NULL,
  legend.right = 50,
  legend.top = 40,
  legend.bottom = NULL,
  legend.backgroundColor = NULL,
  legend.borderColor = NULL,
  legend.borderWidth = NULL,
  legend.borderRadius = NULL,
  legend.shadowBlur = NULL,
  legend.shadowColor = NULL,
  legend.shadowOffsetX = NULL,
  legend.shadowOffsetY = NULL,
  legend.itemStyle.color = NULL,
  legend.itemStyle.borderColor = NULL,
  legend.itemStyle.borderWidth = NULL,
  legend.itemStyle.borderType = NULL,
  legend.itemStyle.shadowBlur = NULL,
  legend.itemStyle.shadowColor = NULL,
  legend.itemStyle.shadowOffsetX = NULL,
  legend.itemStyle.shadowOffsetY = NULL,
  legend.itemStyle.opacity = NULL,
  legend.lineStyle.color = NULL,
  legend.lineStyle.width = NULL,
  legend.lineStyle.type = NULL,
  legend.lineStyle.shadowBlur = NULL,
  legend.lineStyle.shadowColor = NULL,
  legend.lineStyle.shadowOffsetX = NULL,
  legend.lineStyle.shadowOffsetY = NULL,
  legend.lineStyle.opacity = NULL,
  legend.lineStyle.inactiveColor = NULL,
  legend.lineStyle.inactiveWidth = NULL,
  legend.textStyle.color = NULL,
  legend.textStyle.fontStyle = NULL,
  legend.textStyle.fontWeight = NULL,
  legend.textStyle.fontFamily = NULL,
  legend.textStyle.fontSize = NULL,
  legend.textStyle.backgroundColor = NULL,
  legend.textStyle.borderColor = NULL,
  legend.textStyle.borderWidth = NULL,
  legend.textStyle.borderType = NULL,
  legend.textStyle.borderRadius = NULL,
  legend.textStyle.padding = NULL,
  legend.textStyle.shadowColor = NULL,
  legend.textStyle.shadowBlur = NULL,
  legend.textStyle.shadowOffsetX = NULL,
  legend.textStyle.shadowOffsetY = NULL,
  legend.textStyle.width = NULL,
  legend.textStyle.height = NULL,
  legend.textStyle.textBorderColor = NULL,
  legend.textStyle.textBorderWidth = NULL,
  legend.textStyle.textBorderType = NULL,
  legend.textStyle.textShadowColor = NULL,
  legend.textStyle.textShadowBlur = NULL,
  legend.textStyle.textShadowOffsetX = NULL,
  legend.textStyle.textShadowOffsetY = NULL,
  legend.scrollDataIndex = NULL,
  legend.pageButtonItemGap = NULL,
  legend.pageButtonGap = NULL,
  legend.pageButtonPosition = NULL,
  legend.pageIconColor = NULL,
  legend.pageIconInactiveColor = NULL,
  legend.pageIconSize = NULL,
  legend.pageTextStyle = NULL,
  legend.animation = NULL,
  legend.animationDurationUpdate = NULL,
  legend.pageTextStyle.color = NULL,
  legend.pageTextStyle.fontStyle = NULL,
  legend.pageTextStyle.fontWeight = NULL,
  legend.pageTextStyle.fontFamily = NULL,
  legend.pageTextStyle.fontSize = NULL,
  legend.pageTextStyle.lineHeight = NULL,
  legend.pageTextStyle.width = NULL,
  legend.pageTextStyle.height = NULL,
  legend.pageTextStyle.textBorderColor = NULL,
  legend.pageTextStyle.textBorderWidth = NULL,
  legend.pageTextStyle.textBorderType = NULL,
  legend.pageTextStyle.textShadowColor = NULL,
  legend.pageTextStyle.textShadowBlur = NULL,
  legend.pageTextStyle.textShadowOffsetX = NULL,
  legend.pageTextStyle.textShadowOffsetY = NULL,
  legend.emphasis.selectorLabel.show = NULL,
  legend.emphasis.selectorLabel.distance = NULL,
  legend.emphasis.selectorLabel.rotate = NULL,
  legend.emphasis.selectorLabel.color = NULL,
  legend.emphasis.selectorLabel.fontStyle = NULL,
  legend.emphasis.selectorLabel.fontWeight = NULL,
  legend.emphasis.selectorLabel.fontFamily = NULL,
  legend.emphasis.selectorLabel.fontSize = NULL,
  legend.emphasis.selectorLabel.align = NULL,
  legend.emphasis.selectorLabel.verticalAlign = NULL,
  legend.emphasis.selectorLabel.lineHeight = NULL,
  legend.emphasis.selectorLabel.backgroundColor = NULL,
  legend.emphasis.selectorLabel.borderColor = NULL,
  legend.emphasis.selectorLabel.borderWidth = NULL,
  legend.emphasis.selectorLabel.borderType = NULL,
  legend.emphasis.selectorLabel.borderRadius = NULL,
  legend.emphasis.selectorLabel.padding = NULL,
  legend.emphasis.selectorLabel.shadowColor = NULL,
  legend.emphasis.selectorLabel.shadowBlur = NULL,
  legend.emphasis.selectorLabel.shadowOffsetX = NULL,
  legend.emphasis.selectorLabel.shadowOffsetY = NULL,
  legend.emphasis.selectorLabel.width = NULL,
  legend.emphasis.selectorLabel.height = NULL,
  legend.emphasis.selectorLabel.textBorderColor = NULL,
  legend.emphasis.selectorLabel.textBorderWidth = NULL,
  legend.emphasis.selectorLabel.textBorderType = NULL,
  legend.emphasis.selectorLabel.textShadowColor = NULL,
  legend.emphasis.selectorLabel.textShadowBlur = NULL,
  legend.emphasis.selectorLabel.textShadowOffsetX = NULL,
  legend.emphasis.selectorLabel.textShadowOffsetY = NULL
)
}
\arguments{
\item{e}{An echarts4r object}
}
\value{
The modified echarts4r object
}
\description{
Exposes every legend.* option so you don't have to hand-craft the JSON.
}
