% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Bar}
\alias{Plot.Bar}
\title{Plot.Bar}
\usage{
Plot.Bar(
  Engine = "Plotly",
  dt = NULL,
  PreAgg = FALSE,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  AggMethod = "mean",
  Title = "Bar Plot",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{Engine}{'Plotly' or "Echarts"}

\item{dt}{Source data.table}

\item{PreAgg}{logical}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Column name of Group Variable for distinct colored histograms by group levels}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{Title}{title}

\item{EchartsTheme}{"macaron"}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff'}

\item{Debug}{FALSE}

\item{BorderColor}{'darkblue'}
}
\description{
Build a bar plot by simply passing arguments to a single function
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Bar(
  Engine = 'Plotly', # "Echarts"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  data = data,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  AggMethod = 'mean',
  FillColor = "#0066ff",
  ChartColor = 'lightsteelblue1',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  Debug = FALSE)

# # Step through function
# dt <- data
# Engine = 'Plotly' # "Echarts"
# EchartsTheme = "macaron"
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# AggMethod = 'mean'
# GroupVar = NULL
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# FillColor = "#0066ff"
# TextSize = 12
# AngleX = 90
# AngleY = 0
# BackGroundColor = "#6a6969" #, #"#1b1959", #'#00060b',
# ChartColor = '#001534'
# BorderColor = 'darkblue'
# TextColor = 'darkblue'
# GridColor = 'white'
# SubTitleColor = 'blue'
# LegendPosition = 'bottom'
# LegendBorderSize = 0.50
# LegendLineType = 'solid'
# Debug = FALSE
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
