% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.BinaryMetrics}
\alias{Plot.BinaryMetrics}
\title{Plot.BinaryMetrics}
\usage{
Plot.BinaryMetrics(
  dt = NULL,
  PreAgg = FALSE,
  AggMethod = "mean",
  SampleSize = 100000L,
  XVar = NULL,
  YVar = NULL,
  ZVar = NULL,
  Metrics = c("Utility", "MCC", "Accuracy", "F1_Score", "F2_Score", "F0.5_Score",
    "ThreatScore", "TPR", "TNR", "FNR", "FPR", "FDR", "FOR"),
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  ZVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  CostMatrixWeights = c(0, 1, 1, 0),
  NumberBins = 20,
  Height = "600px",
  Width = "1135px",
  Title = "Binary Metrics",
  ShowLabels = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  Engine = "Plotly",
  EchartsTheme = "macarons",
  EchartsLabels = FALSE,
  TimeLine = TRUE,
  X_Scroll = TRUE,
  Y_Scroll = FALSE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{PreAgg}{logical}

\item{AggMethod}{character}

\item{SampleSize}{numeric}

\item{XVar}{character}

\item{YVar}{character}

\item{ZVar}{character}

\item{Metrics}{Multiple selection "Utility","MCC","Accuracy","F1_Score","F2_Score","F0.5_Score","ThreatScore","TPR","TNR","FNR","FPR","FDR","FOR"}

\item{GroupVar}{character}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{ZVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{NumberBins}{numeric}

\item{Title}{character}

\item{ShowLabels}{character}

\item{Title.YAxis}{character}

\item{Title.XAxis}{character}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{EchartsLabels}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{hex character}

\item{ChartColor}{hex character}

\item{FillColor}{hex character}

\item{FillColorReverse}{hex character}

\item{GridColor}{hex character}

\item{TextColor}{hex character}

\item{ZeroLineColor}{hex character}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}
}
\value{
Partial dependence calibration plot
}
\description{
Line plot of evaluation metrics across thresholds
}
\examples{
\dontrun{
# Query postgres
data <- AutoPlots::DM.pgQuery(
  Host = 'localhost',
  DataBase = 'KompsProcessed',
  SELECT = c('ARTICLE','BRAND','CHILLED_Liters_PerDay','CHILLED_Margin_PerDay','CHILLED_Net_Revenue_PerDay','CHILLED_Units_PerDay','CUSTOMER_COD_char','DATE_ISO'),
  AggStat = 'AVG',
  FROM = 'POS_Processed_Long_Daily_backward',
  GroupBy = NULL,
  SamplePercent = 1,
  User = 'postgres',
  Port = 5432,
  Password = 'Aa')

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# PreAgg = FALSE
# dt = data
# SampleSize = 100000L
# PlotEngineType = Engine =  "Echarts" # "Plotly"
# TimeLine = FALSE
# EchartsLabels = FALSE
# EchartsTheme = "purple-passion"
# XVar = "CHILLED_Liters_PerDay" # "Predict"
# YVar = "CHILLED_Margin_PerDay"
# ZVar = "CHILLED_Units_PerDay"
# GroupVar = NULL
# AggMethod = 'mean'
# GroupVar = NULL # "BRAND"
# NumberBins = 20
# AggStat = "mean"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# X_Scroll = TRUE,
# Y_Scroll = FALSE,
# FillColor = "#0066ff"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# Debug = FALSE
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.HeatMap}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.ShapImportance}()},
\code{\link{Plot.VariableImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
