% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Calibration.Box}
\alias{Plot.Calibration.Box}
\title{Plot.Calibration.Box}
\usage{
Plot.Calibration.Box(
  Engine = "Echarts",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  dt = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  NumberBins = 21,
  Title = "Calibration Plot",
  SampleSize = 100000L,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  GridColor = "white",
  TextColor = "white",
  AggMethod = "mean",
  ZeroLineColor = "#ffff",
  FillColor = "#0066ff",
  FillColorReverse = "#ff9900",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{dt}{data.table}

\item{XVar}{character}

\item{YVar}{character}

\item{GroupVar}{character}

\item{NumberBins}{numeric}

\item{Title}{character}

\item{SampleSize}{numeric}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{GridColor}{character hex}

\item{TextColor}{"Not Implemented"}

\item{AggMethod}{character}

\item{ZeroLineColor}{character hex}

\item{FillColor}{character hex}

\item{FillColorReverse}{character hex}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}
}
\value{
Calibration plot or boxplot
}
\description{
This function automatically builds calibration plots and calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.Calibration.Box(
  dt = data,
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  XVar = "Predict",
  YVar = "Adrian",
  AggMethod = 'mean',
  GroupVar = "Factor_1",
  NumberBins = 20,
  AggStat = "mean",
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Title = 'Bar Plot',
  SampleSize = 100000,
  FillColor = "#0066ff",
  FillColorReverse = "#ff9900",
  BackGroundColor = "#6a6969",
  ChartColor = '#001534',
  GridColor = 'white',
  TextColor = 'white',
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# Engine = "Echarts" # "Plotly"
# TimeLine = FALSE
# EchartsTheme = "macaron"
# X_Scroll = TRUE
# Y_Scroll = TRUE
# XVar = "Predict"
# YVar = "Adrian"
# AggMethod = 'mean'
# GroupVar = "Factor_1"
# NumberBins = 20
# AggStat = "mean"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# SampleSize = 100000
# FillColor = "#0066ff"
# FillColorReverse = "#ff9900"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# Debug = FALSE
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{VI_Plot}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
