% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Calibration.Line}
\alias{Plot.Calibration.Line}
\title{Plot.Calibration.Line}
\usage{
Plot.Calibration.Line(
  dt = NULL,
  Engine = "Echarts",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  AggMethod = "mean",
  NumBuckets = 21,
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Title = "Calibration Plot",
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  GridColor = "white",
  TextColor = "white",
  SubTitleColor = "white",
  BorderColor = "white",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{Engine}{"Echarts" or "Plotly"}

\item{EchartsTheme}{Provide an "Echarts" theme}

\item{TimeLine}{TRUE or FALSE}

\item{XVar}{Column name of the predicted values from your model}

\item{YVar}{Column name of the target variable from your model}

\item{GroupVar}{One Grouping Variable}

\item{AggMethod}{"mean", "median", "sd", "skewnewss", "kurtosis", "CoeffVar"}

\item{NumBuckets}{Number of groups for calibration. 21 groups is equivalent to every 5th percentile, since the 0th and 100th percentiles are always included}

\item{ZeroLineColor}{color}

\item{ZeroLineWidth}{1}

\item{Title}{"Title"}

\item{FacetVar1}{"Not Implemented"}

\item{FacetVar2}{"Not Implemented"}

\item{FillColor}{color}

\item{TextSize}{1.25}

\item{AngleX}{"Not Implemented"}

\item{AngleY}{"Not Implemented"}

\item{BackGroundColor}{color}

\item{ChartColor}{color}

\item{GridColor}{color}

\item{TextColor}{"Not Implemented"}

\item{SubTitleColor}{"Not Implemented"}

\item{BorderColor}{"Not Implemented"}

\item{LegendPosition}{"Not Implemented"}

\item{LegendBorderSize}{"Not Implemented"}

\item{LegendLineType}{"Not Implemented"}

\item{Debug}{FALSE}
}
\value{
Calibration plot or boxplot
}
\description{
This function automatically builds calibration plots and calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.Calibration.Line(
  dt = data,
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  AggMethod = 'mean',
  NumBuckets = 20,
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Title = 'Bar Plot',
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  BackGroundColor = "#6a6969", #"#1b1959", #'#00060b',
  ChartColor = '#001534',
  GridColor = 'white',
  TextColor = 'white',
  SubTitleColor = 'white',
  BorderColor = 'white',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# Engine = "Echarts" # "Plotly"
# TimeLine = FALSE
# EchartsTheme = "macaron"
# XVar = "Predict"
# YVar = "Adrian"
# AggMethod = 'mean'
# GroupVar = "Factor_1"
# NumBuckets = 20
# AggStat = "mean"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# FacetVar1 = NULL
# FacetVar2 = NULL
# FillColor = "#0066ff"
# TextSize = 12
# AngleX = 90
# AngleY = 0
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# SubTitleColor = 'white'
# BorderColor = 'white'
# LegendPosition = 'bottom'
# LegendBorderSize = 0.50
# LegendLineType = 'solid'
# Debug = FALSE
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{PlotREsim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
