% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_plots.R
\name{Plot.ClusterSizes}
\alias{Plot.ClusterSizes}
\title{Plot cluster sizes}
\usage{
Plot.ClusterSizes(
  cluster_sizes,
  Title = "Cluster sizes",
  EchartsTheme = "macarons",
  ...
)
}
\arguments{
\item{cluster_sizes}{Named or unnamed cluster counts.}

\item{Title}{Plot title.}

\item{EchartsTheme}{AutoPlots theme.}

\item{...}{Additional arguments passed to [Bar()].}
}
\value{
An `echarts4r` htmlwidget.
}
\description{
Plot cluster sizes
}
