% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.ConfusionMatrix}
\alias{Plot.ConfusionMatrix}
\title{Plot.ConfusionMatrix}
\usage{
Plot.ConfusionMatrix(
  dt = NULL,
  PreAgg = FALSE,
  XVar = NULL,
  YVar = NULL,
  ZVar = "N",
  PercentileBuckets_Y = 0.1,
  PercentileBuckets_X = 0.1,
  NLevels_Y = 50,
  NLevels_X = 50,
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  Title = "Confusion Matrix",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  AggMethod = "count",
  GroupVar = NULL,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{PreAgg}{FALSE}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{ZVar}{= "N"}

\item{PercentileBuckets_Y}{= 0.10,}

\item{PercentileBuckets_X}{= 0.10,}

\item{NLevels_Y}{= NumLevels_Y,}

\item{NLevels_X}{= NumLevels_X,}

\item{Engine}{'Plotly' or "Echarts"}

\item{EchartsTheme}{"macaron"}

\item{TimeLine}{logical}

\item{Title}{title}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff'}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{GroupVar}{= NULL}

\item{Debug}{FALSE}

\item{Algo}{'catboost', 'xgboost', 'h2o'}

\item{BorderColor}{'darkblue'}
}
\value{
ROC Plot for classification models
}
\description{
Generate variable importance plots
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML2_ScoringData.csv")
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_Test_VI_Data.csv")

# Step Through Function
library(AutoPlots)
library(data.table)
dt = data
Engine =  "Plotly" #@ "Echarts"
TimeLine = FALSE
PreAgg = FALSE
EchartsTheme = "purple-passion"
XVar = "Importance" # "Predict"
YVar = "Variable" # "CHILLED_Margin_PerDay"
AggMethod = 'mean'
GroupVar = NULL # "BRAND"
NumBuckets = 20
AggStat = "mean"
ZeroLineColor = '#ffff'
ZeroLineWidth = 1.25
Title = 'Bar Plot'
FillColor = "#0066ff"
BackGroundColor = "#6a6969"
ChartColor = '#001534'
GridColor = 'white'
TextColor = 'white'
Debug = FALSE

AutoPlots:::VI_Plot(
# Algo = "catboost",
# dt = data,
# PreAgg = FALSE
# Engine = 'Plotly',
# EchartsTheme = "macaron",
# TimeLine = TRUE,
# XVar = "Importance",
# YVar = "Variable",
# ZVar = "N"
# PercentileBuckets_Y = 0.10,
# PercentileBuckets_X = 0.10,
# NLevels_Y = NumLevels_Y,
# NLevels_X = NumLevels_X,
# GroupVar = NULL,
# AggMethod = 'mean',
# Title = 'Variable Importance Plot',
# BackGroundColor = "#6a6969",
# ChartColor = '#001534',
# FillColor = "#0066ff",
# GridColor = 'white',
# TextColor = 'white',
# ZeroLineColor = '#ffff',
# ZeroLineWidth = 1.25,
# Debug = FALSE)
}

}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{VI_Plot}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
