% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Copula}
\alias{Plot.Copula}
\title{Plot.Copula}
\usage{
Plot.Copula(
  Engine = "Plotly",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  X_Index = TRUE,
  Y_Index = TRUE,
  data = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Copula Plot",
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  SampleSize = 50000,
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  GridColor = "#ffff",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  TextColor = "white",
  SubTitleColor = "white",
  BorderColor = "#ffff",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{Engine}{= "Plotly",}

\item{EchartsTheme}{= "dark-blue",}

\item{TimeLine}{Logical}

\item{X_Index}{logical}

\item{Y_Index}{logical}

\item{data}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Requires an XVar and YVar already be defined}

\item{Title}{'Violin Plot'}

\item{FacetVar1}{Column name of facet variable 1. If NULL then ignored}

\item{FacetVar2}{Column name of facet variable 2. If NULL then ignored}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{FillColor}{'gray'}

\item{TextSize}{14}

\item{AngleX}{90}

\item{AngleY}{0}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{GridColor}{'white'}

\item{ZeroLineColor}{= '#ffff',}

\item{ZeroLineWidth}{= 2,}

\item{TextColor}{'darkblue'}

\item{SubTitleColor}{'darkblue'}

\item{BorderColor}{'darkblue'}

\item{LegendPosition}{'bottom'}

\item{LegendBorderSize}{0.50}

\item{LegendLineType}{'solid'}

\item{Debug}{FALSE}
}
\description{
Build a copula plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoQuant)
library(data.table)
library(echarts4r)

# Load data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Scatter(
  TimeLine = FALSE,
  X_Index = TRUE,
   Y_Index = TRUE,
  data = data,
  XVar = 'CHILLED_Units_PerDay',
  YVar = 'CHILLED_Margin_PerDay',
  GroupVar = 'BRAND',
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  SampleSize = 100000,
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = 'lightsteelblue1',
  BorderColor = 'darkblue',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  SubTitleColor = 'blue',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  Debug = FALSE)

# Step through function
data# = data |> dplyr::group_by(Region)
Engine = "Echarts"
EchartsTheme = "macaron"
TimeLine = FALSE
XVar = 'CHILLED_Units_PerDay'
YVar = 'CHILLED_Margin_PerDay'
GroupVar = 'Store'
SampleSize = 100000
FillColor = "#0066ff"
TextSize = 12
AngleX = 90
AngleY = 0
ChartColor = 'lightsteelblue1'
BorderColor = 'darkblue'
TextColor = 'darkblue'
GridColor = 'white'
BackGroundColor = 'gray95'
SubTitleColor = 'blue'
LegendPosition = 'bottom'
LegendBorderSize = 0.50
LegendLineType = 'solid'
Debug = FALSE
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
