% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Copula}
\alias{Plot.Copula}
\title{Plot.Copula}
\usage{
Plot.Copula(
  dt = NULL,
  SampleSize = 30000L,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Copula Plot",
  Engine = "Plotly",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  yaxis.fontSize = 14,
  xaxis.fontSize = 14,
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = "#63aeff",
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Requires an XVar and YVar already be defined}

\item{Title}{'Violin Plot'}

\item{Engine}{= "Plotly",}

\item{EchartsTheme}{= "dark-blue",}

\item{TimeLine}{Logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{FillColorReverse}{character}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff',}

\item{ZeroLineWidth}{= 2,}

\item{Debug}{FALSE}
}
\description{
Build a copula plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoQuant)
library(data.table)
library(echarts4r)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Scatter(
  dt = data,
  SampleSize = 100000,
  XVar = 'CHILLED_Units_PerDay',
  YVar = 'CHILLED_Margin_PerDay',
  GroupVar = 'BRAND',
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = 'gray95',
  ChartColor = 'lightsteelblue1',
  FillColor = "#0066ff",
  GridColor = 'white',
  TextColor = 'darkblue',
  Debug = FALSE)

# Step through function
data# = data |> dplyr::group_by(Region)
Engine = "Echarts"
EchartsTheme = "macaron"
TimeLine = FALSE
XVar = 'CHILLED_Units_PerDay'
YVar = 'CHILLED_Margin_PerDay'
GroupVar = 'Store'
SampleSize = 100000
BackGroundColor = 'gray95'
ChartColor = 'lightsteelblue1'
FillColor = "#0066ff"
TextColor = 'darkblue'
GridColor = 'white'
Debug = FALSE
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
