% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.CorrMatrix}
\alias{Plot.CorrMatrix}
\title{Plot.CorrMatrix}
\usage{
Plot.CorrMatrix(
  dt = NULL,
  CorrVars = NULL,
  Method = "spearman",
  PreAgg = FALSE,
  Title = "Correlation Matrix",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{CorrVars}{character}

\item{Method}{character}

\item{PreAgg}{logical}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{= "macaron"}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{FillColorReverse}{character}

\item{GridColor}{character hex}

\item{TextColor}{character hex}

\item{Debug}{logical}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
