% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.CorrMatrix}
\alias{Plot.CorrMatrix}
\title{Plot.CorrMatrix}
\usage{
Plot.CorrMatrix(
  Engine = "Plotly",
  EchartsTheme = "macaron",
  X_Index = TRUE,
  Y_Index = TRUE,
  data = NULL,
  CorrVars = NULL,
  Method = "spearman",
  PreAgg = FALSE,
  Title = "Correlation Matrix",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{Engine}{= "Plotly"}

\item{EchartsTheme}{= "macaron"}

\item{X_Index}{logical}

\item{Y_Index}{logical}

\item{data}{Source data.table}

\item{CorrVars}{Column names of variables you want included in the correlation matrix}

\item{Method}{'spearman' default, 'pearson' otherwise}

\item{PreAgg}{logical}

\item{Title}{"Correlation Matrix"}

\item{BackGroundColor}{= "#6a6969"}

\item{ChartColor}{= '#001534'}

\item{FillColor}{= "#0066ff"}

\item{GridColor}{= 'white'}

\item{TextColor}{= 'white'}

\item{Debug}{= FALSE}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")
CorrVars <- c('Rolling_365_Liters','Rolling_365_Units','Rolling_365_Margin')
p <- cor(data[, .SD, .SDcols = c(CorrVars)])
p1 <- heatmaply::heatmaply_cor(
  p,
  colors = c('red', 'white', 'blue'),
  xlab = "Features",
  ylab = "Features",
  k_col = 2,
  k_row = 2)
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
