% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Density}
\alias{Plot.Density}
\title{Plot.Density}
\usage{
Plot.Density(
  Engine = "Plotly",
  EchartsTheme = "macarons",
  TimeLine = FALSE,
  X_Index = TRUE,
  Y_Index = TRUE,
  data = NULL,
  MeasureVars = NULL,
  GroupVariables = NULL,
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  BackGroundColor = "#17108db8",
  ChartColor = "#001534",
  GridColor = "white",
  TextColor = "white",
  SubTitleColor = "white",
  BorderColor = "white",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{Engine}{= "Plotly",}

\item{EchartsTheme}{= "macarons",}

\item{TimeLine}{logical}

\item{X_Index}{logical}

\item{Y_Index}{logical}

\item{data}{From App}

\item{GroupVariables}{From App}

\item{TextSize}{From App}

\item{AngleX}{From App}

\item{AngleY}{From App}

\item{BackGroundColor}{From App}

\item{ChartColor}{From App}

\item{GridColor}{From App}

\item{TextColor}{From App}

\item{SubTitleColor}{From App}

\item{BorderColor}{From App}

\item{LegendPosition}{From App}

\item{LegendBorderSize}{From App}

\item{LegendLineType}{From App}

\item{Debug}{From App}

\item{MeasureVariables}{From App}
}
\description{
Density plots, by groups, with transparent continuous plots
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")

# Run function
p1 <- AutoPlots:::Plot.Density(
  TimeLine = FALSE,
  data = data,
  MeasureVars = NULL,
  GroupVariables = 'Weekly_Sales',
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = 'lightsteelblue1',
  BorderColor = 'darkblue',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  SubTitleColor = 'blue',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  Debug = FALSE)

# Step through function
# # plotly::ggplotly(p1)
# TimeLine = FALSE
# Engine = "Plotly"
# EchartsTheme = "macarons"
# data = data
# MeasureVars = 'Weekly_Sales'
# GroupVariables = c('Store','Dept')
# TextSize = 12
# AngleX = 90
# AngleY = 0
# ChartColor = 'lightsteelblue1'
# BorderColor = 'darkblue'
# TextColor = 'darkblue'
# GridColor = 'white'
# BackGroundColor = 'gray95'
# SubTitleColor = 'blue'
# LegendPosition = 'bottom'
# LegendBorderSize = 0.50
# LegendLineType = 'solid'
# Debug = FALSE
# Bins
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\concept{Graphics}
