% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Density}
\alias{Plot.Density}
\title{Plot.Density}
\usage{
Plot.Density(
  dt = NULL,
  SampleSize = 100000L,
  YVar = NULL,
  XVar = NULL,
  GroupVar = NULL,
  Title = "Density Plot",
  Engine = "Plotly",
  EchartsTheme = "macarons",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{From App}

\item{SampleSize}{= 100000L}

\item{YVar}{From App}

\item{XVar}{character}

\item{GroupVar}{character}

\item{Title}{= "Density Plot"}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{"#6a6969",}

\item{ChartColor}{"#001534",}

\item{FillColor}{"#0066ff",}

\item{FillColorReverse}{"#97ff00",}

\item{GridColor}{"white",}

\item{TextColor}{"white",}

\item{Debug}{From App}
}
\description{
Density plots, by groups, with transparent continuous plots
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
p1 <- AutoPlots:::Plot.Density(
  TimeLine = FALSE,
  data = data,
  YVar = NULL,
  GroupVar = 'Weekly_Sales',
  ChartColor = 'lightsteelblue1',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  Debug = FALSE)

# Step through function
# # plotly::ggplotly(p1)
# TimeLine = FALSE
# Engine = "Plotly"
# EchartsTheme = "macarons"
# data = data
# YVar = 'Weekly_Sales'
# GroupVar = c('Store','Dept')
# ChartColor = 'lightsteelblue1'
# TextColor = 'darkblue'
# GridColor = 'white'
# BackGroundColor = 'gray95'
# Debug = FALSE
# NumberBins
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\concept{Standard Plots}
