% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_plots.R
\name{Plot.ExplainedVariance}
\alias{Plot.ExplainedVariance}
\title{Plot component explained variance}
\usage{
Plot.ExplainedVariance(
  explained_variance,
  cumulative = NULL,
  Title = "Explained variance",
  EchartsTheme = "macarons",
  ...
)
}
\arguments{
\item{explained_variance}{Numeric explained-variance proportions.}

\item{cumulative}{Optional cumulative proportions.}

\item{Title}{Plot title.}

\item{EchartsTheme}{AutoPlots theme.}

\item{...}{Additional arguments passed to [Plot.Bar()].}
}
\value{
An `echarts4r` htmlwidget.
}
\description{
Plot component explained variance
}
