% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Gains}
\alias{Plot.Gains}
\title{Plot.Gains}
\usage{
Plot.Gains(
  dt = NULL,
  PreAgg = FALSE,
  XVar = NULL,
  YVar = NULL,
  ZVar = "N",
  GroupVar = NULL,
  NumberBins = 20,
  Title = "Gains Plot",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{PreAgg}{logical}

\item{XVar}{character}

\item{YVar}{character}

\item{ZVar}{character}

\item{GroupVar}{character}

\item{NumberBins}{numeric}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{FillColorReverse}{character hex}

\item{GridColor}{character hex}

\item{TextColor}{character hex}

\item{ZeroLineColor}{character hex}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}
}
\description{
Create a cumulative gains chart
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.BinaryMetrics}()},
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Heatmap}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.ShapImportance}()},
\code{\link{Plot.VariableImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
