% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Histogram}
\alias{Plot.Histogram}
\title{Plot.Histogram}
\usage{
Plot.Histogram(
  dt = NULL,
  SampleSize = 30000L,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  NumberBins = 30,
  Height = "600px",
  Width = "1135px",
  Title = "Histogram",
  Engine = "Plotly",
  EchartsTheme = "macarons",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineWidth = 1.25,
  ZeroLineColor = "white",
  title.fontSize = 22,
  title.fontWeight = "bold",
  title.textShadowColor = "#63aeff",
  title.textShadowBlur = 3,
  title.textShadowOffsetY = 1,
  title.textShadowOffsetX = -1,
  xaxis.fontSize = 14,
  yaxis.fontSize = 14,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Column name of Group Variable for distinct colored histograms by group levels}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{NumberBins}{= 30}

\item{Height}{= NULL,}

\item{Width}{= NULL,}

\item{Engine}{= PlotEngineType,}

\item{EchartsTheme}{= EchartsTheme,}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{FillColorReverse}{character}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineWidth}{= 1.25,}

\item{ZeroLineColor}{= "white",}

\item{Debug}{FALSE}
}
\description{
Build a histogram plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
p1 <- AutoPlots:::Plot.Histogram(
  dt = data,
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  SampleSize = 100000,
  NumberBins = 20,
  FillColor = "#0066ff",
  ChartColor = 'lightsteelblue1',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  Debug = FALSE)

# # Step through function
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# TimeLine = FALSE
# AggMethod = 'mean'
# ZeroLineWidth = 1.25
# Title = 'Histogram'
# NumberBins = 20
# SampleSize = 100000
# FillColor = "#0066ff"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'darkblue'
# Debug = FALSE
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Violin}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
