% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Histogram}
\alias{Plot.Histogram}
\title{Plot.Histogram}
\usage{
Plot.Histogram(
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Index = TRUE,
  Y_Index = TRUE,
  data = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Histogram",
  SampleSize = 1e+05,
  Bins = 30,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineWidth = 1.25,
  ZeroLineColor = "white",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  SubTitleColor = "white",
  BorderColor = "white",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{Engine}{= PlotEngineType,}

\item{EchartsTheme}{= EchartsTheme,}

\item{TimeLine}{logical}

\item{X_Index}{logical}

\item{Y_Index}{logical}

\item{data}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Column name of Group Variable for distinct colored histograms by group levels}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{Bins}{= 30}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineWidth}{= 1.25,}

\item{ZeroLineColor}{= "white",}

\item{TextSize}{14}

\item{AngleX}{90}

\item{AngleY}{0}

\item{FacetVar1}{Column name of facet variable 1. If NULL then ignored}

\item{FacetVar2}{Column name of facet variable 2. If NULL then ignored}

\item{SubTitleColor}{'darkblue'}

\item{BorderColor}{'darkblue'}

\item{LegendPosition}{'bottom'}

\item{LegendBorderSize}{0.50}

\item{LegendLineType}{'solid'}

\item{Debug}{FALSE}
}
\description{
Build a histogram plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoPlots)
library(data.table)

# Load data
data <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML1_ScoringData.csv")

# Run function
p1 <- AutoPlots:::Plot.Histogram(
  data = data,
  TimeLine = FALSE,
  X_Index = TRUE,
  Y_Index = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  FacetVar1 = 'Store',
  FacetVar2 = 'Dept',
  SampleSize = 100000,
  Bins = 20,
  FillColor = "#0066ff",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = 'lightsteelblue1',
  BorderColor = 'darkblue',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  SubTitleColor = 'blue',
  LegendPosition = 'bottom',
  LegendBorderSize = 0.50,
  LegendLineType = 'solid',
  Debug = FALSE)

# # Step through function
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay"
# GroupVar = "BRAND"
# TimeLine = FALSE
# AggMethod = 'mean'
# ZeroLineWidth = 1.25
# Title = 'Histogram'
# FacetVar1 = NULL
# FacetVar2 = NULL
# Bins = 20
# SampleSize = 100000
# FillColor = "#0066ff"
# TextSize = 12
# AngleX = 90
# AngleY = 0
# ChartColor = '#001534'
# BackGroundColor = "#6a6969"
# GridColor = 'white'
# BorderColor = 'darkblue'
# TextColor = 'darkblue'
# SubTitleColor = 'blue'
# LegendPosition = 'bottom'
# LegendBorderSize = 0.50
# LegendLineType = 'solid'
# Debug = FALSE
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
