% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Lift}
\alias{Plot.Lift}
\title{Plot.Lift}
\usage{
Plot.Lift(
  dt = NULL,
  PreAgg = FALSE,
  XVar = NULL,
  YVar = NULL,
  ZVar = "N",
  NumBuckets = 20,
  PercentileBuckets_Y = 0.1,
  PercentileBuckets_X = 0.1,
  NLevels_Y = 50,
  NLevels_X = 50,
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  Title = "Confusion Matrix",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  GroupVar = NULL,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  SubTitleColor = "white",
  BorderColor = "white",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{dt}{= NULL}

\item{PreAgg}{= FALSE}

\item{XVar}{= NULL}

\item{YVar}{= NULL}

\item{ZVar}{= "N"}

\item{NumBuckets}{= 20}

\item{PercentileBuckets_Y}{= 0.10}

\item{PercentileBuckets_X}{= 0.10}

\item{NLevels_Y}{= 50}

\item{NLevels_X}{= 50}

\item{Engine}{= 'Plotly'}

\item{EchartsTheme}{= "macaron"}

\item{TimeLine}{= TRUE}

\item{Title}{= "Confusion Matrix"}

\item{BackGroundColor}{= "#6a6969"}

\item{ChartColor}{= '#001534'}

\item{FillColor}{= "#0066ff"}

\item{GridColor}{= 'white'}

\item{TextColor}{= 'white'}

\item{ZeroLineColor}{= '#ffff'}

\item{ZeroLineWidth}{= 1.25}

\item{GroupVar}{= NULL}

\item{FacetVar1}{= NULL}

\item{FacetVar2}{= NULL}

\item{TextSize}{= 12}

\item{AngleX}{= 90}

\item{AngleY}{= 0}

\item{SubTitleColor}{= 'white'}

\item{BorderColor}{= 'white'}

\item{LegendPosition}{= 'bottom'}

\item{LegendBorderSize}{= 0.50}

\item{LegendLineType}{= 'solid'}

\item{Debug}{= FALSE}
}
\description{
Create a cumulative gains chart
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{VI_Plot}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
