% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_plots.R
\name{Plot.NeighborDistances}
\alias{Plot.NeighborDistances}
\title{Plot ordered nearest-neighbor distances}
\usage{
Plot.NeighborDistances(
  distances,
  Title = "Neighbor distances",
  EchartsTheme = "macarons",
  ...
)
}
\arguments{
\item{distances}{Numeric neighbor distances.}

\item{Title}{Plot title.}

\item{EchartsTheme}{AutoPlots theme.}

\item{...}{Additional arguments passed to [Line()].}
}
\value{
An `echarts4r` htmlwidget.
}
\description{
Plot ordered nearest-neighbor distances
}
