% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.PartialDependence.Line}
\alias{Plot.PartialDependence.Line}
\title{Plot.PartialDependence.Line}
\usage{
Plot.PartialDependence.Line(
  dt = NULL,
  Engine = "Plotly",
  EchartsTheme = "macaron",
  EchartsLabels = FALSE,
  TimeLine = TRUE,
  XVar = NULL,
  YVar = NULL,
  ZVar = NULL,
  GroupVar = NULL,
  NumBuckets = 20,
  PreAgg = FALSE,
  AggMethod = "mean",
  Title = "Gains Plot",
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{EchartsLabels}{character}

\item{TimeLine}{logical}

\item{XVar}{character}

\item{YVar}{character}

\item{ZVar}{character}

\item{GroupVar}{character}

\item{NumBuckets}{numeric}

\item{PreAgg}{logical}

\item{AggMethod}{character}

\item{Title}{character}

\item{BackGroundColor}{hex character}

\item{ChartColor}{hex character}

\item{FillColor}{hex character}

\item{GridColor}{hex character}

\item{TextColor}{hex character}

\item{ZeroLineColor}{hex character}

\item{ZeroLineWidth}{numeric}

\item{Debug}{= FALSE}

\item{FillColorReverse}{hex character}
}
\value{
Partial dependence calibration plot
}
\description{
This function automatically builds partial dependence calibration plots
}
\examples{
\dontrun{
# Query postgres
data <- AutoPlots::DM.pgQuery(
  Host = 'localhost',
  DataBase = 'KompsProcessed',
  SELECT = c('ARTICLE','BRAND','CHILLED_Liters_PerDay','CHILLED_Margin_PerDay','CHILLED_Net_Revenue_PerDay','CHILLED_Units_PerDay','CUSTOMER_COD_char','DATE_ISO'),
  AggStat = 'AVG',
  FROM = 'POS_Processed_Long_Daily_backward',
  GroupBy = NULL,
  SamplePercent = 1,
  User = 'postgres',
  Port = 5432,
  Password = 'Aa1028#@')

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# PreAgg = FALSE
# dt = data
# SampleSize = 100000L
# PlotEngineType = Engine =  "Echarts" # "Plotly"
# TimeLine = FALSE
# EchartsLabels = FALSE
# EchartsTheme = "purple-passion"
# XVar = "CHILLED_Liters_PerDay" # "Predict"
# YVar = "CHILLED_Margin_PerDay"
# ZVar = "CHILLED_Units_PerDay"
# GroupVar = NULL
# AggMethod = 'mean'
# GroupVar = NULL # "BRAND"
# NumBuckets = 20
# AggStat = "mean"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# FillColor = "#0066ff"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# Debug = FALSE
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{PlotREsim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
