% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Residuals.Histogram}
\alias{Plot.Residuals.Histogram}
\title{Plot.Residuals.Histogram}
\usage{
Plot.Residuals.Histogram(
  dt = NULL,
  SampleSize = 1e+05,
  AggMethod = "mean",
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  NumberBins = 20,
  Title = "Calibration Plot",
  Engine = "Echarts",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{SampleSize}{numeric}

\item{AggMethod}{character}

\item{XVar}{character}

\item{YVar}{character}

\item{GroupVar}{character}

\item{NumberBins}{numeric}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{GridColor}{character hex}

\item{TextColor}{Not Implemented}

\item{ZeroLineColor}{character hex}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}
}
\description{
Residuals Plot
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.Residuals.Histogram(
  dt = data,
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  AggMethod = 'mean',
  GroupVar = "BRAND",
  NumberBins = 20,
  AggStat = "mean",
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Title = 'Bar Plot',
  SampleSize = 100000,
  FillColor = "#0066ff",
  BackGroundColor = "#6a6969", #"#1b1959", #'#00060b',
  ChartColor = '#001534',
  GridColor = 'white',
  TextColor = 'white',
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# Engine = "Echarts" # "Plotly"
# TimeLine = FALSE
# EchartsTheme = "macaron"
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay",
# AggMethod = 'mean',
# GroupVar = "BRAND",
# NumberBins = 20
# AggStat = "mean"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# SampleSize = 100000
# FillColor = "#0066ff"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# Debug = FALSE

# dt = dt1
# NumberBins = NumberBins
# Engine = Engine
# EchartsTheme = EchartsTheme
# TimeLine = tl
# YVar = "Target - Predicted"
# XVar = XVar
# GroupVar = GroupVar
# ZeroLineColor = ZeroLineColor
# ZeroLineWidth = ZeroLineWidth
# Title = Title
# SampleSize = SampleSize
# FillColor = FillColor
# BackGroundColor = BackGroundColor
# ChartColor = ChartColor
# GridColor = GridColor
# TextColor = GridColor
# Debug = Debug
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{VI_Plot}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
