% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Residuals.Histogram}
\alias{Plot.Residuals.Histogram}
\title{Plot.Residuals.Histogram}
\usage{
Plot.Residuals.Histogram(
  dt = NULL,
  AggMethod = "mean",
  SampleSize = 1e+05,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  NumberBins = 20,
  Title = "Calibration Plot",
  Engine = "Echarts",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{AggMethod}{character}

\item{SampleSize}{numeric}

\item{XVar}{character}

\item{YVar}{character}

\item{GroupVar}{character}

\item{NumberBins}{numeric}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{FillColorReverse}{character hex}

\item{GridColor}{character hex}

\item{TextColor}{Not Implemented}

\item{ZeroLineColor}{character hex}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}
}
\description{
Residuals Plot
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.Residuals.Histogram(
  dt = data,
  AggMethod = 'mean',
  SampleSize = 100000,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  GroupVar = "BRAND",
  NumberBins = 20,
  Title = 'Bar Plot',
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  BackGroundColor = "#6a6969", #"#1b1959", #'#00060b',
  ChartColor = '#001534',
  FillColor = "#0066ff",
  GridColor = 'white',
  TextColor = 'white',
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# AggMethod = 'mean',
# SampleSize = 100000
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay",
# GroupVar = "BRAND",
# NumberBins = 20
# Title = 'Bar Plot'
# Engine = "Echarts" # "Plotly"
# EchartsTheme = "macaron"
# TimeLine = FALSE
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# FillColor = "#0066ff"
# GridColor = 'white'
# TextColor = 'white'
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Debug = FALSE
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.BinaryMetrics}()},
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Heatmap}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.ShapImportance}()},
\code{\link{Plot.VariableImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
