% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.ROC}
\alias{Plot.ROC}
\title{Plot.ROC}
\usage{
Plot.ROC(
  dt = NULL,
  SampleSize = 1e+05,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  AggMethod = "mean",
  Height = "600px",
  Width = "1135px",
  Title = "Calibration Plot",
  Engine = "Echarts",
  EchartsTheme = "macarons",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{SampleSize}{numeric}

\item{XVar}{character}

\item{YVar}{character}

\item{GroupVar}{character}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{AggMethod}{character}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hex}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{FillColorReverse}{character hex}

\item{GridColor}{character hex}

\item{TextColor}{character hex}

\item{ZeroLineColor}{character hex}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}

\item{NumberBins}{numeric}
}
\value{
Calibration plot or boxplot
}
\description{
ROC Plot
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Build Plot
AutoPlots::Plot.ROC(
  dt = data,
  Engine = "Echarts", # "Plotly"
  EchartsTheme = "macarons",
  TimeLine = TRUE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  XVar = "Predict",
  YVar = "CHILLED_Margin_PerDay",
  AggMethod = 'mean',
  GroupVar = "BRAND",
  NumberBins = 20,
  AggStat = "mean",
  ZeroLineColor = '#ffff',
  ZeroLineWidth = 1.25,
  Title = 'Bar Plot',
  SampleSize = 100000,
  FillColor = "#0066ff",
  BackGroundColor = "#6a6969", #"#1b1959", #'#00060b',
  ChartColor = '#001534',
  GridColor = 'white',
  TextColor = 'white',
  Debug = FALSE)

# # Step Through Function
# library(AutoPlots)
# library(data.table)
# dt = data
# Engine = "Echarts" # "Plotly"
# TimeLine = FALSE
# X_Scroll = TRUE,
# Y_Scroll = TRUE,
# EchartsTheme = "macaron"
# XVar = "Predict"
# YVar = "CHILLED_Margin_PerDay",
# AggMethod = 'mean',
# GroupVar = "BRAND",
# NumberBins = 20
# AggStat = "mean"
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Title = 'Bar Plot'
# SampleSize = 100000
# FillColor = "#0066ff"
# BackGroundColor = "#6a6969"
# ChartColor = '#001534'
# GridColor = 'white'
# TextColor = 'white'
# Debug = FALSE
}
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.BinaryMetrics}()},
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.HeatMap}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.ShapImportance}()},
\code{\link{Plot.VariableImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
