% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Scatter}
\alias{Plot.Scatter}
\title{Plot.Scatter}
\usage{
Plot.Scatter(
  dt = NULL,
  SampleSize = 30000L,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Title = "Scatter Plot",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  GridColor = "#ffff",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{SampleSize}{numeric}

\item{XVar}{character}

\item{YVar}{character}

\item{GroupVar}{character}

\item{Title}{character}

\item{Engine}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{character hexb}

\item{ChartColor}{character hex}

\item{FillColor}{character hex}

\item{GridColor}{character hex}

\item{TextColor}{character hex}

\item{ZeroLineColor}{character hex}

\item{ZeroLineWidth}{numeric}

\item{Debug}{logical}

\item{BorderColor}{character hex}
}
\description{
Build a copula plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoQuant)
library(data.table)
library(echarts4r)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Scatter(
  dt = data,
  SampleSize = 100000,
  XVar = 'CHILLED_Units_PerDay',
  YVar = 'CHILLED_Margin_PerDay',
  GroupVar = 'BRAND',
  Engine = "Echarts",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  FillColor = "#0066ff",
  BackGroundColor = 'gray95',
  ChartColor = 'lightsteelblue1',
  BorderColor = 'darkblue',
  TextColor = 'darkblue',
  GridColor = 'white',
  Debug = FALSE)

# Step through function
# dt = data |> dplyr::group_by(Region)
# SampleSize = 100000
# XVar = 'XREG1'
# YVar = 'XREG2'
# GroupVar = 'Store'
# Engine = "Echarts"
# EchartsTheme = "macaron"
# TimeLine = FALSE
# X_Scroll = TRUE
# Y_Scroll = TRUE
# Title = "Bla"
# FillColor = "#0066ff"
# BackGroundColor = 'gray95'
# ChartColor = 'lightsteelblue1'
# BorderColor = 'darkblue'
# TextColor = 'darkblue'
# GridColor = 'white'
# Debug = FALSE
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
