% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Scatter3D}
\alias{Plot.Scatter3D}
\title{Plot.Scatter3D}
\usage{
Plot.Scatter3D(
  dt = NULL,
  SampleSize = 1e+05,
  XVar = NULL,
  YVar = NULL,
  ZVar = NULL,
  GroupVar = NULL,
  Title = "3D Scatter",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  FillColor = "#0066ff",
  BackGroundColor = "#bbbec7",
  ChartColor = "#001534",
  GridColor = "#ffff",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{ZVar}{Column name of Z-Axis variable. If NULL then ignored}

\item{GroupVar}{Requires an XVar and YVar already be defined}

\item{Title}{'Violin Plot'}

\item{Engine}{= "Plotly"}

\item{EchartsTheme}{= "macaron"}

\item{TimeLine}{Logical}

\item{FillColor}{'gray'}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff',}

\item{ZeroLineWidth}{= 2,}

\item{Debug}{FALSE}
}
\description{
Build a 3D-copula plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Load packages
library(AutoQuant)
library(data.table)

# Load data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")

# Run function
AutoPlots:::Plot.Scatter3D(
  dt = data,
  SampleSize = 100000,
  XVar = 'ARTICLE',
  YVar = 'BRAND',
  ZVar = 'CHILLED_Units_PerDay',
  GroupVar = 'Store',
  Title = "Something",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = FALSE,
  FillColor = "#0066ff",
  ChartColor = 'lightsteelblue1',
  TextColor = 'darkblue',
  GridColor = 'white',
  BackGroundColor = 'gray95',
  Debug = FALSE)

# Step through function
# dt = data
# SampleSize = 100000
# XVar = "Predict",
# YVar = "CHILLED_Margin_PerDay",
# GroupVar = "BRAND",
# Title = 'Copula 3D'
# Engine = "Plotly"
# EchartsTheme = "macaron"
# TimeLine = FALSE
# FillColor = "#0066ff"
# BackGroundColor = "#bbbec7"
# ChartColor = '#001534'
# GridColor = '#ffff'
# TextColor = '#FFFFFF'
# ZeroLineColor = '#ffff'
# ZeroLineWidth = 1.25
# Debug = FALSE
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
