% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.ShapImportance}
\alias{Plot.ShapImportance}
\title{Plot.ShapImportance}
\usage{
Plot.ShapImportance(
  dt,
  PreAgg = FALSE,
  AggMethod = "meanabs",
  YVar = NULL,
  GroupVar = NULL,
  NumberBins = 21,
  NumLevels_X = 33,
  NumLevels_Y = 33,
  Title = "Heatmap",
  Engine = "Plotly",
  EchartsTheme = "dark",
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{AggMethod}{"mean", "median", "sum", "sd", "skewness","kurtosis", "coeffvar", "meanabs", "medianabs", "sumabs", "sdabs", "skewnessabs", "kurtosisabs", "CoeffVarabs"}

\item{NumberBins}{= 21}

\item{NumLevels_X}{= 20}

\item{NumLevels_Y}{= 20}

\item{Title}{"Heatmap"}

\item{Engine}{"plotly", "echarts4r"}

\item{EchartsTheme}{"dark-blue"}

\item{BackGroundColor}{= "#6a6969"}

\item{ChartColor}{= '#001534'}

\item{FillColor}{= "#0066ff"}

\item{FillColorReverse}{character hex}

\item{GridColor}{= '#ffffff'}

\item{Debug}{= FALSE}
}
\description{
Plot.ShapImportance variable importance
}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.BinaryMetrics}()},
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.VariableImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
