% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.StandardPlots}
\alias{Plot.StandardPlots}
\title{Plot.StandardPlots}
\usage{
Plot.StandardPlots(
  dt = NULL,
  PreAgg = FALSE,
  PlotType = "Scatter",
  SampleSize = 100000L,
  AggMethod = "mean",
  NumberBins = 30,
  YVar = NULL,
  XVar = NULL,
  ZVar = NULL,
  GroupVar = NULL,
  YVarTrans = NULL,
  XVarTrans = NULL,
  ZVarTrans = NULL,
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  Height = NULL,
  Width = NULL,
  PlotEngineType = "Plotly",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  Title = NULL,
  ShowLabels = FALSE,
  Title.YAxis = NULL,
  Title.XAxis = NULL,
  NumLevels_Y = 75,
  NumLevels_X = 40,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  FontSize = 14,
  Debug = FALSE
)
}
\arguments{
\item{dt}{source data.table}

\item{PreAgg}{FALSE}

\item{PlotType}{character}

\item{SampleSize}{character}

\item{AggMethod}{character}

\item{NumberBins}{For histograms}

\item{YVar}{Y-Axis variable name}

\item{XVar}{X-Axis variable name}

\item{ZVar}{Z-Axis variable name}

\item{GroupVar}{Character variable variable}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{ZVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{Height}{NULL or valid css unit}

\item{Width}{NULL or valid css unit}

\item{PlotEngineType}{"Echarts" or "Plotly"}

\item{EchartsTheme}{"auritus","azul","bee-inspired","blue","caravan","carp","chalk","cool","dark-bold","dark","eduardo", #' "essos","forest","fresh-cut","fruit","gray","green","halloween","helianthus","infographic","inspired", #' "jazz","london","dark","macarons","macarons2","mint","purple-passion","red-velvet","red","roma","royal", #' "sakura","shine","tech-blue","vintage","walden","wef","weforum","westeros","wonderland"}

\item{TimeLine}{character}

\item{Title}{character}

\item{ShowLabels}{character}

\item{Title.YAxis}{character}

\item{Title.XAxis}{character}

\item{BackGroundColor}{color outside of plot window. Rcolors and hex outside of plot window. Rcolors and hex character}

\item{ChartColor}{character}

\item{FillColor}{character}

\item{FillColorReverse}{character}

\item{GridColor}{character}

\item{TextColor}{character}

\item{FontSize}{numeric}

\item{Debug}{Debugging purposes}
}
\description{
Helper for standard plots
}
\seealso{
Other Auto Plotting: 
\code{\link{Plots.ModelEvaluation}()}
}
\author{
Adrian Antico
}
\concept{Auto Plotting}
