% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.StandardPlots}
\alias{Plot.StandardPlots}
\title{Plot.StandardPlots}
\usage{
Plot.StandardPlots(
  dt = NULL,
  PreAgg = FALSE,
  PlotType = "Scatter",
  SampleSize = 100000L,
  YVar = NULL,
  XVar = NULL,
  ZVar = NULL,
  GroupVar = NULL,
  PlotEngineType = "Plotly",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  NumLevels_Y = 75,
  NumLevels_X = 40,
  AggMethod = "mean",
  NumberBins = 30,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{character}

\item{PreAgg}{= FALSE}

\item{PlotType}{character}

\item{SampleSize}{character}

\item{YVar}{character}

\item{XVar}{character}

\item{ZVar}{character}

\item{GroupVar}{character}

\item{PlotEngineType}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{character}

\item{AggMethod}{character}

\item{NumberBins}{character}

\item{BackGroundColor}{character}

\item{ChartColor}{character}

\item{FillColor}{character}

\item{FillColorReverse}{character}

\item{GridColor}{character}

\item{TextColor}{character}

\item{Debug}{character}
}
\description{
Helper for standard plots
}
\seealso{
Other Auto Plotting: 
\code{\link{Plots.ModelEvaluation}()}
}
\author{
Adrian Antico
}
\concept{Auto Plotting}
