% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.StandardPlots}
\alias{Plot.StandardPlots}
\title{Plot.StandardPlots}
\usage{
Plot.StandardPlots(
  dt = NULL,
  PlotType = "Scatter",
  SampleSize = 100000L,
  YVar = NULL,
  XVar = NULL,
  ZVar = NULL,
  GroupVar = NULL,
  PlotEngineType = "Plotly",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  NumLevels_Y = 75,
  NumLevels_X = 40,
  BarPlotAggMethod = "mean",
  NumberBins = 30,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#ff9900",
  GridColor = "white",
  TextColor = "white",
  Debug = FALSE
)
}
\arguments{
\item{dt}{character}

\item{PlotType}{character}

\item{SampleSize}{character}

\item{YVar}{character}

\item{XVar}{character}

\item{ZVar}{character}

\item{GroupVar}{character}

\item{PlotEngineType}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{character}

\item{BarPlotAggMethod}{character}

\item{NumberBins}{character}

\item{BackGroundColor}{character}

\item{ChartColor}{character}

\item{FillColor}{character}

\item{FillColorReverse}{character}

\item{GridColor}{character}

\item{TextColor}{character}

\item{Debug}{character}

\item{SizeVar1}{character}

\item{FacetVar1}{character}

\item{FacetVar2}{character}

\item{AngleX}{character}

\item{AngleY}{character}

\item{BorderColor}{character}
}
\description{
Helper for standard plots
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{Plots.ModelEvaluation}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
