% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Stock}
\alias{Plot.Stock}
\title{Plot.Stock}
\usage{
Plot.Stock(StockDataOutput, Type = "candlestick")
}
\arguments{
\item{StockDataOutput}{PolyOut returned from StockData()}

\item{Type}{'candlestick', 'ohlc'}
}
\description{
Create a candlestick plot for stocks. See https://plotly.com/r/figure-labels/
}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
