% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.VariableImportance}
\alias{Plot.VariableImportance}
\title{Plot.VariableImportance}
\usage{
Plot.VariableImportance(
  Algo = "CatBoost",
  dt = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  AggMethod = "mean",
  Height = "600px",
  Width = "1135px",
  Title = "Variable Importance Plot",
  Engine = "Plotly",
  EchartsTheme = "macaron",
  TimeLine = TRUE,
  X_Scroll = TRUE,
  Y_Scroll = TRUE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{Algo}{'catboost', 'xgboost', 'h2o'}

\item{dt}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Column name of Group Variable for distinct colored histograms by group levels}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{AggMethod}{Choose from 'mean', 'sum', 'sd', and 'median'}

\item{Title}{title}

\item{Engine}{'Plotly' or "Echarts"}

\item{EchartsTheme}{"macaron"}

\item{TimeLine}{logical}

\item{X_Scroll}{logical}

\item{Y_Scroll}{logical}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{FillColorReverse}{character hex}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff'}

\item{Debug}{FALSE}
}
\value{
ROC Plot for classification models
}
\description{
Generate variable importance plots
}
\examples{
\dontrun{
# Create fake data
data <- data.table::fread("./CatBoost_ML1_ScoringData.csv")
data <- data.table::fread("./CatBoost_ML2_ScoringData.csv")
data <- data.table::fread("./CatBoost_ML1_Test_VI_Data.csv")

# Step Through Function
library(AutoPlots)
library(data.table)
dt = data
Engine =  "Plotly"
TimeLine = FALSE
X_Scroll = TRUE,
Y_Scroll = TRUE,
EchartsTheme = "purple-passion"
XVar = "Importance" # "Predict"
YVar = "Variable" # "CHILLED_Margin_PerDay"
AggMethod = 'mean'
GroupVar = NULL # "BRAND"
NumberBins = 20
AggStat = "mean"
ZeroLineColor = '#ffff'
ZeroLineWidth = 1.25
Title = 'Bar Plot'
FillColor = "#0066ff"
BackGroundColor = "#6a6969"
ChartColor = '#001534'
GridColor = 'white'
TextColor = 'white'
Debug = FALSE

Rappture::Plot.VariableImportance(
# Algo = "catboost",
# dt = data,
# Engine = 'Plotly',
# EchartsTheme = "macaron",
# TimeLine = TRUE,
# X_Scroll = TRUE,
# Y_Scroll = TRUE,
# XVar = "Importance",
# YVar = "Variable",
# GroupVar = NULL,
# AggMethod = 'mean',
# Title = 'Variable Importance Plot',
# BackGroundColor = "#6a6969",
# ChartColor = '#001534',
# FillColor = "#0066ff",
# GridColor = 'white',
# TextColor = 'white',
# ZeroLineColor = '#ffff',
# ZeroLineWidth = 1.25,
# Debug = FALSE)
}

}
\seealso{
Other Model Evaluation: 
\code{\link{Plot.BinaryMetrics}()},
\code{\link{Plot.Calibration.Box}()},
\code{\link{Plot.Calibration.Line}()},
\code{\link{Plot.ConfusionMatrix}()},
\code{\link{Plot.Gains}()},
\code{\link{Plot.Lift}()},
\code{\link{Plot.PartialDependence.Box}()},
\code{\link{Plot.PartialDependence.HeatMap}()},
\code{\link{Plot.PartialDependence.Line}()},
\code{\link{Plot.ROC}()},
\code{\link{Plot.Residuals.Histogram}()},
\code{\link{Plot.Residuals.Scatter}()},
\code{\link{Plot.ShapImportance}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
