% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plot.Violin}
\alias{Plot.Violin}
\title{Plot.Violin}
\usage{
Plot.Violin(
  dt = NULL,
  XVar = NULL,
  YVar = NULL,
  GroupVar = NULL,
  Height = "600px",
  Width = "1135px",
  Title = "Violin Plot",
  SampleSize = 1e+05,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  ZeroLineColor = "#ffff",
  ZeroLineWidth = 1.25,
  Debug = FALSE
)
}
\arguments{
\item{dt}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{GroupVar}{Requires an XVar and YVar already be defined}

\item{Height}{= NULL,}

\item{Width}{= NULL,}

\item{Title}{'Violin Plot'}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{BackGroundColor}{'gray95'}

\item{ChartColor}{'lightsteelblue'}

\item{FillColor}{'gray'}

\item{FillColorReverse}{character}

\item{GridColor}{'white'}

\item{TextColor}{'darkblue'}

\item{ZeroLineColor}{= '#ffff',}

\item{ZeroLineWidth}{= 2,}

\item{Debug}{FALSE}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
# Step Through Function
library(Rappture)
library(data.table)
dt <- data.table::fread("C:/Users/Bizon/Documents/GitHub/rappwd/CatBoost_ML_Regression_ScoringData.csv")
PreAgg = TRUE
DataReady = FALSE
PlotEngineType = Engine =  "Echarts" # "Plotly"
TimeLine = FALSE
EchartsTheme = "purple-passion"
GroupVar = NULL# "ARTICLE" # "BRAND" #c("BRAND",'ARTICLE')
GroupVars = NULL# c("BRAND",'ARTICLE')
CorrVars = c("CHILLED_Net_Revenue_PerDay","CHILLED_Margin_PerDay","CHILLED_Liters_PerDay","CHILLED_Units_PerDay")
Method = 'spearman'
NumLevelsDisplay = 50L
AggMethod = 'mean'
NumberBins = 20
ZeroLineColor = '#ffff'
ZeroLineWidth = 1.25
Title = 'Bar Plot'
SampleSize = 100000
FillColor = "#0066ff"
FillColorReverse = "#97ff00"
BackGroundColor = "#6a6969"
ChartColor = '#001534'
GridColor = 'white'
TextColor = 'white'
X_Scroll = TRUE
Y_Scroll = TRUE
NumLevels_X = 15
NumLevels_Y = 15
Debug = FALSE
Alpha = 0.8

# BoxPlot
XVar <- NULL
GroupVar <- "BRAND"
YVar <- "CHILLED_Margin_PerDay"
}

}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Stock}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
