% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plots.ModelEvaluation}
\alias{Plots.ModelEvaluation}
\title{Plots.ModelEvaluation}
\usage{
Plots.ModelEvaluation(
  dt = NULL,
  AggMethod = "mean",
  SampleSize = 100000L,
  PlotType = NULL,
  YVar = NULL,
  TargetLevel = NULL,
  ZVar = NULL,
  XVar = NULL,
  GroupVar = NULL,
  YVarTrans = "Identity",
  XVarTrans = "Identity",
  ZVarTrans = "Identity",
  FacetRows = 1,
  FacetCols = 1,
  FacetLevels = NULL,
  NumLevels_Y = 75,
  NumLevels_X = 40,
  Height = "600px",
  Width = "1100px",
  Title = NULL,
  PlotEngineType = "Echarts",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  FillColor = "#0066ff",
  FillColorReverse = "#97ff00",
  GridColor = "white",
  TextColor = "white",
  FontSize = 14L,
  NumberBins = 20,
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{AggMethod}{character}

\item{SampleSize}{100000L}

\item{PlotType}{character}

\item{YVar}{character}

\item{TargetLevel}{character}

\item{ZVar}{character}

\item{XVar}{character}

\item{GroupVar}{character}

\item{YVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{XVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{ZVarTrans}{"Asinh", "Log", "LogPlus1", "Sqrt", "Asin", "Logit", "PercRank", "Standardize", "BoxCox", "YeoJohnson"}

\item{FacetRows}{Defaults to 1 which causes no faceting to occur vertically. Otherwise, supply a numeric value for the number of output grid rows}

\item{FacetCols}{Defaults to 1 which causes no faceting to occur horizontally. Otherwise, supply a numeric value for the number of output grid columns}

\item{FacetLevels}{Faceting rows x columns is the max number of levels allowed in a grid. If your GroupVar has more you can supply the levels to display.}

\item{NumLevels_Y}{= 75}

\item{NumLevels_X}{= 40}

\item{Height}{= "600px",}

\item{Width}{= "1100px",}

\item{PlotEngineType}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{BackGroundColor}{hex}

\item{ChartColor}{hex}

\item{FillColor}{hex}

\item{FillColorReverse}{hex}

\item{GridColor}{hex}

\item{TextColor}{hex}

\item{NumberBins}{numeric}

\item{Debug}{logical}
}
\description{
Plot helper for model evaluation plot types
}
\examples{
\dontrun{
dt = NULL,
SampleSize = 100000L,
PlotType = NULL,
YVar = NULL,
TargetLevel = NULL,
ZVar = NULL,
XVar = NULL,
Title = NULL,
PlotEngineType = "Plotly",
EchartsTheme = "dark-blue",
TimeLine = FALSE,
BackGroundColor = "#6a6969",
ChartColor = '#001534',
GridColor = 'white',
FillColor = "#0066ff",
TextColor = 'white',
NumberBins = 20,
Debug = FALSE
}

}
\seealso{
Other Auto Plotting: 
\code{\link{Plot.StandardPlots}()}
}
\author{
Adrian Antico
}
\concept{Auto Plotting}
