% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Plots.ModelEvaluation}
\alias{Plots.ModelEvaluation}
\title{Plots.ModelEvaluation}
\usage{
Plots.ModelEvaluation(
  dt = NULL,
  SampleSize = 100000L,
  PlotType = NULL,
  YVar = NULL,
  TargetLevel = NULL,
  ZVar = NULL,
  XVar = NULL,
  GroupVar = NULL,
  Title = NULL,
  PlotEngineType = "Echarts",
  EchartsTheme = "dark-blue",
  TimeLine = FALSE,
  BackGroundColor = "#6a6969",
  ChartColor = "#001534",
  GridColor = "white",
  FillColor = "#0066ff",
  FillColorReverse = "#ff9900",
  TextColor = "white",
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  GamFit = FALSE,
  Buckets = 20,
  ShapAgg = "mean",
  Debug = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{SampleSize}{100000L}

\item{PlotType}{character}

\item{YVar}{character}

\item{TargetLevel}{character}

\item{ZVar}{character}

\item{XVar}{character}

\item{GroupVar}{character}

\item{PlotEngineType}{character}

\item{EchartsTheme}{character}

\item{TimeLine}{logical}

\item{BackGroundColor}{hex}

\item{ChartColor}{hex}

\item{GridColor}{hex}

\item{FillColor}{hex}

\item{FillColorReverse}{hex}

\item{TextColor}{hex}

\item{FacetVar1}{placeholder}

\item{FacetVar2}{placeholder}

\item{GamFit}{placeholder}

\item{Buckets}{numeric}

\item{ShapAgg}{character. A string for aggregating shapely values for importances. Choices include, 'mean', 'absmean', 'meanabs', 'geomean', 'harmmean', 'sd', 'median', 'absmedian', 'medianabs'}

\item{Debug}{logical}
}
\description{
Plot helper for model evaluation plot types
}
\examples{
\dontrun{
dt = NULL,
PlotEngineType = "Plotly",
EchartsTheme = "dark-blue",
TimeLine = FALSE,
SampleSize = 100000L,
PlotType = NULL,
YVar = NULL,
TargetLevel = NULL,
ZVar = NULL,
XVar = NULL,
BackGroundColor = "#6a6969",
ChartColor = '#001534',
GridColor = 'white',
FillColor = "#0066ff",
TextColor = 'white',
FacetVar1 = NULL,
FacetVar2 = NULL,
GamFit = FALSE,
Buckets = 20,
ShapAgg = 'mean',
Debug = FALSE
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{ChartTheme}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StandardPlots}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()},
\code{\link{StockData}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
