% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_echarts4r_functions.R
\name{save_image}
\alias{save_image}
\title{Save plot as image}
\usage{
save_image(plot, path, name, height = 1027, width = 1500, delay = 1, zoom = 2)
}
\arguments{
\item{plot}{A plot object or display_plot_grid object}

\item{path}{Directory to save image}

\item{name}{name of output image. name only}

\item{height}{Default 1027 good for PowerPoint}

\item{width}{Default 1500 good for PowerPoint}

\item{delay}{Default 1 good for PowerPoint}

\item{zoom}{Default 2 good for PowerPoint}
}
\value{
A browsable HTML grid for use in Rmarkdown, Shiny, or viewer pane.
}
\description{
Save plot as image
}
