% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{StockData}
\alias{StockData}
\title{StockData}
\usage{
StockData(
  PolyOut = NULL,
  Symbol = "TSLA",
  CompanyName = "Tesla Inc. Common Stock",
  Metric = "Stock Price",
  TimeAgg = "days",
  StartDate = "2021-01-01",
  EndDate = "2022-01-01",
  APIKey = NULL
)
}
\arguments{
\item{PolyOut}{NULL. If NULL, data is pulled. If supplied, data is not pulled.}

\item{Symbol}{ticker symbol string}

\item{CompanyName}{company name if you have it. ends up in title, that is all}

\item{Metric}{Stock Price, Percent Returns (use symbol for percent), Percent Log Returns (use symbol for percent), Index, Quadratic Variation}

\item{TimeAgg}{= 'days', 'weeks', 'months'}

\item{StartDate}{Supply a start date. E.g. '2022-01-01'}

\item{EndDate}{Supply an end date. E.g. `Sys.Date()`}

\item{APIKey}{Supply your polygon API key}

\item{Type}{'candlestick', 'ohlc'}
}
\description{
Create stock data for plotting using Plot.Stock()
}
\seealso{
Other Standard Plots: 
\code{\link{Plot.Area}()},
\code{\link{Plot.BarPlot3D}()},
\code{\link{Plot.Bar}()},
\code{\link{Plot.Box}()},
\code{\link{Plot.Copula3D}()},
\code{\link{Plot.Copula}()},
\code{\link{Plot.CorrMatrix}()},
\code{\link{Plot.Density}()},
\code{\link{Plot.HeatMap}()},
\code{\link{Plot.Histogram}()},
\code{\link{Plot.Line}()},
\code{\link{Plot.Pie}()},
\code{\link{Plot.River}()},
\code{\link{Plot.Scatter3D}()},
\code{\link{Plot.Scatter}()},
\code{\link{Plot.StackedBar}()},
\code{\link{Plot.Step}()},
\code{\link{Plot.Stock}()},
\code{\link{Plot.Violin}()}
}
\author{
Adrian Antico
}
\concept{Standard Plots}
